% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_ArrayDistribution.R
\name{ArrayDistribution}
\alias{ArrayDistribution}
\title{Product ArrayDistribution}
\description{
A special case joint distribution where each indepdent distribution is the same Distribution
but not necessarily with the same parameters.
}
\details{
Exploits the following relationships of independent distributions
\deqn{f_P(x, y) = f_X(x) * f_Y(y)}
\deqn{F_P(x, y) = F_X(x) * F_Y(y)}
where f_P/F_P is the pdf/cdf of the joint (product) distribution P and X,Y are independent distributions.

\code{JointDistribution} inherits all methods from \code{\link{Distribution}} and \code{\link{DistributionWrapper}}.
}
\section{Constructor Arguments}{

\tabular{ll}{
\code{distribution} \tab Distribution. \cr
\code{paramVector} \tab vector of parameters, see details. \cr
}
}

\examples{
\dontrun{
prodBin <- ArrayDistribution$new(Binomial$new(prob = 0.5, size = 10),
Binomial$new(prob = 0.5, size = 20))
prodBin$pdf(2)
prodBin$cdf(5)
}
}
\seealso{
\code{\link{ProductDistribution}} and \code{\link{Distribution}}. As well as \code{\link{DistributionWrapper}}
for wrapper details.
}
