% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{as.data.table.ParameterSet}
\alias{as.data.table.ParameterSet}
\title{Coerce ParameterSet to data.table}
\usage{
\method{as.data.table}{ParameterSet}(x, ...)
}
\arguments{
\item{x}{ParameterSet}

\item{...}{Ignored.}
}
\value{
A data.table.
}
\description{
Coerces a ParameterSet to a data.table.
}
