% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{as.data.table}
\alias{as.data.table}
\title{Coerce ParameterSet to data.table}
\usage{
as.data.table(object)
}
\arguments{
\item{object}{ParameterSet}
}
\description{
Coerces a ParameterSet to a data.table.
}
\section{R6 Usage}{
 $as.data.table()
}

\seealso{
\code{\link{ParameterSet}}
}
