% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_MixtureDistribution.R
\name{as.MixtureDistribution}
\alias{as.MixtureDistribution}
\title{Coercion to Mixture Distribution}
\usage{
as.MixtureDistribution(object, weights = "uniform")
}
\arguments{
\item{object}{\link{ProductDistribution} or \link{VectorDistribution}}
}
\description{
Helper functions to quickly convert compatible objects to
a \link{MixtureDistribution}.
}
