% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_Interval.R
\name{as.numeric.Interval}
\alias{as.numeric.Interval}
\title{Coerces Interval to Numeric}
\value{
If the Interval is of class integer then returns a numeric vector of the interval elements.
}
\description{
If possible coerces an integer-class to numeric.
}
\details{
If the interval is of class 'integer', returns the interval as a numeric. Otherwise does nothing.

This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $length()
}

\seealso{
\code{\link{Set}}, \code{\link{length.Interval}}
}
