% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{as.ParameterSet}
\alias{as.ParameterSet}
\alias{as.ParameterSet.data.table}
\alias{as.ParameterSet.list}
\title{Coerce to a ParameterSet}
\usage{
as.ParameterSet(x,...)

\method{as.ParameterSet}{data.table}(x, ...)

\method{as.ParameterSet}{list}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional arguments}
}
\value{
An R6 object of class ParameterSet.
}
\description{
Coerces objects to ParameterSet.
}
\details{
Currently supported coercions are from data tables and lists. Function assumes
that the data table columns are the correct inputs to a ParameterSet, see the constructor
for details. Similarly for lists, names are taken to be ParameterSet parameters and values taken
to be arguments.
}
\seealso{
\code{\link{ParameterSet}}
}
