% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_ProductDistribution.R
\name{as.ProductDistribution}
\alias{as.ProductDistribution}
\title{Coercion to Product Distribution}
\usage{
as.ProductDistribution(object)
}
\arguments{
\item{object}{\link{MixtureDistribution} or \link{VectorDistribution}}
}
\description{
Helper functions to quickly convert compatible objects to
a \link{ProductDistribution}.
}
