% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_VectorDistribution.R
\name{as.VectorDistribution}
\alias{as.VectorDistribution}
\title{Coercion to Vector Distribution}
\usage{
as.VectorDistribution(object)
}
\arguments{
\item{object}{\link{MixtureDistribution} or \link{ProductDistribution}}
}
\description{
Helper functions to quickly convert compatible objects to
a \link{VectorDistribution}.
}
