% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_helpers.R
\name{assertDiscrete}
\alias{assertDiscrete}
\alias{checkDiscrete}
\alias{testDiscrete}
\alias{isDiscrete}
\title{assert/check/test/Discrete}
\usage{
assertDiscrete(x)
checkDiscrete(x)
testDiscrete(x)
isDiscrete(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is discrete
}
\examples{
isDiscrete(Binomial$new()) # FALSE
}
