% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_helpers.R
\name{assertDistributionList}
\alias{assertDistributionList}
\alias{checkDistributionList}
\alias{testDistributionList}
\alias{isDistributionList}
\title{assert/check/test/isDistributionList}
\usage{
assertDistributionList(x)
checkDistributionList(x)
testDistributionList(x)
isDistributionList(x)
}
\arguments{
\item{x}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if a given object is a list of R6 Distributions
}
\examples{
isDistributionList(list(Binomial$new(),5)) # FALSE
isDistributionList(list(Binomial$new(),Exponential$new())) # TRUE
}
