% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_helpers.R
\name{assertSkewness}
\alias{assertSkewness}
\alias{checkSkewness}
\alias{testSkewness}
\alias{isSkewness}
\title{assert/check/test/Skewness}
\usage{
assertSkewness(x, type)
checkSkewness(x, type)
testSkewness(x, type)
isSkewness(x, type)
}
\arguments{
\item{x}{Distribution}

\item{type}{skewness type, Negative Skew/No Skew/Positive Skew}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test the skewnewss of a Distribution
}
\examples{
testSkewness(Binomial$new(), "Negative Skew")
testSkewness(Binomial$new(), "Positive Skew")
}
