% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDistribution_BetaNoncentral.R
\name{BetaNoncentral}
\alias{BetaNoncentral}
\title{Noncentral Beta Distribution Class}
\description{
Mathematical and statistical functions for the Noncentral Beta distribution, which
is commonly used as the prior in Bayesian modelling.
}
\details{
The Noncentral Beta distribution parameterised with  two shape parameters, \eqn{\alpha, \beta}, and location, \eqn{\lambda}, is defined by the pdf, \deqn{f(x) = exp(-\lambda/2) \sum_{r=0}^\infty ((\lambda/2)^r/r!) (x^{\alpha+r-1}(1-x)^{\beta-1})/B(\alpha+r, \beta)} for \eqn{\alpha, \beta > 0, \lambda \geq 0}, where \eqn{B} is the Beta function.

The distribution is supported on
}
\examples{
x = BetaNoncentral$new(shape1 = 2, shape2 = 5, location = 3)

# Update parameters
x$setParameterValue(shape1 = 1)
x$parameters()

# d/p/q/r
x$pdf(5)
x$cdf(5)
x$quantile(0.42)
x$rand(4)

summary(x)

}
