% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{cdf}
\alias{cdf}
\title{Cumulative Distribution Function}
\usage{
cdf(object, x1, ..., lower.tail = TRUE, log.p = FALSE, simplify = TRUE)
}
\arguments{
\item{object}{Distribution.}

\item{x1}{vector of numerics to evaluate function at.}

\item{...}{additional arguments.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X \le x)} otherwise, \eqn{P(X > x)}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{simplify}{if TRUE (default) returns results in simplest form (vector or data.table) otherwise as data.table.}
}
\value{
Cumulative distribution function evaluated at given points as either a numeric if \code{simplify} is TRUE
or as a data.table.
}
\description{
Returns the cumulative distribution function for a distribution evaluated at a given
point.
}
\details{

}
