% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{cdfAntiDeriv}
\alias{cdfAntiDeriv}
\title{Cumulative Distribution Function Anti-Derivative}
\arguments{
\item{lower}{lower limit for integration, default is infimum.}

\item{upper}{upper limit for integration, default is supremum.}
}
\description{
The anti-derivative of the cumulative distribution function between given limits or
over the full support.
}
\details{
The cdf anti-derivative is defined by
\deqn{acdf(a, b) = \int_a^b F_X(x) dx}
where X is the distribution, F_X is the cdf of the distribution X and a,b are the limits of integration.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method.
}
