% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Kernel.R
\name{cdfSquared2Norm}
\alias{cdfSquared2Norm}
\title{Squared Cumulative Distribution Function 2-Norm}
\usage{
cdfSquared2Norm(object, x = 0, upper = Inf)
}
\arguments{
\item{object}{Distribution.}

\item{x}{amount to shift the result.}

\item{upper}{upper limit of the integral.}
}
\value{
Squared 2-norm of cdf evaluated between limits as a numeric.
}
\description{
The squared 2-norm of the cdf evaluated up to a given limit, possibly shifted.
}
