% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{cf}
\alias{cf}
\alias{cf.Distribution}
\title{Characteristic Function}
\usage{
\method{cf}{Distribution}(object, t)
}
\arguments{
\item{object}{an object used to select a method.}

\item{t}{integer to evaluate characteristic function at}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Characteristic function of a distribution
}
\details{
The characteristic function is defined by
\deqn{cf_X(t) = E_X[exp(xti)]}
where X is the distribution and E_X is the expectation of the distribution X.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.
}
