% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{class.SetInterval}
\alias{class.SetInterval}
\title{SetInterval Minimum Accessor}
\description{
Returns the SetInterval minimum.
}
\details{
This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $min()
}

\seealso{
\code{\link{SetInterval}}
}
