% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{complement}
\alias{complement}
\alias{-.SetInterval}
\title{Symbolic Complement for SetInterval}
\usage{
complement(...)

\method{-}{SetInterval}(x, y)
}
\arguments{
\item{...}{sets and/or intervals to take the union of.}

\item{x}{distribution}

\item{y}{distribution}
}
\value{
An R6 object of class SetInterval.
}
\description{
Makes a symbolic representation for the complement of sets/intervals.
}
\details{
This does not calculate the complement of the arguments but
  is just a symbolic representation using unicode.
}
\seealso{
\code{\link{product}} and \code{\link{union}}.
}
