% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{complement.SetInterval}
\alias{complement.SetInterval}
\alias{-.SetInterval}
\title{Symbolic complement.SetInterval for SetInterval}
\usage{
complement.SetInterval(...)

\method{-}{SetInterval}(x, y)
}
\arguments{
\item{...}{sets and/or intervals to take the union.SetInterval of.}

\item{x}{distribution}

\item{y}{distribution}
}
\value{
An R6 object of class SetInterval.
}
\description{
Makes a symbolic representation for the complement.SetInterval of sets/intervals.
}
\details{
This does not calculate the complement.SetInterval of the arguments but
  is just a symbolic representation using unicode.
}
\seealso{
\code{\link{product.SetInterval}} and \code{\link{union.SetInterval}}.
}
