% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{ConcreteWrapper}
\alias{ConcreteWrapper}
\title{Concrete Wrapper for Distributions}
\description{
An R6 concrete wrapper class for use with decorators.
}
\details{
Should not be constructed by the user. The purpose of this class is to provide a non-abstract
interface that can be initialized by decorators when required.
}
\seealso{
\code{\link{DistributionWrapper}}
}
