% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{decorators}
\alias{decorators}
\title{Decorators Accessor}
\usage{
decorators(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
Character vector of decorators.
}
\description{
Returns the decorators added to a distribution.
}
\section{R6 Usage}{
 $decorators
}

