% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{dimension.SetInterval}
\alias{dimension.SetInterval}
\title{SetInterval Dimension Accessor}
\description{
Returns the SetInterval dimension.
}
\details{
This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $dimension()
}

\seealso{
\code{\link{SetInterval}}
}
