% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_ArrayDistribution.R,
%   R/distr6-deprecated.R, R/listSpecialSets.R
\name{ArrayDistribution}
\alias{ArrayDistribution}
\alias{distr6-deprecated}
\alias{listSpecialSets}
\title{Deprecated distr6 Functions and Classes}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("-deprecated")}.
}
\section{Deprecated Functions}{

\tabular{ll}{
\strong{Deprecated} \tab \strong{Replacement} \cr
\code{ArrayDistribution} \tab \link{ProductDistribution} \cr
\code{listSpecialSets} \tab \code{\link[set6:listSpecialSets]{set6::listSpecialSets()}} \cr
}
}

\keyword{internal}
