% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr6.news.R
\name{distr6News}
\alias{distr6News}
\title{Show distr6 NEWS.md File}
\usage{
distr6News()
}
\value{
NEWS.md in viewer.
}
\description{
Displays the contents of the NEWS.md file for viewing distr6
release information.
}
\examples{
\dontrun{
distr6News()
}
}
