% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{distrDomain}
\alias{distrDomain}
\title{Distribution Domain Accessor}
\usage{
distrDomain(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
Returns the distribution domain.
}
\details{
The domain of a probability distribution is the set of values returned by the pdf/pdf,
including zero.
}
\section{R6 Usage}{
 $distrDomain()
}

\seealso{
\code{\link{SetInterval}} and \code{\link{properties}}
}
