% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator.R
\name{DistributionDecorator}
\alias{DistributionDecorator}
\title{Abstract Decorator for Distribution}
\arguments{
\item{distribution}{distribution object}

\item{pos}{position of environment to assign decorated distribution}
}
\value{
error
}
\description{
An abstract decorator object for distributions to enhance functionality.
}
\details{
This is an abstract class that cannot be directly constructed.
  See the concete child classes for initializable statistical decorators.
}
\seealso{
\code{\link{CoreStatistics}} and \code{\link{ExoticStatistics}}.
}
