% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{dmax}
\alias{dmax}
\title{Distribution Maximum Accessor}
\usage{
dmax(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
Maximum as a numeric.
}
\description{
Returns the distribution maximum as the maximum of the support. If the support is
not bounded above then maximum is given by
\deqn{maximum = supremum - 1.1e-15}
}
\section{R6 Usage}{
 $dmax
}

\seealso{
\code{\link{support}}, \code{\link{dmin}}, \code{\link{sup}}, \code{\link{inf}}
}
