% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator.R
\name{.assign_distribution}
\alias{.assign_distribution}
\alias{.assign_wrappedDistribution}
\title{Internal Helper Functions for Decorators}
\usage{
.assign_distribution(distribution, pos, decorators, distname)

.assign_wrappedDistribution(distribution, pos, decorators, distname)
}
\arguments{
\item{distribution}{distribution object}

\item{pos}{position of environment to assign decorated distribution}

\item{decorators}{decorators to decorate Distribution with}

\item{distname}{name of distribution to assign to environment}
}
\value{
Assigns distribution to given environment.
}
\description{
Used by decorators to overload distributiond with an extra decorator interface.
}
\details{
These are internal helper functions that should not be directly used.
}
\seealso{
\code{\link{DistributionDecorator}}.
}
