% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dparse.R
\name{dparse}
\alias{dparse}
\title{String parses}
\usage{
dparse(toparse)
}
\arguments{
\item{toparse}{character}
}
\value{
Returns an R6 distribution object
}
\description{
Parse a custom string that represents an R6 distribution
}
\details{
Transform a custom (user) input to a R6 object.

This function is specially useful when you expect a user input which should
result in specific distribution. The distribution name must be the
ShortName, ClassName or Alias listed in the package. You can list them
using the \link{listDistributions} function.

Text call must be in the following format: Distribution([params]), e.g.
T() or Normal(mean = 3).

The parameters can be entered in any order, but they must be named. If no
parameter is given, the default parametrization provided in $new() call
is used.
}
\examples{
dparse("N()")
dparse("T(df = 4)")
dparse("chisq(df = 3)")
}
