% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_Set.R
\name{elements}
\alias{elements}
\title{Set Elements Accessor}
\value{
Elements in the set.
}
\description{
Returns the elements in a Set.
}
\details{
This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $elements()
}

\seealso{
\code{\link{Set}}
}
