% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\name{Empty}
\alias{Empty}
\title{Empty Set}
\description{
The mathematical Empty, or Null, set.
}
\details{
The Empty set is the set defined by having no elements,
\deqn{Empty = \{ \}}{Empty = { }}
}
\section{Constructor}{
 Empty$new()
}

\examples{
Empty$new()

}
\seealso{
\code{\link{listSpecialSets}}
}
