% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Epanechnikov.R
\name{Epanechnikov}
\alias{Epanechnikov}
\title{Epanechnikov Kernel}
\description{
Mathematical and statistical functions for the Epanechnikov kernel defined by the
pdf, \deqn{f(x) = \frac{3}{4}(1-x^2)}{f(x) = 3/4(1-x^2)}
over the support \eqn{x \in (-1,1)}{x \epsilon (-1,1)}.
}
\details{
The quantile function is omitted as no closed form analytic expressions could
be found, decorate with FunctionImputation for numeric results.
}
\seealso{
Other kernels: 
\code{\link{Cosine}},
\code{\link{LogisticKernel}},
\code{\link{NormalKernel}},
\code{\link{Quartic}},
\code{\link{Sigmoid}},
\code{\link{Silverman}},
\code{\link{TriangularKernel}},
\code{\link{Tricube}},
\code{\link{Triweight}},
\code{\link{UniformKernel}}
}
\concept{kernels}
\section{Super classes}{
\code{\link[distr6:Distribution]{distr6::Distribution}} -> \code{\link[distr6:Kernel]{distr6::Kernel}} -> \code{Epanechnikov}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Full name of distribution.}

\item{\code{short_name}}{Short name of distribution for printing.}

\item{\code{description}}{Brief description of the distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Epanechnikov-pdfSquared2Norm}{\code{Epanechnikov$pdfSquared2Norm()}}
\item \href{#method-Epanechnikov-cdfSquared2Norm}{\code{Epanechnikov$cdfSquared2Norm()}}
\item \href{#method-Epanechnikov-variance}{\code{Epanechnikov$variance()}}
\item \href{#method-Epanechnikov-clone}{\code{Epanechnikov$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="cdf"><a href='../../distr6/html/Distribution.html#method-Distribution-cdf'><code>distr6::Distribution$cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="confidence"><a href='../../distr6/html/Distribution.html#method-Distribution-confidence'><code>distr6::Distribution$confidence()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="correlation"><a href='../../distr6/html/Distribution.html#method-Distribution-correlation'><code>distr6::Distribution$correlation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="getParameterValue"><a href='../../distr6/html/Distribution.html#method-Distribution-getParameterValue'><code>distr6::Distribution$getParameterValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="iqr"><a href='../../distr6/html/Distribution.html#method-Distribution-iqr'><code>distr6::Distribution$iqr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="liesInSupport"><a href='../../distr6/html/Distribution.html#method-Distribution-liesInSupport'><code>distr6::Distribution$liesInSupport()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="liesInType"><a href='../../distr6/html/Distribution.html#method-Distribution-liesInType'><code>distr6::Distribution$liesInType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="parameters"><a href='../../distr6/html/Distribution.html#method-Distribution-parameters'><code>distr6::Distribution$parameters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="pdf"><a href='../../distr6/html/Distribution.html#method-Distribution-pdf'><code>distr6::Distribution$pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="prec"><a href='../../distr6/html/Distribution.html#method-Distribution-prec'><code>distr6::Distribution$prec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="print"><a href='../../distr6/html/Distribution.html#method-Distribution-print'><code>distr6::Distribution$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="quantile"><a href='../../distr6/html/Distribution.html#method-Distribution-quantile'><code>distr6::Distribution$quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="rand"><a href='../../distr6/html/Distribution.html#method-Distribution-rand'><code>distr6::Distribution$rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="setParameterValue"><a href='../../distr6/html/Distribution.html#method-Distribution-setParameterValue'><code>distr6::Distribution$setParameterValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="stdev"><a href='../../distr6/html/Distribution.html#method-Distribution-stdev'><code>distr6::Distribution$stdev()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="strprint"><a href='../../distr6/html/Distribution.html#method-Distribution-strprint'><code>distr6::Distribution$strprint()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="summary"><a href='../../distr6/html/Distribution.html#method-Distribution-summary'><code>distr6::Distribution$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="workingSupport"><a href='../../distr6/html/Distribution.html#method-Distribution-workingSupport'><code>distr6::Distribution$workingSupport()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Kernel" data-id="initialize"><a href='../../distr6/html/Kernel.html#method-Kernel-initialize'><code>distr6::Kernel$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Kernel" data-id="mean"><a href='../../distr6/html/Kernel.html#method-Kernel-mean'><code>distr6::Kernel$mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Kernel" data-id="median"><a href='../../distr6/html/Kernel.html#method-Kernel-median'><code>distr6::Kernel$median()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Kernel" data-id="mode"><a href='../../distr6/html/Kernel.html#method-Kernel-mode'><code>distr6::Kernel$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Kernel" data-id="skewness"><a href='../../distr6/html/Kernel.html#method-Kernel-skewness'><code>distr6::Kernel$skewness()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Epanechnikov-pdfSquared2Norm"></a>}}
\if{latex}{\out{\hypertarget{method-Epanechnikov-pdfSquared2Norm}{}}}
\subsection{Method \code{pdfSquared2Norm()}}{
The squared 2-norm of the pdf is defined by
\deqn{\int_a^b (f_X(u))^2 du}
where X is the Distribution, \eqn{f_X} is its pdf and \eqn{a, b}
are the distribution support limits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Epanechnikov$pdfSquared2Norm(x = 0, upper = Inf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{(numeric(1))} \cr
Amount to shift the result.}

\item{\code{upper}}{\code{(numeric(1))} \cr
Upper limit of the integral.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Epanechnikov-cdfSquared2Norm"></a>}}
\if{latex}{\out{\hypertarget{method-Epanechnikov-cdfSquared2Norm}{}}}
\subsection{Method \code{cdfSquared2Norm()}}{
The squared 2-norm of the cdf is defined by
\deqn{\int_a^b (F_X(u))^2 du}
where X is the Distribution, \eqn{F_X} is its pdf and \eqn{a, b}
are the distribution support limits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Epanechnikov$cdfSquared2Norm(x = 0, upper = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{(numeric(1))} \cr
Amount to shift the result.}

\item{\code{upper}}{\code{(numeric(1))} \cr
Upper limit of the integral.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Epanechnikov-variance"></a>}}
\if{latex}{\out{\hypertarget{method-Epanechnikov-variance}{}}}
\subsection{Method \code{variance()}}{
The variance of a distribution is defined by the formula
\deqn{var_X = E[X^2] - E[X]^2}
where \eqn{E_X} is the expectation of distribution X. If the distribution is multivariate the
covariance matrix is returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Epanechnikov$variance(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Epanechnikov-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Epanechnikov-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Epanechnikov$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
