% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Exponential.R
\docType{data}
\name{Exponential}
\alias{Exponential}
\title{Exponential Distribution}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Exponential
}
\description{
Exponential Distribution
}
\keyword{datasets}
