% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_VectorDistribution.R
\name{Extract.VectorDistribution}
\alias{Extract.VectorDistribution}
\alias{[.VectorDistribution}
\title{Extract one or more Distributions from a VectorDistribution}
\usage{
Extract.VectorDistribution(vecdist, i)

\method{[}{VectorDistribution}(object, i)
}
\arguments{
\item{vecdist}{VectorDistribution from which to extract Distributions.}

\item{i}{indices specifying distributions to extract.}
}
\description{
Once a \code{VectorDistribution} has been constructed, use \code{[}
to extract one or more \code{Distribution}s from inside it.
}
