% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDistribution_FDistributionNoncentral.R
\name{FDistributionNoncentral}
\alias{FDistributionNoncentral}
\title{Noncentral F Distribution Class}
\description{
Mathematical and statistical functions for the Noncentral F distribution, which
is commonly used in ANOVA testing and is the ratio of scaled Chi-Squared distributions..
}
\details{
The Noncentral F distribution   is defined by the
}
\examples{
x <- FDistributionNoncentral$new(df1 = 1, df2 = 3, location = 2)

# Update parameters
x$setParameterValue(df2 = 10)
x$parameters()

# d/p/q/r
x$pdf(5)
x$cdf(5)
x$quantile(0.42)
x$rand(4)

# Statistics
x$mean()
x$variance()

summary(x)

}
