% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_FunctionImputation.R
\name{FunctionImputation}
\alias{FunctionImputation}
\title{Numerical Pdf/Cdf/Quantile/Rand Functions}
\description{
Numeric statistical functions including density/mass function, distribution function,
inverse distribution and simulation.
}
\details{
Decorator objects add functionality to the given Distribution object
 by copying methods in the decorator environment to the chosen Distribution environment. Use the
 \code{\link{decorate}} function to decorate a Distribution. See the help pages for the individual
 CoreStatistics methods to learn more.

 All methods in this decorator use numerical approximations and therefore better results may be available
 from analytic computations.
}
\examples{
x = Exponential$new()
decorate(x, ExoticStatistics, R62S3 = FALSE)
x$survival(1)

x = Exponential$new(decorators = ExoticStatistics, R62S3 = FALSE)
x$survival(4)
}
\seealso{
\code{\link{DistributionDecorator}}
}
