% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Gamma.R
\name{Gamma}
\alias{Gamma}
\title{Gamma Distribution}
\description{
Mathematical and statistical functions for the Gamma distribution parameterised
with mean and shape, rate or scale. By default we use the shape/rate combination, which is most common
in statistics (particularly Bayesian). The shape/rate Gamma distribution is defined by the pdf,
\deqn{f(x) = (b^a)/Gamma(a) * x^(a-1) * exp(-bx)}
where a is the shape parameter and b is the rate parameter.
}
\details{
All results given are analytical but we do make use of the base R gamma function, which is
found in common usage throughout packages in R and we therefore trust the precision of the results.
}
\section{Constructor}{
 Gamma$new(shape = 1,rate = 1, scale = NULL, mean = NULL, decorators = NULL, verbose = FALSE)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{shape} \tab numeric \tab shape parameter. \cr
\code{rate} \tab numeric \tab inverse scale parameter. \cr
\code{scale} \tab numeric \tab scale parameter. \cr
\code{mean} \tab numeric \tab alternate scale parameter. \cr
\code{decorators} \tab Decorator \tab decorators to add functionality. \cr
\code{verbose} \tab logical \tab if TRUE parameterisation messages produced.
}
}

\section{Constructor Details}{
 The Gamma distribution can either be parameterised with rate,
scale or mean. If none are provided then the default is taken to be rate = 1 and shape = 1.
If multiple are provided then parameterisation takes the hierarchy: mean, scale, rate.
Scale is defined by
\deqn{scale = rate^-1}
Mean is defined by
\deqn{mean = shape/rate}
}

\section{Public Variables}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Return} \cr
  \code{name} \tab Name of distribution. \cr
  \code{short_name} \tab Id of distribution. \cr
  \code{description} \tab Brief description of distribution. \cr
  \code{traits} \tab List: type, valueSupport, variateForm.
  }
}

\section{Accessor Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{decorators()} \tab \code{\link{decorators}} \cr
  \code{valueSupport()} \tab \code{\link{valueSupport}} \cr
  \code{variateForm()} \tab \code{\link{variateForm}} \cr
  \code{type()} \tab \code{\link{type}} \cr
  \code{properties()} \tab \code{\link{properties}} \cr
  \code{support()} \tab \code{\link{support}} \cr
  \code{distrDomain()} \tab \code{\link{distrDomain}} \cr
  \code{symmetry()} \tab \code{\link{symmetry}} \cr
  \code{sup()}  \tab \code{\link{sup}} \cr
  \code{inf()} \tab \code{\link{inf}} \cr
  \code{dmax()}  \tab \code{\link{dmax}} \cr
  \code{dmin()} \tab \code{\link{dmin}} \cr
  \code{skewnessType()} \tab \code{\link{skewnessType}} \cr
  \code{kurtosisType()} \tab \code{\link{kurtosisType}} \cr
  }
}

\section{p/d/q/r Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{pdf(x1, ..., log = FALSE)} \tab \code{\link{pdf}} \cr
  \code{cdf(x1, ..., lower.tail = TRUE, log.p = FALSE)} \tab \code{\link{cdf}}\cr
  \code{quantile(p, ..., lower.tail = TRUE, log.p = FALSE)} \tab \code{\link{quantile.Distribution}} \cr
  \code{rand(n)} \tab \code{\link{rand}}
 }
}

\section{Statistical Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{mean()} \tab \code{\link{mean.Distribution}} \cr
  \code{var()} \tab \code{\link{var}} \cr
  \code{skewness()} \tab \code{\link{skewness}} \cr
  \code{kurtosis(excess = TRUE)} \tab \code{\link{kurtosis}} \cr
  \code{entropy(base = 2)} \tab \code{\link{entropy}} \cr
  \code{mgf(t)} \tab \code{\link{mgf}} \cr
  \code{cf(t)} \tab \code{\link{cf}} \cr
  \code{sd()} \tab \code{\link{sd}} \cr
  \code{median()} \tab \code{\link{median.Distribution}} \cr
  \code{iqr()} \tab \code{\link{iqr}} \cr
  }
}

\section{Parameter Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{parameters(id)} \tab \code{\link{parameters}} \cr
  \code{getParameterValue(id, error = "warn")}  \tab \code{\link{getParameterValue}} \cr
  \code{setParameterValue(lst, error = "warn")} \tab \code{\link{setParameterValue}} \cr
}
}

\section{Validation Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{liesInSupport(x, all = TRUE)} \tab \code{\link{liesInSupport}} \cr
  \code{liesInType(x, all = TRUE)} \tab \code{\link{liesInType}} \cr
  \code{liesInDistrDomain(x, all = TRUE)} \tab \code{\link{liesInDistrDomain}} \cr
}
}

\section{Representation Methods}{

\tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{strprint()} \tab \code{\link[RSmisc]{strprint}} \cr
  \code{print()} \tab \code{\link[base]{print}} \cr
  \code{summary(full = T)} \tab \code{\link{summary.Distribution}} \cr
  \code{plot()} \tab Coming Soon. \cr
  \code{qqplot()} \tab Coming Soon. \cr
}
}

