% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{getParameterValue}
\alias{getParameterValue}
\title{Parameter Value Accessor}
\usage{
getParameterValue(object, id, error = "warn")
}
\arguments{
\item{object}{distribution.}

\item{id}{character, id of the parameter to return.}

\item{error}{character, value to pass to \code{stopwarn}.}
}
\description{
Returns the value of the given parameter.
}
\details{
Returns NULL and warning if the given parameter is not in the Distribution, otherwise returns
the value of the given parameter.

\code{stopwarn} either breaks the code with an error if "error" is given or returns \code{NULL}
with warning otherwise.
}
\section{R6 Usage}{
 $getParameterValue(id, error = "warn")
}

\seealso{
\code{\link{parameters}} and \code{\link{setParameterValue}}
}
