% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_helpers.R
\name{getR6Class}
\alias{getR6Class}
\title{Get R6 Class Name}
\usage{
getR6Class(x)
}
\arguments{
\item{x}{R6 object.}
}
\value{
string giving R6 class name.
}
\description{
Gets the name of the given R6 object
}
\details{
The function works for R6 objects not classes. If the function fails make sure the
class is constructed.
}
\examples{
getR6Class(Binomial$new())

}
