% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{getSymbol.SetInterval}
\alias{getSymbol.SetInterval}
\title{SetInterval Symbol Accessor}
\description{
Returns the SetInterval symbol.
}
\details{
This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $getSymbol()
}

\seealso{
\code{\link{SetInterval}}
}
