% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{gprm}
\alias{gprm}
\alias{sprm}
\title{Helper Functionality for Getting and Setting Distribution Parameters}
\usage{
gprm(d, p)

sprm(d, lst)
}
\arguments{
\item{d}{(\code{Distribution(1)}) \cr
Distribution object.}

\item{p}{(\code{character()}) \cr
Name(s) of parameters to written.}

\item{lst}{(\code{list(1)}) \cr
Parameters to update.}
}
\description{
Simple wrapper around d$getParameterValue(p) and
d$setParameterValue(lst).
}
\examples{
d <- dstr("Norm")
gprm(d, "mean")
gprm(d, c("mean", "var"))
sprm(d, list(mean = 1, var = 3))
gprm(d, c("mean", "sd"))
}
