% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_HuberizedDistribution.R
\name{huberize}
\alias{huberize}
\title{Huberize a Distribution}
\usage{
huberize(x, lower, upper)
}
\arguments{
\item{x}{distribution to huberize.}

\item{lower}{lower limit for huberization.}

\item{upper}{upper limit for huberization.}
}
\description{
S3 functionality to huberize an R6 distribution.
}
\seealso{
\link{HuberizedDistribution}
}
