% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{inf.SetInterval}
\alias{inf.SetInterval}
\title{SetInterval Infimum Accessor}
\value{
Infimum as a numeric.
}
\description{
Returns the SetInterval infimum.
}
\details{
This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $inf()
}

\seealso{
\code{\link{SetInterval}}
}
