% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_Interval.R
\name{Interval}
\alias{Interval}
\title{Symbolic Interval Object}
\description{
An R6 set object for symbolic representation of mathematical intervals.
}
\section{Constructor Arguments}{

 \tabular{lll}{
   \strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
   \code{lower = -Inf} \tab numeric \tab Lower limit of set/interval. \cr
   \code{upper = Inf} \tab numeric \tab Upper limit of set/interval. \cr
   \code{type = "[]"} \tab character \tab Interval type, one of (), (], [), []. \cr
}
}

\section{Coercion Methods}{

\tabular{lll}{
\strong{Method} \tab \strong{Return Type} \tab \strong{Details} \cr
\code{as.numeric} \tab numeric \tab Coerces interval range to numeric vector. \cr
}
}

\seealso{
The parent class \code{\link{SetInterval}} for a full list of inherited methods and variables.
}
