% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{isCdf}
\alias{isCdf}
\title{Test the Distribution Cdf Exist?}
\value{
Returns \code{TRUE} if an expression for the Cdf is defined for the distribution, \code{FALSE}
otherwise.
}
\description{
Returns whether or not the distribution has a defined expression for the Cdf.
}
\section{R6 Usage}{
 $isCdf
}

\seealso{
\code{\link{isPdf}}, \code{\link{isQuantile}}, \code{\link{isRand}}
}
