% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{isPdf}
\alias{isPdf}
\title{Test the Distribution Pdf Exist?}
\value{
Returns \code{TRUE} if an expression for the pdf is defined for the distribution, \code{FALSE}
otherwise.
}
\description{
Returns whether or not the distribution has a defined expression for the pdf.
}
\section{R6 Usage}{
 $isPdf
}

\seealso{
\code{\link{isCdf}}, \code{\link{isQuantile}}, \code{\link{isRand}}
}
