% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{isQuantile}
\alias{isQuantile}
\title{Test the Distribution Quantile Exist?}
\value{
Returns \code{TRUE} if an expression for the Quantile is defined for the distribution, \code{FALSE}
otherwise.
}
\description{
Returns whether or not the distribution has a defined expression for the Quantile.
}
\section{R6 Usage}{
 $isQuantile
}

\seealso{
\code{\link{isPdf}}, \code{\link{isCdf}}, \code{\link{isRand}}
}
