% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{isRand}
\alias{isRand}
\title{Test the Distribution Rand Exist?}
\value{
Returns \code{TRUE} if an expression for the Rand is defined for the distribution, \code{FALSE}
otherwise.
}
\description{
Returns whether or not the distribution has a defined expression for the Rand.
}
\section{R6 Usage}{
 $isRand
}

\seealso{
\code{\link{isPdf}}, \code{\link{isCdf}}, \code{\link{isQuantile}}
}
