% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Kernel.R
\name{Kernel}
\alias{Kernel}
\title{Abstract Kernel Class}
\description{
Abstract class for kernels. Call \code{\link[=listKernels]{listKernels()}} for implemented
kernels.
}
\section{Kernels may include the following methods}{

\itemize{
\item \code{mode()} - Mode of the kernel, always \code{0}
\item \code{mean()} - Mean of the kernel, always \code{0}
\item \code{median()} - Median of the kernel, always \code{0}
\item \code{skewness()} - Skewness of the kernel, always \code{0}
\item \code{pdfSquared2Norm} - Squared 2-norm of pdf, will be moved to a new decorator
\item \code{cdfSquared2Norm} - Squared 2-norm of cdf, will be moved to a new decorator
}
}

