% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{kurtosisType}
\alias{kurtosisType}
\title{Type of Kurtosis Accessor}
\usage{
kurtosisType(object)
}
\arguments{
\item{object}{distribution.}
}
\description{
If the distribution kurtosis is present in properties, returns the type of kurtosis
(platykurtic/mesokurtic/leptokurtic), otherwise returns NULL.
}
\section{R6 Usage}{
 $kurtosisType()
}

\seealso{
\code{\link{kurtosis}}, \code{\link{properties}} and \code{\link{skewnessType}}
}
