% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_Interval.R
\name{length.Interval}
\alias{length.Interval}
\title{Length of Interval}
\description{
Returns the length of the Interval after coercing to numeric.
}
\details{
If the interval is of class 'numeric', returns Inf. Otherwise coerces to numeric and returns that
length.

This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $length()
}

\seealso{
\code{\link{Set}}, \code{\link{length.Interval}}
}
