% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_Set.R
\name{length.Set}
\alias{length.Set}
\title{Length of Set}
\value{
Number of elements in the set.
}
\description{
Returns the length of the Set as the number of elements.
}
\details{
This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $length()
}

\seealso{
\code{\link{Set}}
}
