% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_VectorDistribution.R
\name{length.VectorDistribution}
\alias{length.VectorDistribution}
\title{Get Number of Distributions in Vector Distribution}
\usage{
\method{length}{VectorDistribution}(x)
}
\arguments{
\item{x}{\link{VectorDistribution}}
}
\description{
Gets the number of distributions in an object inheriting from
\link{VectorDistribution}.
}
