% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{liesInDistrDomain}
\alias{liesInDistrDomain}
\title{Test if Data Lies in Distribution Domain}
\usage{
liesInDistrDomain(object, x, all = TRUE)
}
\arguments{
\item{object}{distribution.}

\item{x}{vector of numerics to test.}

\item{all}{logical, see details.}
}
\description{
Tests if the given data lies in the domain of the Distribution, either tests if all
data lies in the distribution domain or any of it.
}
\details{
If \code{all} is \code{TRUE} (default) returns \code{TRUE} only if every element in \code{x}
lies in the domain. If \code{all} is \code{FALSE} then returns a vector of logicals for each corresponding element
in the vector \code{x}.
}
\section{R6 Usage}{
 $liesInDistrDomain(x, all = TRUE)
}

\seealso{
\code{\link{liesInSupport}} and \code{\link{liesInType}}
}
