% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{liesInSupport}
\alias{liesInSupport}
\title{Test if Data Lies in Distribution Support}
\usage{
liesInSupport(object, x, all = TRUE)
}
\arguments{
\item{object}{distribution.}

\item{x}{vector of numerics to test.}

\item{all}{logical, see details.}
}
\description{
Tests if the given data lies in the support of the Distribution, either tests if all
data lies in the support or any of it.
}
\details{
If \code{all} is \code{TRUE} (default) returns \code{TRUE} only if every element in \code{x}
lies in the support. If \code{all} is \code{FALSE} then returns a vector of logicals for each corresponding element
in the vector \code{x}.
}
\section{R6 Usage}{
 $liesInSupport(x, all = TRUE)
}

\seealso{
\code{\link{liesInType}} and \code{\link{liesInDistrDomain}}
}
