% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{lines.Distribution}
\alias{lines.Distribution}
\title{Add Distribution Lines to a plot}
\usage{
## S3 method for class "Distribution"
  lines(x, fun, npoints = 3000,...)
}
\arguments{
\item{x}{A distribution object.}

\item{fun}{A list of plottable functions, either one or more of "pdf","cdf","quantile", "survival", "hazard" and "cumhazard".}

\item{npoints}{number of evaluation points.}

\item{...}{graphical parameters to be passed through to plotting methods, see \code{\link[graphics]{par}}.}
}
\description{
A function that adds pdf/pmf, cdf, quantile, survival, hazard and cumulative
  hazard lines of a discrete or continuous distribution to an exisitng plot.
}
\details{
The lines function only works for overlaying distributions of the same value
support type. Therefore bugs can be expected when trying to add continuous to
discrete distributions, and vice versa. The users are responsible for adjusting
the limits of x and y axes based on the parameters of distributions they choose.
}
\examples{
plot(Normal$new(), "pdf")
lines(Normal$new(mean = 0.5, sd = 0.5), "pdf")

}
\seealso{
\code{\link{plot.Distribution}} for drawing a distribution plot.
}
