% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_helpers.R
\name{listDecorators}
\alias{listDecorators}
\title{Lists Implemented Distribution Decorators}
\usage{
listDecorators(simplify = FALSE)
}
\arguments{
\item{simplify}{logical. If FALSE (default) returns result as decorator, otherwise character.}
}
\description{
Lists decorators that can decorate an R6 Distribution.
}
\examples{
listDecorators()
listDecorators(FALSE)
}
