% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listKernels.R
\name{listKernels}
\alias{listKernels}
\title{Lists Implemented R6 Kernels}
\usage{
listKernels(simplify = FALSE)
}
\arguments{
\item{simplify}{logical.}
}
\description{
Lists R6 kernels, either all in a data.table or filtered by chosen
traits and/or properties.
}
\examples{
listKernels()

}
