% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_helpers.R
\name{listWrappers}
\alias{listWrappers}
\title{Lists Implemented Distribution Decorators}
\usage{
listWrappers(simplify = TRUE)
}
\arguments{
\item{simplify}{logical. If TRUE (default) returns results as characters, otherwise R6 classes.}
}
\description{
Lists wrappers that can wrap an R6 Distribution.
}
\examples{
listWrappers()
listWrappers(TRUE)
}
