% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_helpers.R
\name{makeUniqueDistributions}
\alias{makeUniqueDistributions}
\title{De-Duplicate Distributions}
\usage{
makeUniqueDistributions(distlist)
}
\arguments{
\item{distlist}{list of Distributions.}
}
\description{
From a list of Distributions with the same short_name, suffix each with a consecutive
number so that the names are no longer duplicated.
}
\examples{
makeUniqueDistributions(list(Binomial$new(), Binomial$new()))
}
