% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{max.SetInterval}
\alias{max.SetInterval}
\title{SetInterval Maximum Accessor}
\description{
Returns the SetInterval maximum.
}
\details{
This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $max()
}

\seealso{
\code{\link{SetInterval}}
}
