% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{median.Distribution}
\alias{median.Distribution}
\title{Median of a Distribution}
\usage{
\method{median}{Distribution}(x, na.rm = NULL, ...)
}
\arguments{
\item{x}{Distribution.}

\item{na.rm}{ignored, added for consistency with S3 generic.}

\item{...}{ignored, added for consistency with S3 generic.}
}
\value{
Quantile function evaluated at 0.5 as a numeric.
}
\description{
Median of a distribution assuming quantile is provided.
}
\details{
The median is computed as the quantile function evaluated at 0.5.
If the quantile is not found in the distribution (analytically or numerically), returns error.
}
\section{R6 Usage}{
 $median()
}

\seealso{
\code{\link{quantile.Distribution}}
}
