% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{merge.ParameterSet}
\alias{merge.ParameterSet}
\title{Combine ParameterSets}
\usage{
\method{merge}{ParameterSet}(x, y, ...)
}
\arguments{
\item{x}{ParameterSet}

\item{y}{ParameterSet}

\item{...}{ParameterSets}
}
\description{
merge dispatch method to combine parameter sets by rows.
}
\section{R6 Usage}{
 $merge(y, ...)
}

\seealso{
\code{\link{ParameterSet}}
}
