% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{mgf}
\alias{mgf}
\title{Moment Generating Function}
\value{
\code{mgf} gives the moment generating function evaluated at t
}
\description{
Moment Generating Function
}
\section{Usage}{
 $mgf(t)
}

