% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\name{Naturals}
\alias{Naturals}
\title{Set of Natural Numbers}
\description{
The mathematical set of natural numbers.
}
\section{Constructor Arguments}{

 \tabular{lll}{
   \strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
   \code{dim = 1} \tab numeric \tab Dimension of the set. \cr
}
}

\seealso{
\code{\link{listSpecialSets}}
}
