% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDistribution_NegBinomal.R
\name{NegativeBinomial}
\alias{NegativeBinomial}
\title{Negative Binomial Distribution}
\description{
Mathematical and statistical functions for the Negative Binomial distribution parameterised
with size and prob or \eqn{qprob = 1 - prob}. The size-prob Negative Binomial distribution is defined
by the pmf,
\deqn{f(x) = (x + n - 1)C(n - 1) p^n (1 - p)^x}
where \eqn{n = 1,2,...} is the size parameter and \eqn{0 \le p \le 1} is the prob parameter.
}
\section{Constructor}{
 NegativeBinomial$new(size = 10, prob = 0.5, decorators = NULL, verbose = FALSE)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{size} \tab numeric \tab number of failures. \cr
\code{prob} \tab numeric \tab probability of success. \cr
\code{qprob} \tab numeric \tab probability of failure. \cr
\code{decorators} \tab Decorator \tab decorators to add functionality. \cr
\code{verbose} \tab logical \tab if TRUE parameterisation messages produced.
}
}

\section{Constructor Details}{

The Negative Binomial distribution with size = n (positive) and
 prob = p (0 <= p <= 1) has density function defined as
,
where n is the number of successes (size), p is the probability of each
success (prob) and x is interpreted as the number of failures before
the n-th success.If \code{qprob} is given then \code{prob} is ignored. This quantifies
the probability of having x failuers in a sequence of independent and identical
Bernoulli before the n-th success is obtained.
Compared with a Binomial distribution, which counts the number of successes in a
fixed number of trials, a Negative Binomial distribution pre-specifies the target
number of successes and counts the number of failures required to reach this
target number. A NegBinomial(1,p) is the same as a Geometric(p).
}

\section{Statistical Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{mean()} \tab \code{\link{mean.Distribution}} \cr
  \code{var()} \tab \code{\link{var}} \cr
  \code{skewness()} \tab \code{\link{skewness}} \cr
  \code{kurtosis(excess = TRUE)} \tab \code{\link{kurtosis}} \cr
  \code{mgf(t)} \tab \code{\link{mgf}} \cr
  \code{pgf(z)} \tab \code{\link{pgf}} \cr
  \code{cf(t)} \tab \code{\link{cf}} \cr
  \code{sd()} \tab \code{\link{sd}} \cr
  \code{median()} \tab \code{\link{median.Distribution}} \cr
  \code{iqr()} \tab \code{\link{iqr}} \cr
  }
}

\section{Public Variables}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Return} \cr
  \code{name} \tab Name of distribution. \cr
  \code{short_name} \tab Id of distribution. \cr
  \code{description} \tab Brief description of distribution. \cr
  \code{traits} \tab List: type, valueSupport, variateForm.
  }
}

\section{Accessor Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{decorators()} \tab \code{\link{decorators}} \cr
  \code{valueSupport()} \tab \code{\link{valueSupport}} \cr
  \code{variateForm()} \tab \code{\link{variateForm}} \cr
  \code{type()} \tab \code{\link{type}} \cr
  \code{properties()} \tab \code{\link{properties}} \cr
  \code{support()} \tab \code{\link{support}} \cr
  \code{distrDomain()} \tab \code{\link{distrDomain}} \cr
  \code{symmetry()} \tab \code{\link{symmetry}} \cr
  \code{sup()}  \tab \code{\link{sup}} \cr
  \code{inf()} \tab \code{\link{inf}} \cr
  \code{dmax()}  \tab \code{\link{dmax}} \cr
  \code{dmin()} \tab \code{\link{dmin}} \cr
  \code{skewnessType()} \tab \code{\link{skewnessType}} \cr
  \code{kurtosisType()} \tab \code{\link{kurtosisType}} \cr
  }
}

\section{p/d/q/r Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{pdf(x1, ..., log = FALSE)} \tab \code{\link{pdf}} \cr
  \code{cdf(x1, ..., lower.tail = TRUE, log.p = FALSE)} \tab \code{\link{cdf}}\cr
  \code{quantile(p, ..., lower.tail = TRUE, log.p = FALSE)} \tab \code{\link{quantile.Distribution}} \cr
  \code{rand(n)} \tab \code{\link{rand}}
 }
}

\section{Parameter Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{parameters(id)} \tab \code{\link{parameters}} \cr
  \code{getParameterValue(id, error = "warn")}  \tab \code{\link{getParameterValue}} \cr
  \code{setParameterValue(lst, error = "warn")} \tab \code{\link{setParameterValue}} \cr
}
}

\section{Validation Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{liesInSupport(x, all = TRUE, bound = FALSE)} \tab \code{\link{liesInSupport}} \cr
  \code{liesInType(x, all = TRUE)} \tab \code{\link{liesInType}} \cr
  \code{liesInDistrDomain(x, all = TRUE)} \tab \code{\link{liesInDistrDomain}} \cr
}
}

\section{Representation Methods}{

\tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{strprint()} \tab \code{\link[RSmisc]{strprint}} \cr
  \code{print()} \tab \code{\link[base]{print}} \cr
  \code{summary(full = T)} \tab \code{\link{summary.Distribution}} \cr
  \code{plot()} \tab Coming Soon. \cr
  \code{qqplot()} \tab Coming Soon. \cr
}
}

