% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\name{NegReals}
\alias{NegReals}
\title{Set of Negative Reals}
\arguments{
\item{dim}{numeric dimension of set.}

\item{zero}{logical; if FALSE (default) zero is not included in the set}
}
\value{
Returns an R6 object of class NegReals.
}
\description{
The mathematical set of negative reals.
}
\section{Usage}{
 NegReals$new(dim = 1)
}

\seealso{
\code{\link{Naturals}}, \code{\link{PosNaturals}}, \code{\link{Integers}},
  \code{\link{PosIntegers}}, \code{\link{NegIntegers}}, \code{\link{Rationals}}, \code{\link{PosRationals}},
  \code{\link{NegRationals}}, \code{\link{Reals}}, \code{\link{PosReals}}, \code{\link{NegReals}},
  \code{\link{ExtendedReals}}, \code{\link{Complex}}.
}
