% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Normal.R
\name{Normal}
\alias{Normal}
\title{Normal Distribution}
\description{
Mathematical and statistical functions for the Normal distribution parameterised
with mean and variance, sd or precision.
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{mean} \tab numeric \tab mean, location parameter. \cr
\code{var} \tab numeric \tab variance, squared scale parameter. \cr
\code{sd} \tab numeric \tab standard deviation, scale parameter. \cr
\code{precision} \tab numeric \tab precision, squared scale parameter. \cr
\code{decorators} \tab Decorator \tab decorators to add functionality. See details. \cr
}
}

\section{Constructor Details}{
 The Normal distribution can either be parameterised with variance,
standard deviation or precision. If none are provided then var parameterisation is used with var = 1.
If multiple are provided then parameterisatin takes the hierarchy: var, sd, prec.
sd is defined by
\deqn{sd = var^2}
prec is defined by
\deqn{prec = var^-1}
}

\section{Public Methods}{

 \tabular{lrr}{
  \strong{Method} \tab \strong{Return Type} \tab \strong{Details} \cr
  \code{pdf(x1, log = FALSE)} \tab character \tab Evaluates density at x1. \cr
  \code{cdf(x1, lower.tail = TRUE, log.p = FALSE)} \tab numeric \tab Evaluates distribution function at x1. \cr
  \code{quantile(p, lower.tail = TRUE, log.p = FALSE)} \tab numeric \tab Evalutes inverse distribution at p.  \cr
  \code{rand(n)} \tab numeric \tab Randomly generates n samples from the distribution.  \cr
  \code{normectation()} \tab numeric \tab normectation.  \cr
  \code{var()} \tab numeric \tab Variance.  \cr
  \code{skewness()} \tab numeric \tab Skewness. \cr
  \code{kurtosis(excess = TRUE)} \tab numeric \tab Kurtosis. Kurtosis - 3 if excess = TRUE. \cr
  \code{entropy(base = 2)} \tab numeric \tab Entropy. Shannon if base = 2. \cr
  \code{mode()} \tab numeric \tab Mode. \cr
  \code{mgf(t)} \tab numeric \tab Evaluates moment generating function at t. \cr
  \code{cf(t)} \tab numeric \tab Evaluates characteristic function at t. \cr
  \code{survival(x1, log.p = FALSE)} \tab numeric \tab Evaluates survival function at x1. \cr
  \code{hazard(x1)} \tab numeric \tab Evaluates hazard function at x1. \cr
  \code{cumHazard(x1)} \tab numeric \tab Evaluates cumulative hazard function at x1. \cr
  }
}

\section{Public Variables}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Return} \cr
  \code{name} \tab Name of distribution. \cr
  \code{short_name} \tab Id of distribution. \cr
  \code{description} \tab Brief description of distribution. \cr
  \code{traits} \tab List: type, valueSupport, variateForm.
  }
}

\section{Accessor Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{decorators()} \tab \code{\link{decorators}} \cr
  \code{valueSupport()} \tab \code{\link{valueSupport}} \cr
  \code{variateForm()} \tab \code{\link{variateForm}} \cr
  \code{type()} \tab \code{\link{type}} \cr
  \code{properties()} \tab \code{\link{properties}} \cr
  \code{support()} \tab \code{\link{support}} \cr
  \code{distrDomain()} \tab \code{\link{distrDomain}} \cr
  \code{symmetry()} \tab \code{\link{symmetry}} \cr
  \code{sup()}  \tab \code{\link{sup}} \cr
  \code{inf()} \tab \code{\link{inf}} \cr
  \code{skewnessType()} \tab \code{\link{skewnessType}} \cr
  \code{kurtosisType()} \tab \code{\link{kurtosisType}} \cr
  }
}

