% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_helpers.R
\name{operation}
\alias{operation}
\title{Symbolic Operations for SetInterval}
\usage{
operation(unicode,...)
}
\arguments{
\item{unicode}{unicode symbol for the operation.}

\item{...}{sets and/or intervals to combine via the operation.}
}
\value{
An R6 object of class SetInterval.
}
\description{
Operations for SetInterval objects and subclasses, symbolic only.
}
\details{
Generally not recommended to use this function directly but instead
  via \code{\link{product}} or \code{\link{union}}.
}
\seealso{
\code{\link{SetInterval}}.
}
