% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{pdf}
\alias{pdf}
\title{Probability Density/Mass Function}
\usage{
pdf(object, x1, ..., log = FALSE, simplify = TRUE)
}
\arguments{
\item{object}{Distribution.}

\item{x1}{vector of numerics to evaluate function at.}

\item{...}{additional arguments.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{simplify}{if TRUE (default) returns results in simplest form (vector or data.table) otherwise as data.table.}
}
\value{
Probability density function evaluated at given points as either a numeric if \code{simplify} is TRUE
or as a data.table.
}
\description{
Returns the probability density/mass function for continuous/discrete (or mixture)
distributions evaluated at a given point.
}
