% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{pdfPNorm}
\alias{pdfPNorm}
\title{Probability Density Function P-Norm}
\usage{
pdfPNorm(object, p = 2, lower = NULL, upper = NULL)
}
\arguments{
\item{object}{Distribution.}

\item{p}{p-norm to calculate.}

\item{lower}{lower limit for integration, default is infimum.}

\item{upper}{upper limit for integration, default is supremum.}
}
\description{
The p-norm of the pdf evaluated between given limits or over the whole support.
}
\details{
The p-norm of the pdf is defined by
\deqn{\int_a^b |f_X|^p d\mu)^1/p}
where X is the distribution, \eqn{f_X} is the pdf and a,b are the limits of integration.

Returns NULL if distribution is not continuous.

If an analytic expression isn't available, returns error. To impute a numerical expression, use the
\code{\link{ExoticStatistics}} decorator.
}
\section{R6 Usage}{
 $pdfPNorm(object, p = 2, lower = NULL, upper = NULL)
}

\seealso{
\code{\link{ExoticStatistics}} and \code{\link{decorate}}
}
