% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Kernel.R
\name{pdfSquared2Norm}
\alias{pdfSquared2Norm}
\title{Squared Probability Density Function 2-Norm}
\usage{
pdfSquared2Norm(object, x = 0)
}
\arguments{
\item{object}{Distribution.}

\item{x}{amount to shift the result.}
}
\value{
Squared 2-norm of pdf evaluated between limits as a numeric.
}
\description{
The squared 2-norm of the pdf evaluated over the whole support by default or given
limits, possibly shifted.
}
