% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Distribution}
\alias{plot.Distribution}
\title{Plot Distribution Functions for a distr6 Object}
\usage{
\method{plot}{Distribution}(x, fun = c("pdf", "cdf"), npoints = 3000,
  plot = TRUE, ask = FALSE, arrange = TRUE, ...)
}
\arguments{
\item{x}{\code{distr6} object.}

\item{fun}{vector of functions to plot, one or more of: "pdf","cdf","quantile", "survival", "hazard", "cumhazard", and "all"; partial matching available.}

\item{npoints}{number of evaluation points.}

\item{plot}{logical; if TRUE (default), figures are displayed in the plot window; otherwise a \code{data.table} of points and calculated values is returned.}

\item{ask}{logical; if TRUE, the user is asked before each plot, see \code{\link[graphics]{par}}.}

\item{arrange}{logical; if TRUE (default), margins are automatically adjusted with \code{\link[graphics]{layout}} to accomodate all plotted functions.}

\item{...}{graphical parameters to be passed through to plotting functions.}
}
\description{
Six plots, which can be selected with \code{fun} are available for discrete and
continuous univarite distributions: pdf, cdf, quantile, survival, hazard and cumulative
hazard. By default, the first two are plotted side by side.
}
\details{
The evaluation points are calculted using inverse transform on a uniform grid between 0 and 1 with
length given by \code{npoints}. Therefore any distribution without an analytical \code{quantile} method
will first need to be imputed with the \code{\link{FunctionImputation}} decorator.

The order that the functions are supplied to \code{fun} determines the order in which they are
plotted. If \code{ask} is \code{TRUE} then \code{arrange} is ignored. Titles, labels and other
parameters are set automatically, note that parameters supplied in \code{...} will be passed to
all plots. For maximum flexibility in plotting, set \code{arrange} and \code{ask} to \code{FALSE}.
}
\examples{
plot(Normal$new())
plot(Normal$new(), lwd = 2, col = "red")
plot(Geometric$new(), ask = TRUE)
plot(Binomial$new(), fun = "all")
x = plot(Gamma$new(), plot = FALSE)

}
\seealso{
\code{\link{lines.Distribution}} for superimposing a distr6 object and \code{\link{listDistributions}}
for plottable distributions.
}
