% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vectordistribution.R
\name{plot.VectorDistribution}
\alias{plot.VectorDistribution}
\title{Plotting Distribution Functions for a VectorDistribution}
\usage{
\method{plot}{VectorDistribution}(vecdist, fun = "pdf", n, ind, cols, ...)
}
\arguments{
\item{vecdist}{\link{VectorDistribution}.}

\item{fun}{function to plot, one of: "pdf","cdf","quantile", "survival", "hazard", "cumhazard".}

\item{n}{\code{integer}. First n distributions in the \link{VectorDistribution} to plot.}

\item{...}{Other parameters passed to \link{plot.Distribution}.}

\item{`ind`}{\code{integer}. Indices of the distributions in the \link{VectorDistribution} to plot. If given then \code{n} is ignored.}

\item{`cols`.}{Vector of colours for plotting the curves. If missing \code{1:9} are used.}
}
\description{
Helper function to more easily plot distributions inside a \link{VectorDistribution}.
}
\details{
If \code{n} and \code{ind} are both missing then all distributions are plotted if there are 10 or less in the vector.
Otherwise the function will error.
}
\examples{
# Plot pdf and cdf of Normal
vd = VectorDistribution$new(list(Normal$new(), Normal$new(mean = 2)))
plot(vd)

plot(Normal$new())

# Colour both plots red
plot(Normal$new(), col = "red")

# Change the colours of individual plotted functions
plot(Normal$new(), pdf_col = "red", cdf_col = "green")

# Interactive plotting in order - par still works here
plot(Geometric$new(), fun = "all", ask = TRUE, pdf_col = "black",
  cdf_col = "red", quantile_col = "blue", survival_col = "purple",
  hazard_col = "brown", cumhazard_col = "yellow")

# Return plotting structure
x = plot(Gamma$new(), plot = FALSE)

}
\seealso{
\link{plot.Distribution}
}
