% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_helpers.R
\name{power}
\alias{power}
\alias{^.SetInterval}
\title{Symbolic Exponentiation for SetInterval}
\usage{
power(x, power)

x^power
}
\arguments{
\item{x}{set or interval}

\item{power}{power}
}
\value{
An R6 object of class SetInterval.
}
\description{
Makes a symbolic representation for the exponentiation of a given
  set or interval.
}
\details{
This does not calculate the value of exponentiation but
  is just a symbolic representation using unicode.
}
