% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{power.SetInterval}
\alias{power.SetInterval}
\alias{^.SetInterval}
\title{Symbolic Exponentiation for SetInterval}
\usage{
power.SetInterval(x, power)

\method{^}{SetInterval}(x, power)
}
\arguments{
\item{x}{SetInterval}

\item{power}{power to raise SetInterval to}
}
\description{
Makes a symbolic representation for the exponentiation of a given set/interval.
}
\details{
This does not calculate the exponentiation but
  is just a symbolic representation using unicode.
}
\examples{
PosNaturals$new() ^ 2
power.SetInterval(Reals$new(), 3)

}
\seealso{
\code{\link{product.SetInterval}}, \code{\link{union.SetInterval}},
\code{\link{complement.SetInterval}}
}
