% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{prec}
\alias{prec}
\title{Precision of a Distribution}
\usage{
prec(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
Reciprocal of variance as a numeric.
}
\description{
Precision of a distribution assuming variance is provided.
}
