% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_helpers.R
\name{product}
\alias{product}
\alias{*.SetInterval}
\title{Symbolic Cartesian Product for SetInterval}
\usage{
product(...)

x*y
}
\arguments{
\item{...}{sets and/or intervals to take the cartesian product of.}
}
\value{
An R6 object of class SetInterval.
}
\description{
Makes a symbolic representation for the cartesian product of sets/intervals.
}
\details{
This does not calculate the cartesian product of the arguments but
  is just a symbolic representation using unicode.
}
\seealso{
\code{\link{union}} for the union of two or more intervals/sets.
}
