% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{product.SetInterval}
\alias{product.SetInterval}
\alias{*.SetInterval}
\title{Symbolic Cartesian Product for SetInterval}
\usage{
product.SetInterval(...)

\method{*}{SetInterval}(x, y)
}
\arguments{
\item{x}{SetInterval}

\item{y}{SetInterval}

\item{...}{SetIntervals to take the cartesian product of.}
}
\value{
Returns an R6 object inheriting from SetInterval.
}
\description{
Makes a symbolic representation for the cartesian product of sets/intervals.
}
\details{
This does not calculate the cartesian product of the arguments but
  is just a symbolic representation using unicode.
}
\examples{
PosNaturals$new() * Reals$new()
product.SetInterval(PosNaturals$new(), Reals$new())

}
\seealso{
\code{\link{union.SetInterval}}, \code{\link{complement.SetInterval}},
\code{\link{power.SetInterval}}
}
