% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{properties}
\alias{properties}
\title{Properties Accessor}
\usage{
properties(object)
}
\arguments{
\item{object}{distribution.}
}
\description{
Returns the scientific type of the distribution.
}
\section{R6 Usage}{
 $properties()
}

\seealso{
\code{\link{SetInterval}}
}
