% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{quantile.Distribution}
\alias{quantile.Distribution}
\title{Inverse Cumulative Distribution Function}
\usage{
\method{quantile}{Distribution}(
  x,
  ...,
  lower.tail = TRUE,
  log.p = FALSE,
  simplify = TRUE,
  data = NULL
)
}
\arguments{
\item{x}{Distribution.}

\item{...}{additional arguments.}

\item{lower.tail}{logical; if TRUE, probabilities p are given as log(p).}

\item{log.p}{logical; if TRUE then \eqn{q_X(exp(p))} is returned.}

\item{simplify}{if TRUE (default) returns results in simplest form (vector or data.table) otherwise as data.table.}

\item{p}{vector of probabilities to evaluate function at.}
}
\value{
Inverse cumulative distribution function evaluated at given points as either a numeric if \code{simplify} is TRUE
or as a data.table.
}
\description{
Returns the inverse cumulative distribution, aka quantile, function for a distribution
evaluated at a given point between 0 and 1.
}
