% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{rand}
\alias{rand}
\title{Random Simulation Function}
\usage{
rand(object, n, simplify = TRUE)
}
\arguments{
\item{object}{Distribution.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{simplify}{if TRUE (default) returns results in simplest form (vector or data.table) otherwise as data.table.}
}
\value{
Simulated draws from the distribution as either a numeric if \code{simplify} is TRUE
or as a data.table.
}
\description{
Returns a given number of points sampled from the distribution.
}
