% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{rbind.ParameterSet}
\alias{rbind.ParameterSet}
\title{Combine ParameterSets}
\usage{
\method{rbind}{ParameterSet}(..., ...)
}
\arguments{
\item{...}{ParameterSets}
}
\description{
rbind dispatch method to combine parameter sets by rows.
}
\section{R6 Usage}{
 $rbind(...)
}

\seealso{
\code{\link{ParameterSet}}
}
