% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_Scale.R
\name{Scale}
\alias{Scale}
\title{Distribution Scale Wrapper}
\description{
Scales a distribution to a given mean and standard deviation. By default the distribution
is centered (\code{mean} = 0) with unit variance (\code{sd} = 1)
}
\section{Constructor Arguments}{

\tabular{ll}{
\code{distribution} \tab Distribution object. \cr
\code{mean} \tab desired mean after distribution shift. \cr
\code{sd} \tab desired standard deviation after distribution scale.
}
}

\examples{
Scale$new(Binomial$new())

}
\seealso{
See \code{\link{DistributionWrapper}} for inherited wrapper methods and see \code{\link{Distribution}}
for a full list of inherited distribution methods.
}
