% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_Set.R
\name{Set}
\alias{Set}
\title{Symbolic Set Object}
\description{
An R6 set object for symbolic representation of mathematical sets.
}
\section{Constructor Arguments}{

 \tabular{lll}{
   \strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
   \code{...} \tab ANY \tab See constructor details. \cr
}
}

\section{Constructor Details}{
 Objects of any class can be passed to the set constructor. If a numeric
vector is passed then a discrete set of integers is created in the range from the minimum to the
maximum input. Otherwise all arguments are parsed as elements in a mathematical set.
}

\seealso{
The parent class \code{\link{SetInterval}} for a full list of inherited methods and variables.
}
