% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{SetInterval}
\alias{SetInterval}
\title{R6 Generalised Class for Symbolic Sets and Intervals}
\format{An \code{\link[R6]{R6}} object.}
\description{
The parent class to sets and intervals in R6.
}
\details{
Whilst this is not an abstract class, direct construction is generally not advised.
  Construction should instead be called on 'Set' or 'Interval'
}
\section{Construction}{
 SetInterval$new(symbol, type, lower, upper, dimension)
}

\section{Constructor Arguments}{

 \tabular{lr}{
symbol \tab character. Representation of set/interval. \cr
type \tab character. Interval type, one of (), (], [), [] \cr
lower \tab numeric. Lower limit of set/interval \cr
upper \tab numeric. Upper limit of set/interval \cr
dimension \tab integer. Dimension of set/interval \cr
}
}

\section{Accessor Methods}{

\tabular{lr}{
\code{lower} \tab numeric. Lower limit of set/interval \cr
\code{upper} \tab numeric. Upper limit of set/interval \cr
\code{type} \tab  character. Closed/open interval type \cr
\code{dimension} \tab integer. Dimension of set/interval \cr
\code{max} \tab  numeric. Maximum of set/interval \cr
\code{min} \tab  numeric. Minimum of set/interval \cr
\code{sup} \tab  numeric. Supremum of set/interval \cr
\code{inf} \tab  numeric. Infimum of set/interval \cr
\code{getSymbol} \tab  character. Symbolic representation of set/interval \cr
}
}

\section{Representation Methods}{

\tabular{lr}{
\code{print} \tab  Prints symbolic representation of set/interval \cr
}
}

\seealso{
\code{\link{Set}} for R6 Set objects and \code{\link{Interval}} for R6 Interval objects.
}
