% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{setOperation}
\alias{setOperation}
\title{Symbolic Operations for SetInterval}
\usage{
setOperation(unicode, sets, lower = NULL, upper = NULL, type = NULL,
  dim = NULL)
}
\arguments{
\item{unicode}{unicode symbol for the setOperation.}

\item{sets}{list of sets and/or intervals to combine via the setOperation.}

\item{lower}{lower bound of new SetInterval}

\item{upper}{upper bound of new SetInterval}

\item{type}{type of new SetInterval}

\item{dim}{dimension of new SetInterval}
}
\value{
An R6 object of class SetInterval.
}
\description{
Operations for SetInterval objects and subclasses, symbolic only.
}
\details{
Generally not recommended to use this function directly but instead
  via one of the implemented operations.
}
\seealso{
\code{\link{product.SetInterval}}, \code{\link{union.SetInterval}}, \code{\link{complement.SetInterval}},
\code{\link{power.SetInterval}}
}
