% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_helpers.R
\name{setSymbol}
\alias{setSymbol}
\title{Unicode Symbol of Special Sets}
\usage{
setSymbol(set)
}
\arguments{
\item{set}{special set}
}
\description{
Gets the unicode symbol for standard mathematical special sets.
}
\details{
Special set should be supplied as a character string.
  See \code{\link{SpecialSet}} for full list of currently supported sets.
}
\seealso{
\code{\link{SpecialSet}}
}
