% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{skewness}
\alias{skewness}
\alias{skewness.Distribution}
\title{Skewness}
\usage{
\method{skewness}{Distribution}(object, ...)
}
\arguments{
\item{object}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Skewness of a distribution
}
\details{
The skewness of a distribution is defined by the third standardised moment of the
distribution,
\deqn{sk_X = E_X[(x - \mu)^3]/\sigma^3}
where E_X is the expectation of distribution X, \eqn{\mu} is the mean of the distribution and \eqn{\sigma} is the
standard deviation of the distribution.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.
}
