% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{skewnessType}
\alias{skewnessType}
\title{Type of Skewness Accessor}
\usage{
skewnessType(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
If the distribution skewness is present in properties, returns the type of skewness
(negative/none/positive), otherwise returns NULL.
}
\section{R6 Usage}{
 $skewnessType()
}

\seealso{
\code{\link{skewness}}, \code{\link{properties}} and \code{\link{kurtosisType}}
}
