% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewType.R
\name{skewType}
\alias{skewType}
\title{Skewness Type}
\usage{
skewType(skew)
}
\arguments{
\item{skew}{numeric.}
}
\value{
Returns one of 'negative skew', 'no skew' or 'positive skew'.
}
\description{
Gets the type of skewness
}
\details{
Skewness is a measure of asymmetry of a distribution.

A distribution can either have negative skew, no skew or positive skew. A symmetric distribution
will always have no skew but the reverse relationship does not always hold.
}
\examples{
skewType(1)
skewType(0)
skewType(-1)
}
\seealso{
\code{\link{skewness}}, \code{\link{exkurtosisType}}
}
