% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{stopwarn}
\alias{stopwarn}
\title{Helper Function for Validation Errors}
\usage{
stopwarn(error = "warn", error.msg)
}
\arguments{
\item{error}{one of 'warn' or 'stop'. Taken as warning if not 'stop'.}

\item{error.msg}{string to parse as error message.}
}
\value{
error if 'error' == 'stop'; NULL and warning otherwise.
}
\description{
Matches error type to stop or warning and produces the respective error and
returns NULL.
}
\examples{

\dontrun{
if(2 > 1) stopwarn("warn", "This number is too big")
if(2 < 3) stopwarn("stop", "This number is too small")
}

}
