% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSmisc_helpers.R
\name{strprint}
\alias{strprint}
\alias{strprint.R6}
\alias{strprint.list}
\title{String Representation of Print}
\usage{
strprint(x,...)

strprint.R6(x, ...)

strprint.list(x, ...)
}
\arguments{
\item{x}{R6 object}

\item{...}{Additional arguments}
}
\description{
Extends S3 dispatch of strprint to lists and any R6 class.
}
\details{
strprint is a suggested method that should be included in all R6 classes to be passed to
methods such as \code{cat}, \code{summary} and \code{print}. Additionally can be used to easily
parse R6 objects into data-frames, see examples.
}
\examples{
ClassExample <- R6::R6Class("ClassExample",public=list(
strprint = function() return("Test"),
print = function() print(self$strprint())))
ce = ClassExample$new()
print(ce)
data.frame(A = strprint(ce))
strprint(ce)
strprint(list(ce,ce))

}
