% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{[.ParameterSet}
\alias{[.ParameterSet}
\title{Extract one or more parameters from a ParameterSet}
\usage{
\method{[}{ParameterSet}(ps, ids)
}
\arguments{
\item{ps}{\link{ParameterSet} from which to extract parameters.}

\item{ids}{ids of parameters to extract, if \code{id} ends with \verb{_} then all parameters starting
with \code{ids_} are extracted and the prefix is ignored, \code{prefix} can be left \code{NULL}.
See examples.}

\item{prefix}{An optional prefix to remove from ids after extraction, assumes \verb{_} follows the
prefix name, i.e. \code{prefix_ids}.}
}
\description{
Used to extract one or more parameters from a constructed \link{ParameterSet} or
\link{ParameterSetCollection}.
}
\examples{
ps <- VectorDistribution$new(
  distribution = "Binomial",
  params = data.table::data.table(prob = c(0.1, 0.6, 0.2), size = c(2, 4, 6))
)$parameters()

ps["Binom1_prob"] # extracts just Binom1_prob
ps["Binom1_prob", prefix = "Binom1"] # extracts Binom1_prob and removes prefix
ps["Binom1_"] # extracts all Binom1 parameters and removes prefix

}
