% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{summary.Distribution}
\alias{summary.Distribution}
\title{Distribution Summary}
\usage{
\method{summary}{Distribution}(object, full = TRUE, ...)
}
\arguments{
\item{object}{Distribution.}

\item{full}{logical; if TRUE (default), gives an extended summary, otherwise brief.}

\item{...}{additional arguments.}
}
\value{
Printed summary of the distribution.
}
\description{
Summary method for distribution objects (and all child classes).
}
\section{R6 Usage}{
 $summary(full = TRUE)
}

\seealso{
\code{\link{Distribution}}
}
