% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{sup}
\alias{sup}
\title{Supremum Accessor}
\usage{
sup(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
Supremum as a numeric.
}
\description{
Returns the distribution supremum as the supremum of the support.
}
\section{R6 Usage}{
 $sup
}

