% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{sup.SetInterval}
\alias{sup.SetInterval}
\title{SetInterval Supremum Accessor}
\description{
Returns the SetInterval supremum.
}
\details{
This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $sup()
}

\seealso{
\code{\link{SetInterval}}
}
