% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{support}
\alias{support}
\title{Support Accessor}
\usage{
support(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
Returns the support of the distribution.
}
\details{
The support of a probability distribution is defined as the interval where the pmf/pdf is
greater than zero,
\deqn{Supp(X) = \{ x \in R: f_X(x) > 0\}}
where f_X is the pmf if distribution X is discrete, otherwise the pdf.
}
\section{R6 Usage}{
 $support()
}

\seealso{
\code{\link{SetInterval}} and \code{\link{properties}}
}
