% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{survival}
\alias{survival}
\title{Survival Function}
\usage{
survival(object, x1, log.p = FALSE)
}
\arguments{
\item{object}{Distribution.}

\item{x1}{Point to evaluate the survival function at.}

\item{log.p}{logical, if TRUE then the (natural) logarithm of the survival function is returned.}
}
\description{
The survival function of a probability distribution is the probability of surviving
after a point x.
}
\details{
The survival function is defined by
\deqn{S_X(x) = P(X \geq x) = 1 - F_X(x) \int_x^{Sup(X)} f_X(x) dx}
where X is the distribution, \eqn{S_X} is the survival function, \eqn{F_X} is the cdf, \eqn{f_X} is the pdf and
Sup(X) is the supremum of the distribution X.

If an analytic expression isn't available, returns error. To impute a numerical expression, use the
\code{\link{ExoticStatistics}} decorator.
}
\section{R6 Usage}{
 $survival(x1, log.p = FALSE)
}

\seealso{
\code{\link{ExoticStatistics}} and \code{\link{decorate}}
}
