% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{survivalAntiDeriv}
\alias{survivalAntiDeriv}
\title{Survival Function Anti-Derivative}
\arguments{
\item{lower}{lower limit for integration, default is infimum.}

\item{upper}{upper limit for integration, default is supremum.}
}
\description{
The anti-derivative of the survival function between given limits or
over the full support.
}
\details{
The survival anti-derivative is defined by
\deqn{as(a, b) = \int_a^b S_X(x) dx}
where X is the distribution, S_X is the survival function of the distribution X and a,b are the
limits of integration.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method.
}
