% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{symmetry}
\alias{symmetry}
\title{Symmetry Accessor}
\usage{
symmetry(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
Returns the distribution symmetry, one of "symmetric" or "asymmetric".
}
\section{R6 Usage}{
 $symmetry()
}

\seealso{
\code{\link{properties}}
}
