% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testContinuous}
\alias{testContinuous}
\alias{checkContinuous}
\alias{assertContinuous}
\alias{isContinuous}
\title{assert/check/test/Continuous}
\usage{
testContinuous(object)

checkContinuous(object)

assertContinuous(object)

isContinuous(object)
}
\arguments{
\item{object}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is continuous.
}
\examples{
testContinuous(Binomial$new()) # FALSE

}
