% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testDistribution}
\alias{testDistribution}
\alias{checkDistribution}
\alias{assertDistribution}
\alias{isDistribution}
\title{assert/check/test/isDistribution}
\usage{
assertDistribution(x)
checkDistribution(x)
testDistribution(x)
isDistribution(x)

checkDistribution(x)

assertDistribution(x)

isDistribution(x)
}
\arguments{
\item{x}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if a given object is an R6 Distribution
}
\examples{
testDistribution(5) # FALSE
testDistribution(Binomial$new()) # TRUE

}
