% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testDistributionFeature}
\alias{testDistributionFeature}
\alias{checkDistributionFeature}
\alias{assertDistributionFeature}
\alias{isDistributionFeature}
\title{assert/check/test/isDistributionFeature}
\usage{
assertDistributionFeature(x, accessor, feature)
checkDistributionFeature(x, accessor, feature)
testDistributionFeature(x, accessor, feature)
isDistributionFeature(x, accessor, feature)
}
\arguments{
\item{x}{Distribution}

\item{accessor}{accessor of property/trait to test}

\item{feature}{feature to check if the property possesses}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test for a given feature in a Distribution
}
\examples{
testDistributionFeature(Binomial$new(), "variateForm", "multivariate") # TRUE
testDistributionFeature(Exponential$new(), "valueSupport", "discrete") # FALSE

}
