% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testKurtosis}
\alias{testKurtosis}
\alias{checkKurtosis}
\alias{assertKurtosis}
\alias{isKurtosis}
\title{assert/check/test/Kurtosis}
\usage{
assertKurtosis(x, type)
checkKurtosis(x, type)
testKurtosis(x, type)
isKurtosis(x, type)

checkKurtosis(x, type)

assertKurtosis(x, type)

isKurtosis(x, type)
}
\arguments{
\item{x}{Distribution}

\item{type}{kurtosis type, leptokurtic/mesokurtic/platykurtic}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test the kurtosis of a Distribution
}
\examples{
testKurtosis(Binomial$new(), "leptokurtic")
testKurtosis(Binomial$new(), "platykurtic")

}
