% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_helpers.R
\name{testLeptokurtic}
\alias{testLeptokurtic}
\alias{checkLeptokurtic}
\alias{assertLeptokurtic}
\alias{isLeptokurtic}
\title{assert/check/test/Leptokurtic}
\usage{
assertLeptokurtic(x)
checkLeptokurtic(x)
testLeptokurtic(x)
isLeptokurtic(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is leptokurtic.
}
\examples{
testLeptokurtic(Binomial$new())

}
