% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_helpers.R
\name{testMatrixvariate}
\alias{testMatrixvariate}
\alias{checkMatrixvariate}
\alias{assertMatrixvariate}
\alias{isMatrixvariate}
\title{assert/check/test/Matrixvariate}
\usage{
assertMatrixvariate(x)
checkMatrixvariate(x)
testMatrixvariate(x)
isMatrixvariate(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is matrixvariate.
}
\examples{
testMatrixvariate(Binomial$new()) # FALSE

}
