% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testMesokurtic}
\alias{testMesokurtic}
\alias{checkMesokurtic}
\alias{assertMesokurtic}
\title{assert/check/test/Mesokurtic}
\usage{
testMesokurtic(object)

checkMesokurtic(object)

assertMesokurtic(object)
}
\arguments{
\item{object}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if Distribution is mesokurtic.
}
\examples{
testMesokurtic(Binomial$new())

}
