% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{testMessage}
\alias{testMessage}
\title{Test if Message Produced}
\usage{
testMessage(expr)
}
\arguments{
\item{expr}{expression to test.}
}
\value{
TRUE if a message is prodcued, FALSE otherwise.
}
\description{
Tests if a given expression produces a message.
}
