% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testMixture}
\alias{testMixture}
\alias{checkMixture}
\alias{assertMixture}
\alias{isMixture}
\title{assert/check/test/Mixture}
\usage{
testMixture(x)

checkMixture(x)

assertMixture(x)

isMixture(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is mixture.
}
\examples{
testMixture(Binomial$new()) # FALSE

}
