% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testMultivariate}
\alias{testMultivariate}
\alias{checkMultivariate}
\alias{assertMultivariate}
\alias{isMultivariate}
\title{assert/check/test/Multivariate}
\usage{
assertMultivariate(x)
checkMultivariate(x)
testMultivariate(x)
isMultivariate(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is multivariate.
}
\examples{
testMultivariate(Binomial$new()) # FALSE

}
