% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testNoSkew}
\alias{testNoSkew}
\alias{checkNoSkew}
\alias{assertNoSkew}
\alias{isNoSkew}
\title{assert/check/test/NoSkew}
\usage{
assertNoSkew(x)
checkNoSkew(x)
testNoSkew(x)
isNoSkew(x)

checkNoSkew(x)

assertNoSkew(x)

isNoSkew(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is no skew.
}
\examples{
testNoSkew(Binomial$new())

}
