% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testPlatykurtic}
\alias{testPlatykurtic}
\alias{checkPlatykurtic}
\alias{assertPlatykurtic}
\alias{isPlatykurtic}
\title{assert/check/test/Platykurtic}
\usage{
testPlatykurtic(x)

checkPlatykurtic(x)

assertPlatykurtic(x)

isPlatykurtic(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is platykurtic.
}
\examples{
testPlatykurtic(Binomial$new())

}
