% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testSymmetry}
\alias{testSymmetry}
\alias{checkSymmetry}
\alias{assertSymmetry}
\alias{isSymmetry}
\title{assert/check/test/Symmetry}
\usage{
testSymmetry(x)

checkSymmetry(x)

assertSymmetry(x)

isSymmetry(c)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is symmetric.
}
\examples{
testSymmetry(Binomial$new()) # FALSE

}
