% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testUnivariate}
\alias{testUnivariate}
\alias{checkUnivariate}
\alias{assertUnivariate}
\alias{isUnivariate}
\title{assert/check/test/Univariate}
\usage{
assertUnivariate(x)
checkUnivariate(x)
testUnivariate(x)
isUnivariate(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is univariate
}
\examples{
testUnivariate(Binomial$new()) # TRUE

}
