% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testValueSupport}
\alias{testValueSupport}
\alias{checkValueSupport}
\alias{assertValueSupport}
\alias{isValueSupport}
\title{assert/check/test/ValueSupport}
\usage{
assertValueSupport(x, type)
checkValueSupport(x, type)
testValueSupport(x, type)
isValueSupport(x, type)

checkValueSupport(x, type)

assertValueSupport(x, type)

isValueSupport(x, type)
}
\arguments{
\item{x}{Distribution}

\item{type}{value support type, continuous/discrete/mixture}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test the value support of a Distribution
}
\examples{
testValueSupport(Binomial$new(), "discrete") # silent
testValueSupport(Exponential$new(), "discrete") # FALSE

}
