% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_helpers.R
\name{testVariateForm}
\alias{testVariateForm}
\alias{checkVariateForm}
\alias{assertVariateForm}
\alias{isVariateForm}
\title{assert/check/test/VariateForm}
\usage{
assertVariateForm(x, type)
checkVariateForm(x, type)
testVariateForm(x, type)
isVariateForm(x, type)
}
\arguments{
\item{x}{Distribution}

\item{type}{variate form type, univariate/multivariate/matrixvariate}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test the variate form of a Distribution
}
\examples{
testVariateForm(Binomial$new(), "univariate") # silent
testVariateForm(Exponential$new(), "multivariate") # FALSE

}
