% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{toproper}
\alias{toproper}
\title{Coerce String to Proper Case}
\usage{
toproper(str, split = " ", fixed = TRUE)
}
\arguments{
\item{str}{String to coerce}

\item{split}{see \code{\link[base]{strsplit}}}

\item{fixed}{see \code{\link[base]{strsplit}}}
}
\description{
Helper function for string coercion to proper case
}
