% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{traits}
\alias{traits}
\title{Traits Accessor}
\usage{
traits(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
List of traits.
}
\description{
Returns the traits of the distribution.
}
\section{R6 Usage}{
 $traits
}

