% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_TruncatedDistribution.R
\name{truncate}
\alias{truncate}
\title{Truncate a Distribution}
\usage{
truncate(x, lower = NULL, upper = NULL)
}
\arguments{
\item{x}{Distribution.}

\item{lower}{lower limit for truncation.}

\item{upper}{upper limit for truncation.}
}
\description{
S3 functionality to truncate an R6 distribution.
}
\seealso{
\link{TruncatedDistribution}
}
