% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{type.SetInterval}
\alias{type.SetInterval}
\title{SetInterval Type Accessor}
\value{
Returns the type of SetInterval, one of "()","(]","[]","[)".
}
\description{
Returns the SetInterval 'type', one of "()","(]","[]","[)".
}
\details{
This is an R6 method only, no S3 dispatch is available.
}
\section{R6 Usage}{
 $type()
}

\seealso{
\code{\link{SetInterval}}
}
