% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_helpers.R
\name{union}
\alias{union}
\alias{+.SetInterval}
\title{Symbolic Union for SetInterval}
\usage{
\code{union(...)}

x+y
}
\arguments{
\item{\code{...}}{sets and/or intervals to take the union of.}
}
\value{
An R6 object of class SetInterval.
}
\description{
Makes a symbolic representation for the union of sets/intervals.
}
\details{
This does not calculate the union of the arguments but
  is just a symbolic representation using unicode.
}
\seealso{
\code{\link{product}} for the cartesian product of two or more intervals/sets.
}
