% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{union.SetInterval}
\alias{union.SetInterval}
\alias{+.SetInterval}
\title{Symbolic union.SetInterval for SetInterval}
\usage{
union.SetInterval(..., dim = 1)

\method{+}{SetInterval}(x, y)
}
\arguments{
\item{...}{sets and/or intervals to take the union.SetInterval of.}

\item{dim}{dimension of new SetInterval.}

\item{x}{distribution}

\item{y}{distribution}
}
\value{
An R6 object of class SetInterval.
}
\description{
Makes a symbolic representation for the union.SetInterval of sets/intervals.
}
\details{
This does not calculate the union.SetInterval of the arguments but
  is just a symbolic representation using unicode.
}
\seealso{
\code{\link{product.SetInterval}} for the cartesian product.SetInterval of two or more intervals/sets.
}
