% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{update.ParameterSet}
\alias{update.ParameterSet}
\title{Updates a ParameterSet}
\usage{
\method{update}{ParameterSet}(object, ...)
}
\arguments{
\item{object}{ParameterSet}

\item{...}{ignored, added for S3 consistency}
}
\description{
Updates parameter in a ParameterSet using \code{updateFunc}s.
}
\details{
In general this method should never need to be called manually by the user as it is internally
called in \code{setParameterValue}.

The method works by cycling through parameters in a \code{ParameterSet} that have non-NA \code{updateFunc}s
and parses these as expressions, thereby updating their values.
}
\section{R6 Usage}{
 $update()
}

\seealso{
\code{\link{ParameterSet}}
}
