% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{var.Distribution}
\alias{var.Distribution}
\title{Numeric Variance of a Distribution}
\usage{
var()
}
\description{
A numerc variance calculation for distributions.
}
\details{
The variance of a probability distribution can be numerically calculated via the generalised
expectation function and the formula
\deqn{var_X = E[X^2] - E[X]^2}

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.
}
