% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{var}
\alias{var}
\title{Distribution Variance}
\usage{
var(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
The variance of a distribution, either calculated analytically if possible otherwise
estimated numerically.
}
\details{
The variance of a distribution is defined by the formula
\deqn{var_X = E[X^2] - E[X]^2}

If an analytic expression isn't available, returns error. To impute a numerical expression, use the
\code{\link{CoreStatistics}} decorator.
}
\section{R6 Usage}{
 $var()
}

\seealso{
\code{\link{CoreStatistics}}, \code{\link{decorate}} and \code{\link{genExp}}.
}
