% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{workingSupport}
\alias{workingSupport}
\title{Approximate Finite Support}
\usage{
workingSupport(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
\CRANpkg{set6} object.
}
\description{
If the distribution has an infinite support then this function calculates
the approximate finite limits by finding the largest small number for which \code{cdf == 0} and the
smallest large number for which \code{cdf == 1}.
}
