% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_ArrayDistribution.R
\name{ArrayDistribution}
\alias{ArrayDistribution}
\title{Product ArrayDistribution}
\description{
A special case joint distribution where each independent distribution is the same
Distribution class but not necessarily with the same parameters.
}
\details{
Exploits the following relationships of independent distributions
\deqn{f_A(X1 = x1,...,XN = xN) = f_X1(x1) * ... * f_XN(xn)}
\deqn{F_A(X1 = x1,...,XN = xN) = F_X1(x1) * ... * F_XN(xn)}
where f_A/F_A is the pdf/cdf of the joint (product) distribution P and X1,...,XN are independent distributions.
}
\section{Constructor}{
 ArrayDistribution$new(distribution, paramList)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{distribution} \tab distribution \tab Distribution to wrap. \cr
\code{paramList} \tab list \tab List of parameters, see examples. \cr
}
}

\section{Public Methods}{

See \code{\link{Distribution}} and \code{\link{DistributionWrapper}}.
}

\examples{
a = ArrayDistribution$new(Binomial, list(list(prob = 0.1, size = 2), list(prob = 0.6, size = 4),
                                        list(prob = 0.2, size = 6)))
a$pdf(x1=1,x2=2,x3=3)
a$cdf(x1=1,x2=2,x3=3)
}
\seealso{
\code{\link{ProductDistribution}}
}
