% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDistribution_BetaNoncentral.R
\name{BetaNoncentral}
\alias{BetaNoncentral}
\title{Noncentral Beta Distribution Class}
\description{
Mathematical and statistical functions for the Noncentral Beta distribution, which
is commonly used as the prior in Bayesian modelling.
}
\details{
The Noncentral Beta distribution   is defined by the
}
\examples{
x = BetaNoncentral$new(shape1 = 2, shape2 = 5, location = 3)

# Update parameters
x$setParameterValue(shape1 = 1)
x$parameters()

# d/p/q/r
x$pdf(5)
x$cdf(5)
x$quantile(0.42)
x$rand(4)

summary(x)

}
