% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Binomial.R
\name{Binomial}
\alias{Binomial}
\title{Binomial Distribution}
\description{
Mathematical and statistical functions for the Binomial distribution parameterised
with probability of success and size (number of trials).
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{size} \tab numeric \tab number of trials, default = 10. \cr
\code{prob} \tab numeric \tab probability of success, default = 0.5. \cr
\code{decorators} \tab Decorator \tab decorators to add functionality. See details. \cr
}
}

\section{Constructor Details}{
 The Binomial distribution is parameterised with size (number of trials)
as an integer and prob (probability of success) as a number between 0 and 1.
}

\section{Public Methods}{

 \tabular{lrr}{
  \strong{Method} \tab \strong{Return Type} \tab \strong{Details} \cr
  \code{pdf(x1, log = FALSE)} \tab character \tab Evaluates density at x1. \cr
  \code{cdf(x1, lower.tail = TRUE, log.p = FALSE)} \tab numeric \tab Evaluates distribution function at x1. \cr
  \code{quantile(p, lower.tail = TRUE, log.p = FALSE)} \tab numeric \tab Evalutes inverse distribution at p.  \cr
  \code{rand(n)} \tab numeric \tab Randomly generates n samples from the distribution.  \cr
  \code{expectation()} \tab numeric \tab Expectation.  \cr
  \code{var()} \tab numeric \tab Variance.  \cr
  \code{skewness()} \tab numeric \tab Skewness. \cr
  \code{kurtosis(excess = TRUE)} \tab numeric \tab Kurtosis. Kurtosis - 3 if excess = TRUE. \cr
  \code{entropy(base = 2)} \tab numeric \tab Entropy. Shannon if base = 2. \cr
  \code{mgf(t)} \tab numeric \tab Evaluates moment generating function at t. \cr
  \code{pgf(t)} \tab numeric \tab Evaluates probability generating function at t. \cr
  \code{cf(t)} \tab numeric \tab Evaluates characteristic function at t. \cr
  \code{survival(x1, log.p = FALSE)} \tab numeric \tab Evaluates survival function at x1. \cr
  \code{hazard(x1)} \tab numeric \tab Evaluates hazard function at t. \cr
  \code{cumHazard(x1)} \tab numeric \tab Evaluates cumulative hazard function at t. \cr
  }
}

\section{Public Variables}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Return} \cr
  \code{name} \tab Name of distribution. \cr
  \code{short_name} \tab Id of distribution. \cr
  \code{description} \tab Brief description of distribution. \cr
  \code{traits} \tab List: type, valueSupport, variateForm.
  }
}

\section{Accessor Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{decorators()} \tab \code{\link{decorators}} \cr
  \code{valueSupport()} \tab \code{\link{valueSupport}} \cr
  \code{variateForm()} \tab \code{\link{variateForm}} \cr
  \code{type()} \tab \code{\link{type}} \cr
  \code{properties()} \tab \code{\link{properties}} \cr
  \code{support()} \tab \code{\link{support}} \cr
  \code{distrDomain()} \tab \code{\link{distrDomain}} \cr
  \code{symmetry()} \tab \code{\link{symmetry}} \cr
  \code{sup()}  \tab \code{\link{sup}} \cr
  \code{inf()} \tab \code{\link{inf}} \cr
  \code{skewnessType()} \tab \code{\link{skewnessType}} \cr
  \code{kurtosisType()} \tab \code{\link{kurtosisType}} \cr
  }
}

