% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Binomial.R
\docType{data}
\name{Binomial}
\alias{Binomial}
\title{Binomial Distribution}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Binomial
}
\description{
Binomial Distribution
}
\keyword{datasets}
