% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{cdfAntiDeriv}
\alias{cdfAntiDeriv}
\title{Cumulative Distribution Function Anti-Derivative}
\usage{
cdfAntiDeriv(object, lower = NULL, upper = NULL)
}
\arguments{
\item{object}{Distribution.}

\item{lower}{lower limit for integration, default is infimum.}

\item{upper}{upper limit for integration, default is supremum.}
}
\value{
Antiderivative of the cdf evaluated between limits as a numeric.
}
\description{
The anti-derivative of the cumulative distribution function between given limits or
over the full support.
}
