% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{cdfPNorm}
\alias{cdfPNorm}
\title{Cumulative Distribution Function P-Norm}
\usage{
cdfPNorm(object, p = 2, lower = NULL, upper = NULL)
}
\arguments{
\item{object}{Distribution.}

\item{p}{p-norm to calculate.}

\item{lower}{lower limit for integration, default is infimum.}

\item{upper}{upper limit for integration, default is supremum.}
}
\value{
Given p-norm of cdf evaluated between limits as a numeric.
}
\description{
The p-norm of the cdf evaluated between given limits or over the whole support.
}
