% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{cf}
\alias{cf}
\title{Characteristic Function}
\usage{
cf(object, t)
}
\arguments{
\item{object}{Distribution.}

\item{t}{integer to evaluate characteristic function at.}
}
\description{
Characteristic function of a distribution
}
\details{
The characteristic function is defined by
\deqn{cf_X(t) = E_X[exp(xti)]}
where X is the distribution and \eqn{E_X} is the expectation of the distribution X.

If an analytic expression isn't available, returns error. To impute a numerical expression, use the
\code{\link{CoreStatistics}} decorator.
}
\section{R6 Usage}{
 $cf(t)
}

\seealso{
\code{\link{CoreStatistics}} and \code{\link{decorate}}
}
