% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{cf}
\alias{cf}
\title{Characteristic Function}
\usage{
cf(object, t, ...)
}
\arguments{
\item{object}{Distribution.}

\item{t}{integer to evaluate characteristic function at.}

\item{...}{Passed to \verb{$genExp}.}
}
\value{
Characteristic function evaluated at t as a numeric.
}
\description{
Characteristic function of a distribution
}
