% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{complement.SetInterval}
\alias{complement.SetInterval}
\alias{-.SetInterval}
\title{Symbolic Complement for SetInterval}
\usage{
complement.SetInterval(...)

\method{-}{SetInterval}(x, y)
}
\arguments{
\item{x}{SetInterval}

\item{y}{SetInterval}

\item{...}{SetIntervals to take the complement of.}
}
\value{
Returns an R6 object inheriting from SetInterval.
}
\description{
Makes a symbolic representation for the complement of sets/intervals.
}
\details{
This does not calculate the complement of the arguments but
is just a symbolic representation using unicode.
}
\examples{
PosNaturals$new() - Interval$new(-1,10)
complement.SetInterval(Reals$new(), Interval$new(10,Inf))

}
\seealso{
\code{\link{product.SetInterval}}, \code{\link{union.SetInterval}},
\code{\link{power.SetInterval}}
}
