% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{cor}
\alias{cor}
\title{Numeric Correlation a Distribution}
\usage{
cor()
}
\description{
A numerc calculation for the correlation of a (multivariate) distribution.
}
\details{
If the distribution is univariate then nothing is returned, otherwise the
correlation is calculated numerically.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.
}
