% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{cor}
\alias{cor}
\title{Numeric Correlation a Distribution}
\usage{
cor(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
A numeric calculation for the correlation of a (multivariate) distribution.
}
\details{
If the distribution is univariate then nothing is returned, otherwise the
correlation is calculated numerically.
}
\section{R6 Usage}{
 $cor()
}

