% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{CoreStatistics}
\alias{CoreStatistics}
\alias{mgf}
\alias{cf}
\alias{pgf}
\alias{iqr}
\alias{entropy}
\alias{skewness}
\alias{skewnessType}
\alias{kurtosis}
\alias{kurtosisType}
\alias{kthmoment}
\title{Core Statistics Methods for Distributions}
\usage{
$new(distribution)

$mgf(t)

$cf(t)

$pgf(t)

$iqr()

$entropy(base = 2)

$skewness()

$skewnessType()

$kurtosis(excess = TRUE)

$kurtosisType()

$kthmoment(k, type = "central")
}
\arguments{
\item{distribution}{distribution object.}

\item{t}{integer. Input for function evaluation.}

\item{base}{integer. Logarithmic base for entropy}

\item{excess}{logical. If TRUE (default) Excess Kurtosis is calculated}

\item{k}{integer. Moment to calculate.}

\item{type}{string. One of "central" or "standard".}
}
\value{
\code{new} constructs an R6 object of class Distribution.

\code{mgf} gives the moment generating function evaluated at t

\code{cf} gives the characteristic function evaluated at t

\code{pgf} gives the probability generating function evaluated at t

\code{iqr} gives the interquartile range of the distribution

\code{entropy} gives the (information) entropy of a distribution,
  default is Shannon entropy (base = 2)

\code{skewness} gives the 3rd standardised moment of a distribution.

\code{skewnessType} is an accessor for the type of Skewness

\code{kurtosis} gives the 4th standardised moment of a distribution.
  Excess (kurtosis - 3) is default.

\code{kurtosisType} is an accessor for the type of Kurtosis

\code{kthmoment} gives the kth central (default) or standardized moment
}
\description{
Added functionality to distribution objects for statistical
  methods that can be considered core but lie outside of the p/d/q/r generation
  functions.
}
\details{
Decorator objects add functionality to the given Distribution object
 by overwriting the object in the Global Environment. They can be specified
 in construction of the Distribution or by constructing the given Decorator.

 Methods act on the distribution and not the constructor therefore method chaining of the form
 \code{CoreStatistics$new(distribution)$iqr()} is not supported but \code{distribution$new(decorator=CoreStatistics)$iqr()} is.

 Generating functions are evaluated at a particular point \code{t} and do not
  give specific analytic generating functions. \code{type} of moment is one of,
  "central" for the kth moment about the mean, or "standard" for the central moment
  standardised by, kthCentralMoment / standard deviation^k.

  Many functions rely on the expectation of a distribution and therefore results may be
  approximate when numerical integration is used.
}
\examples{
X = Binomial$new(decorator = "CoreStatistics")
X$iqr()
X$kurtosis()

X = Binomial$new()
CoreStatistics$new(X)
X$kthmoment(4)
}
\seealso{
\code{\link{ExoticStatistics}} for more available methods.
}
