% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{CoreStatistics}
\alias{CoreStatistics}
\alias{cf}
\alias{pgf}
\alias{iqr}
\alias{entropy}
\alias{skewness}
\alias{kurtosis}
\alias{kthmoment}
\alias{genExp}
\alias{var}
\alias{cov}
\alias{cor}
\alias{mode}
\title{Core Statistical Methods for Distributions}
\arguments{
\item{distribution}{distribution object.}

\item{t}{integer. Input for function evaluation.}

\item{base}{integer. Logarithmic base for entropy}

\item{excess}{logical. If TRUE (default) Excess Kurtosis is calculated}

\item{k}{integer. Moment to calculate.}

\item{type}{string. One of "central" or "standard".}
}
\value{
\code{cf} gives the characteristic function evaluated at t

\code{pgf} gives the probability generating function evaluated at t

\code{iqr} gives the interquartile range of the distribution

\code{entropy} gives the (information) entropy of a distribution,
  default is Shannon entropy (base = 2)

\code{skewness} gives the 3rd standardised moment of a distribution.

\code{kurtosis} gives the 4th standardised moment of a distribution.
  Excess (kurtosis - 3) is default.

\code{kthmoment} gives the kth central (default) or standardized moment

\code{genExp} gives the expectation (default)

\code{var} gives the variance

\code{cov} gives the covariance

\code{cor} gives the correlation

\code{mode} gives the mode
}
\description{
Added functionality to distribution objects for numerical statistical
  methods. Including a generalised expectation function for more complex numerical calculations
}
\details{
Decorator objects add functionality to the given Distribution object
 by copying methods in the decorator environment to the chosen Distribution environment. Use the
 \code{\link{decorate}} function to decorate a Distribution. See the help pages for the individual
 CoreStatistics methods to learn more.

 All methods in this decorator use numerical approximations and therefore better results may be available
 from analytic computations.


 Generating functions are evaluated at a particular point \code{t} and do not
  give specific analytic generating functions. \code{type} of moment is one of,
  "central" for the kth moment about the mean, or "standard" for the central moment
  standardised by, kthCentralMoment / standard deviation^k.
}
\section{Usage}{
 $cf(t)

 $pgf(t)

 $iqr()

 entropy(base = 2)

 $skewness()

 $kurtosis(excess = TRUE)

 $kthmoment(k, type = "central")

 $genExp(trafo)

 $var()

 $cov()

 $cor()

 $mode()
}

\examples{
x = Binomial$new()
decorate(x, CoreStatistics)
x$iqr()

x = Binomial$new(decorators = CoreStatistics)
x$kthmoment(4)
}
\seealso{
\code{\link{decorate}} for the decorate function and \code{\link{ExoticStatistics}} for
more available methods.
}
