% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{CoreStatistics}
\alias{CoreStatistics}
\title{Core Statistical Methods for Distributions}
\description{
Added functionality to distribution objects for numerical statistical
  methods. Including a generalised expectation function for more complex numerical calculations.
}
\details{
Decorator objects add functionality to the given Distribution object
 by copying methods in the decorator environment to the chosen Distribution environment. Use the
 \code{\link{decorate}} function to decorate a Distribution.

 All methods in this decorator use numerical approximations and therefore better results may be available
 from analytic computations.
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{dist} \tab distribution \tab Distribution to decorate. \cr
\code{R62S3} \tab logical \tab If TRUE (default), S3 methods are added for decorators in construction.
}
}

\section{Public Methods}{

\tabular{lll}{
\strong{Method} \tab \strong{Input -> Output} \tab \strong{Details} \cr
\code{mgf(t)} \tab numeric -> numeric \tab Moment generating function evaluated at t. \cr
\code{pgf(t)} \tab numeric -> numeric \tab Probability generating function evaluated at t. \cr
\code{cf(t)} \tab numeric -> numeric \tab Characteristic function evaluated at t. \cr
\code{iqr()} \tab -> numeric \tab Interquartile range of the distribution. \cr
\code{entropy(base = 2)} \tab integer -> numeric \tab Distribution entropy, Shannon by default. \cr
\code{skewness()} \tab -> numeric \tab Third stndardised moment of the distribution. \cr
\code{kurtosis(excess = TRUE)} \tab logical -> numeric \tab Fourth standardised moment of the distribution. \cr
By default excess kurtosis (kurtosis - 3). \cr
\code{kthmoment(type = "central")} \tab character -> numeric \tab Central (default) or standardised kth moment (type = "standard"). \cr
\code{genExp(trafo)} \tab function -> numeric \tab Generalised expectation formula. If trafo = NULL, returns arithmetic mean. \cr
\code{var()} \tab -> numeric \tab Numeric variance of distribution. \cr
\code{cov()} \tab -> numeric \tab Covariance of distribution. \cr
\code{cor()} \tab -> numeric \tab Correlation of distribution. \cr
\code{mode(which = 1)} \tab integer -> numeric \tab Mode of distribution, if which = 1, returns first, otherwise all.
}
}

\examples{
x = Binomial$new()
decorate(x, CoreStatistics, R62S3 = FALSE)
x$iqr()

x = Binomial$new(decorators = CoreStatistics, R62S3 = FALSE)
x$kthmoment(4)
}
\seealso{
\code{\link{DistributionDecorator}}
}
