% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{CoreStatistics}
\alias{CoreStatistics}
\title{Core Statistical Methods for Distributions}
\description{
Added functionality to distribution objects for numerical statistical
  methods. Including a generalised expectation function for more complex numerical calculations.
}
\details{
Decorator objects add functionality to the given Distribution object
 by copying methods in the decorator environment to the chosen Distribution environment. Use the
 \code{\link{decorate}} function to decorate a Distribution.

 All methods in this decorator use numerical approximations and therefore better results may be available
 from analytic computations. See below for the methods added to a distribution after decorating with
 \code{CoreStatistics}.
}
\section{Constructor}{
 CoreStatistics$new(dist)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{dist} \tab distribution \tab Distribution to decorate. \cr
}
}

\section{Added Methods}{

\tabular{lll}{
\strong{Method} \tab \strong{Name} \tab \strong{Link} \cr
\code{mgf(t)} \tab Moment generating function \tab \code{\link{mgf}} \cr
\code{pgf(t)} \tab Probability generating function \tab \code{\link{pgf}} \cr
\code{cf(t)} \tab Characteristic function \tab \code{\link{cf}} \cr
\code{entropy(base = 2)} \tab (Shannon) Entropy \tab \code{\link{entropy}} \cr
\code{skewness()} \tab Skewness \tab \code{\link{skewness}} \cr
\code{kurtosis(excess = TRUE)} \tab Kurtosis \tab \code{\link{kurtosis}} \cr
\code{kthmoment(type = "central")} \tab Kth Moment \tab \code{\link{kthmoment}} \cr
\code{genExp(trafo)} \tab Generalised Expectation \tab \code{\link{genExp}} \cr
\code{mode(which = "all")} \tab Mode \tab \code{\link{mode}} \cr
\code{var()} \tab Variance \tab \code{\link{var}} \cr
\code{cor()} \tab Correlation \tab \code{\link{cor}} \cr
}
}

\examples{
x = Binomial$new()
decorate(x, CoreStatistics)
x$genExp()

x = Binomial$new(decorators = CoreStatistics)
x$kthmoment(4)

}
\seealso{
\code{\link{DistributionDecorator}}, \code{\link{decorate}} and \code{\link{ExoticStatistics}}
}
