% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{correlation}
\alias{correlation}
\title{Distribution Correlation}
\usage{
correlation(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
Either '1' if distribution is univariate or the correlation as a numeric or matrix.
}
\description{
Correlation of a distribution.
}
