% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{cov.Distribution}
\alias{cov.Distribution}
\title{Numeric Covariance a Distribution}
\arguments{
\item{object}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A numeric calculation for the covariance of a (multivariate) distribution.
}
\details{
If the distribution is univariate then the variance is returned, otherwise the
covariance is calculated numerically.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.

export
}
