% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{cov}
\alias{cov}
\title{Numeric Covariance a Distribution}
\usage{
cov(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
A numeric calculation for the covariance of a (multivariate) distribution.
}
\details{
If the distribution is univariate then the variance is returned, otherwise the
covariance is calculated numerically.
}
\section{R6 Usage}{
 $cov()
}

