% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{cov}
\alias{cov}
\title{Distribution Covariance}
\usage{
cov(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
Covariance of a distribution.
}
\details{
The covariance of a distribution is defined by the equation,
\deqn{\sigma_{XY} = E[(X - E[X])(Y - E[Y])], X != Y}
where \eqn{E_X} is the expectation with respect to distribution X.

If the distribution is univariate then returns the variance.

If an analytic expression isn't available, returns error. To impute a numerical expression, use the
\code{\link{CoreStatistics}} decorator.
}
\section{R6 Usage}{
 $cov()
}

\seealso{
\code{\link{CoreStatistics}} and \code{\link{decorate}}
}
