% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{cumHazard}
\alias{cumHazard}
\title{Cumulative Hazard Function}
\usage{
cumHazard(object, x1, log = FALSE)
}
\arguments{
\item{object}{Distribution.}

\item{x1}{Point to evaluate the cumulative hazard function at.}

\item{log}{logical, if TRUE then the (natural) logarithm of the cumulative hazard function is returned.}
}
\value{
Cumulative hazard function as a numeric, natural logarithm returned if \code{log} is TRUE.
}
\description{
The cumulative hazard function of a probability distribution is the anti-derivative of
the hazard function.
}
