% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate}
\title{Decorate Distributions}
\usage{
decorate(distribution, decorators)
}
\arguments{
\item{distribution}{distribution to decorate}

\item{decorators}{list of decorators}
}
\description{
A convenient S3 function to decorate R6 distributions
}
\seealso{
\code{\link{DistributionDecorator}} for the abstract decorator class and
\code{\link{CoreStatistics}}, \code{\link{ExoticStatistics}}, \code{\link{FunctionImputation}} for
available decorators.
}
