% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate}
\title{Decorate Distributions}
\usage{
decorate(distribution, decorators, R62S3 = TRUE)
}
\arguments{
\item{distribution}{distribution to decorate}

\item{decorators}{list of decorators}
}
\description{
Decorate Distributions
}
