% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate}
\title{Decorate Distributions}
\usage{
decorate(distribution, decorators)
}
\arguments{
\item{distribution}{distribution to decorate}

\item{decorators}{list of decorators}
}
\description{
A convenient S3 function to decorate R6 distributions
}
\details{
Decorating is the process of adding methods to classes that are not part of the core
interface. Use \code{listDecorators} to see wich decorators are currently available. The primary
use-cases are to add numeric results when analytic ones are missing, to add complex modelling functions,
to impute missing p/d/q/r functions.

The \code{decorators} parameter should either be a list of decorator classes (i.e. not as strings)
or a single decorator class; see examples.
}
\examples{
B <- Binomial$new()
decorate(B, CoreStatistics)

E <- Exponential$new()
decorate(E, list(CoreStatistics, ExoticStatistics))

}
\seealso{
\code{\link{listDecorators}} for available decorators.
}
