% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_ArrayDistribution.R,
%   R/distr6-deprecated.R
\name{ArrayDistribution}
\alias{ArrayDistribution}
\alias{distr6-deprecated}
\title{Deprecated distr6 Functions and Classes}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\section{Deprecated Functions}{

\tabular{ll}{
\strong{Deprecated} \tab \strong{Replacement} \cr
\code{ArrayDistribution} \tab \code{\link{ProductDistribution}} \cr
}
}

\keyword{internal}
