% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Distribution}
\alias{Distribution}
\title{Generalised Distribution Object}
\description{
A generalised distribution object for defining custom probability distributions
  as well as serving as the parent class to specific, familiar distributions. Common
  mathematical and statistical methods for distributions are defined here with approximate numerical
  calculations (as opposed to analytical results).
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{name} \tab character \tab Full name of distribution. \cr
\code{short_name} \tab character \tab Short name to identify distribution. \cr
\code{type} \tab SetInterval \tab Scientific type. \cr
\code{support} \tab SetInterval \tab Distribution support. See Details. \cr
\code{distrDomain} \tab SetInterval \tab Distribution domain See Details. \cr
\code{symmetric} \tab logical \tab Is distribution symmetric? \cr
\code{pdf} \tab function \tab See Details. \cr
\code{cdf} \tab function \tab See Details. \cr
\code{quantile} \tab function \tab See Details. \cr
\code{rand} \tab function \tab See Details. \cr
\code{parameters} \tab ParameterSet \tab See Details. \cr
\code{decorators} \tab list \tab R6 decorators to add in construction. \cr
\code{valueSupport} \tab character \tab continuous, discrete, mixture. See Details. \cr
\code{variateForm} \tab character \tab univariate, multivariate, matrixvariate. See Details. \cr
\code{description} \tab character \tab short description of distribution. \cr
\code{R62S3} \tab logical \tab if TRUE (default), S3 methods are added for decorators in construction.
}
}

\section{Constructor Details}{
 The primary purpose of the Distribution object is to serve as the parent class
  to all other distributions, therefore all methods are approximate numeric calculations
  and the user may prefer to utilise decorators to improve accuracy.

  \code{type}, \code{support} and \code{distrDomain} should be given as an R6 SetInterval
  object. If none are supplied then the set of Reals is taken to be the type, support and domain
  of the distribution. If only \code{type} is supplied then this is taken to also be the support
  and domain.

  By default, missing \code{pdf}, \code{cdf} and \code{quantile} are not automatically imputed.
  Use the imputation wrappers (see below) to generate these with a selected method.
  The \code{rand} function is automatically generated depending on which of the above are supplied.
  The generation for this is performed according to the hierarchy: quantile -> rand, cdf -> rand, pdf -> rand.

  \code{parameters} should be supplied as a ParameterSet. The distribution parameterisation
  is taken to be whichever parameters are flagged as 'settable', any others in the ParameterSet
  are automatically updated by a given function.

  \code{decorators} is a list of decorators (R6 environments not strings) to decorate the
  Distribution with in construction. Decorators can also be added after construction. See
  \code{\link{DistributionDecorator}} for more details.

  \code{valueSupport} should be one of continuous/discrete/mixture if supplied.
  \code{variateForm} should be one of univariate/multivariate/matrixvariate if supplied.
  If not given these are automatically filled from \code{type} and \code{support}.
}

\section{Accessor Methods}{

 \tabular{lrr}{
  \strong{Method} \tab \strong{Return Type} \tab \strong{Details} \cr
  \code{name()} \tab character \cr
  \code{short_name()} \tab character \cr
  \code{description()} \tab character \cr
  \code{decorators()} \tab character \cr
  \code{traits()} \tab list \cr
  \code{valueSupport()} \tab character \cr
  \code{variateForm()} \tab character \cr
  \code{type()} \tab Set \tab \code{\link{Set}} \cr
  \code{properties()} \tab list \cr
  \code{support()} \tab Set \tab \code{\link{Set}} \cr
  \code{distrDomain()} \tab Set \tab \code{\link{Set}} \cr
  \code{symmetry()} \tab character \cr
  \code{parameters(id,as.df = FALSE)} \tab ParameterSet or data.frame \tab \code{\link{ParameterSet}} \cr
  \code{getParameterValue(id)} \tab numeric \tab \code{\link{ParameterSet}} \cr
  \code{sup()} \tab numeric \tab supremum of distribution \cr
  \code{inf()} \tab numeric \tab infimum of distribution \cr
  }
}

\section{Math/Stats Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Details} \cr
  \code{pdf(x, log = F)} \tab Evaluate density/mass at x \cr
  \code{cdf(q, lower.tail = T, log.p = F)} \tab Evaluate distribution function at q.\cr
  \code{quantile(p, lower.tail = T, log.p = F)} \tab Evaluate quantile function at p \cr
  \code{rand(n)} \tab Simulate n draws from distribution \cr
  \code{expectation(trafo)} \tab Calculate expectation \cr
  \code{var()} \tab Calculate variance \cr
  \code{sd()} \tab Calculate standard deviation \cr
  \code{cov()} \tab Calculate covariance. See Details \cr
  \code{cor()} \tab Calculate correlation. See Details \cr
  \code{median()} \tab Calculate median \cr
  \code{mode(which = 1)} \tab Calculate mode. See Details \cr
 }
}

\section{Other Methods}{

 \tabular{lrr}{
  \strong{Method} \tab \strong{Input -> Output} \tab \strong{Details} \cr
  \code{setParameterValue(lst)} \tab list -> invisible(self) \tab Set parameter value. See \code{\link{ParameterSet}}. \cr
  \code{liesInSupport(x, all = TRUE)} \tab numeric x logical -> logical \tab Does x lie in the support of distribution? See Details. \cr
  \code{liesInType(x)} \tab numeric -> logical \tab Does x lie in the type of distribution? \cr
  \code{liesInDistrDomain(x)} \tab numeric -> logical \tab Does x lie in the domain of distribution? \cr
}
}

\section{Representation Methods}{

\tabular{ll}{
  \strong{Method} \tab \strong{Details} \cr
  \code{strprint()} \tab Character representation of print \cr
  \code{print()} \tab Print method \cr
  \code{summary(full = T)} \tab Summary method \cr
  \code{plot()} \tab Plotting method \cr
  \code{qqplot()} \tab QQ-Plots \cr
}
}

\section{Public Methods Details}{

\code{cov} defaults to \code{var} for univariate distributions and \code{cor} returns NULL.
\code{mode} returns by default the first mode of the distribution where applicable, otherwise a specified
integer or all.

If \code{liesInSupport(x, all = TRUE)} then returns TRUE only if every numeric in vector \code{x} lies
in the support of the distribution, otherwise returns a vector of logicals.
}

\seealso{
See \code{\link{SetInterval}} and \code{\link{SpecialSet}} for details on Sets and
Intervals. See \code{\link{ParameterSet}} for parameter details. See
\code{\link{DistributionDecorator}} for Decorator details. See \code{\link[stats]{Binomial}} for
details on the arguments to \code{pdf}/\code{cdf}/\code{quantile}/\code{rand}.
}
