% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Distribution}
\alias{Distribution}
\title{Generalised Distribution Object}
\description{
A generalised distribution object for defining custom probability distributions
  as well as serving as the parent class to specific, familiar distributions. Common
  mathematical and statistical methods for distributions are defined here with approximate numerical
  calculations (as opposed to analytical results).
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{name} \tab character \tab Full name of distribution. \cr
\code{short_name} \tab character \tab Short name to identify distribution. \cr
\code{type} \tab SetInterval \tab Scientific type. \cr
\code{support} \tab SetInterval \tab Distribution support. See Details. \cr
\code{distrDomain} \tab SetInterval \tab Distribution domain See Details. \cr
\code{symmetric} \tab logical \tab Is distribution symmetric? \cr
\code{pdf} \tab function \tab See Details. \cr
\code{cdf} \tab function \tab See Details. \cr
\code{quantile} \tab function \tab See Details. \cr
\code{rand} \tab function \tab See Details. \cr
\code{parameters} \tab ParameterSet \tab See Details. \cr
\code{decorators} \tab list \tab R6 decorators to add in construction. \cr
\code{valueSupport} \tab character \tab continuous, discrete, mixture. See Details. \cr
\code{variateForm} \tab character \tab univariate, multivariate, matrixvariate. See Details. \cr
\code{description} \tab character \tab short description of distribution. \cr
}
}

\section{Constructor Details}{


  The most basic Distribution object consists of a name and one of pdf/cdf.

  If supplied, \code{type}, \code{support} and \code{distrDomain} should be given as an R6 SetInterval
  object. If none are supplied then the set of Reals is taken to be the type and the dimension is the
  number of formal arguments in the pdf/cdf. If only \code{type} is supplied then this is taken to also
  be the support and domain.

  By default, missing \code{pdf}, \code{cdf}, \code{quantile} and \code{rand} are not automatically imputed.
  Use the \code{\link{FunctionImputation}} decorator to generate these.

  See \code{\link{ParameterSet}} for more details on construction of a ParameterSet.

  \code{decorators} is an optional list of decorators (R6 environments not strings) to decorate the
  Distribution in construction. Decorators can also be added after construction. See
  \code{\link{DistributionDecorator}} and \code{\link{decorate}} for more details.

  \code{valueSupport} should be one of continuous/discrete/mixture if supplied.
  \code{variateForm} should be one of univariate/multivariate/matrixvariate if supplied.
  If not given these are automatically filled from \code{type} and \code{support}.
}

\section{Public Variables}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Return} \cr
  \code{name} \tab Name of distribution. \cr
  \code{short_name} \tab Id of distribution. \cr
  \code{description} \tab Brief description of distribution. \cr
  \code{traits} \tab List: type, valueSupport, variateForm.
  }
}

\section{Public Methods}{

 \tabular{ll}{
  \strong{Accessor Methods} \tab \strong{Link} \cr
  \code{decorators()} \tab \code{\link{decorators}} \cr
  \code{valueSupport()} \tab \code{\link{valueSupport}} \cr
  \code{variateForm()} \tab \code{\link{variateForm}} \cr
  \code{type()} \tab \code{\link{type}} \cr
  \code{properties()} \tab \code{\link{properties}} \cr
  \code{support()} \tab \code{\link{support}} \cr
  \code{distrDomain()} \tab \code{\link{distrDomain}} \cr
  \code{symmetry()} \tab \code{\link{symmetry}} \cr
  \code{sup()}  \tab \code{\link{sup}} \cr
  \code{inf()} \tab \code{\link{inf}} \cr
  \code{dmax()}  \tab \code{\link{dmax}} \cr
  \code{dmin()} \tab \code{\link{dmin}} \cr
  \code{skewnessType()} \tab \code{\link{skewnessType}} \cr
  \code{kurtosisType()} \tab \code{\link{kurtosisType}} \cr
  \tab \cr \tab \cr \tab \cr
  \strong{p/d/q/r Methods} \tab \strong{Link} \cr
  \code{pdf(x1, ..., log = FALSE, simplify = TRUE)} \tab \code{\link{pdf}} \cr
  \code{cdf(x1, ..., lower.tail = TRUE, log.p = FALSE, simplify = TRUE)} \tab \code{\link{cdf}}\cr
  \code{quantile(p, ..., lower.tail = TRUE, log.p = FALSE, simplify = TRUE)} \tab \code{\link{quantile.Distribution}} \cr
  \code{rand(n, simplify = TRUE)} \tab \code{\link{rand}} \cr
  \tab \cr \tab \cr \tab \cr
  \strong{Parameter Methods} \tab \strong{Link} \cr
  \code{parameters(id)} \tab \code{\link{parameters}} \cr
  \code{getParameterValue(id, error = "warn")}  \tab \code{\link{getParameterValue}} \cr
  \code{setParameterValue(lst, error = "warn")} \tab \code{\link{setParameterValue}} \cr
  \tab \cr \tab \cr \tab \cr
  \strong{Validation Methods} \tab \strong{Link} \cr
  \code{liesInSupport(x, all = TRUE, bound = FALSE)} \tab \code{\link{liesInSupport}} \cr
  \code{liesInType(x, all = TRUE)} \tab \code{\link{liesInType}} \cr
  \code{liesInDistrDomain(x, all = TRUE)} \tab \code{\link{liesInDistrDomain}} \cr
  \tab \cr \tab \cr \tab \cr
  \strong{Representation Methods} \tab \strong{Link} \cr
  \code{strprint()} \tab \code{\link{strprint}} \cr
  \code{print()} \tab \code{\link[base]{print}} \cr
  \code{summary(full = T)} \tab \code{\link{summary.Distribution}} \cr
  \code{plot()} \tab Coming Soon. \cr
  \code{qqplot()} \tab Coming Soon. \cr
  }
}

\seealso{
See \code{\link{SetInterval}} and \code{\link{SpecialSet}} for details on Sets and
Intervals. See \code{\link{ParameterSet}} for parameter details. See
\code{\link{DistributionDecorator}} for Decorator details.
}
