% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator.R
\name{DistributionDecorator}
\alias{DistributionDecorator}
\title{Abstract Decorator for Distribution}
\description{
An abstract decorator object for distributions to enhance functionality.
}
\details{
This is an abstract class that cannot be directly constructed.
  See \code{\link{decorate}} for the function to decorate Distributions.
}
\seealso{
\code{\link{decorate}} for the decorating function. \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for available decorators. \code{\link{Distribution}} for decoratable
distributions.
}
