% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{DistributionWrapper}
\alias{DistributionWrapper}
\title{Abstract Wrapper for Distributions}
\description{
An R6 abstract wrapper class with methods implemented for child classes.
}
\details{
Cannot be implemented directly.
}
\section{Public Methods}{

\tabular{ll}{
\code{wrappedModels(model = NULL)} \tab Get internally wrapped model by name. Or list of all models if model = NULL. \cr
\code{setParameterValue(lst)} \tab Sets the value of an internal models parameter. See Details.
}
}

\section{Public Method Arguments}{

\tabular{ll}{
\code{model} \tab Wrapped model to access. \cr
\code{lst} \tab list. Names are parameter IDs, values are values to set parameters.
}
}

\section{Public Methods Details}{

Wrapped models overload the minimum number of functions necessary. The abstract class overloads
the \code{setParameterValue} method only as wrapped models alter paramater IDs. After wrapping
a model, parameter IDs are altered by prefixing the ID with "model_". For example wrapping Model1 with
a parameter 'param1' results in 'Model1_param1'.
}

\seealso{
\code{\link{TruncatedDistribution}}, \code{\link{HuberizedDistribution}}
}
