% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{dmin}
\alias{dmin}
\title{Distribution Minimum Accessor}
\usage{
dmin(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
Minimum as a numeric.
}
\description{
Returns the distribution minimum as the minimum of the support. If the support is
not bounded below then minimum is given by
\deqn{minimum = infimum + 1.1e-15}
}
\section{R6 Usage}{
 $dmin
}

