% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{entropy}
\alias{entropy}
\title{Entropy}
\arguments{
\item{object}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{base}{base of the entropy logarithm, default = 2 (Shannon entropy)}
}
\description{
(Information) Entropy of a distribution
}
\details{
The entropy of a distribution is defined by
\deqn{- sum f_X * log(f_X)}
where f_X is the pdf of distribution X. The base of the logarithm of the equation determines the
type of entropy computed. By default we use base 2 to compute entropy in 'Shannons' or 'bits'.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method.
}
