% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{entropy}
\alias{entropy}
\title{Distribution Entropy}
\usage{
entropy(object, base = 2, ...)
}
\arguments{
\item{object}{Distribution.}

\item{base}{base of the entropy logarithm, default = 2 (Shannon entropy)}

\item{...}{Passed to \verb{$genExp}.}
}
\value{
Entropy with given base as a numeric.
}
\description{
(Information) Entropy of a distribution
}
