% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_helpers.R
\name{exkurtosisType}
\alias{exkurtosisType}
\title{Kurtosis Type}
\usage{
exkurtosisType(kurtosis)
}
\arguments{
\item{kurtosis}{numeric.}
}
\description{
Gets the type of (excess) kurtosis
}
\examples{
exkurtosisType(1)
}
