% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{ExoticStatistics}
\alias{ExoticStatistics}
\title{Exotic Statistics Methods for Distributions}
\arguments{
\item{distribution}{distribution object.}
}
\value{
\code{ExoticStatistics$new} constructs an R6 object of class Distribution.
}
\description{
Further functionality to distribution objects for statistical
  methods that can be considered more exotic than core, such as survival modelling
  and p-norms.
}
\details{
Decorator objects add functionality to the given Distribution object
 by overwriting the object in the Global Environment. They can be specified
 in construction of the Distribution or by constructing the given Decorator.

 Methods act on the distribution and not the constructor therefore method chaining of the form
 \code{ExoticStatistics$new(distribution)$hazard(1)} is not supported but \code{distribution$new(decorator=ExoticStatistics)$hazard(1)} is.
}
\section{Usage}{
 ExoticStatistics$new(distribution)
}

\examples{
\dontrun{
X = Binomial$new(decorator = "ExoticStatistics")
X$survival(1)
X$pdfPNorm()
}

\dontrun{
X = Binomial$new()
ExoticStatistics$new(X)
X$pdfPNorm(4)
}
}
\seealso{
\code{\link{CoreStatistics}} for more available methods.
}
