% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{ExoticStatistics}
\alias{ExoticStatistics}
\title{Exotic Statistical Methods for Distributions}
\description{
Further functionality to distribution objects for numerical statistical
  methods that can be considered more exotic than core, such as survival modelling
  and p-norms.
}
\details{
Decorator objects add functionality to the given Distribution object
 by copying methods in the decorator environment to the chosen Distribution environment. Use the
 \code{\link{decorate}} function to decorate a Distribution.

 All methods in this decorator use numerical approximations and therefore better results may be available
 from analytic computations. See below for the methods added to a distribution after decorating with
 \code{ExoticStatistics}.
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{dist} \tab distribution \tab Distribution to decorate. \cr
}
}

\section{Added Methods}{

\tabular{lll}{
\strong{Method} \tab \strong{Name} \tab \strong{Link} \cr
\code{survival(x1, log = FALSE)} \tab Survival function \tab \code{\link{survival}} \cr
\code{hazard(x1, log = FALSE)} \tab Hazard function \tab \code{\link{hazard}} \cr
\code{cumHazard(x1, log = FALSE)} \tab Cumulative hazard function \tab \code{\link{cumHazard}} \cr
\code{cdfAntiDeriv(lower = self$inf(), upper = self$sup())} \tab Anti-derivative of cdf \tab \code{\link{cdfAntiDeriv}} \cr
\code{survivalAntiDeriv(lower = self$inf(), upper = self$sup())} \tab Anti-derivative of survival function \tab \code{\link{survivalAntiDeriv}} \cr
\code{cdfPNorm(p = 2, lower = self$inf(), upper = self$sup())} \tab P-norm of cdf \tab \code{\link{cdfPNorm}} \cr
\code{pdfPNorm(p = 2, lower = self$inf(), upper = self$sup())} \tab P-norm of pdf \tab \code{\link{pdfPNorm}} \cr
\code{survivalPNorm(p = 2, lower = self$inf(), upper = self$sup())} \tab P-norm of survival function \tab \code{\link{survivalPNorm}} \cr
}
}

\examples{
x = Exponential$new()
decorate(x, ExoticStatistics)
x$survival(1)

x = Exponential$new(decorators = ExoticStatistics)
x$survival(4)
}
\seealso{
\code{\link{DistributionDecorator}} and \code{\link{CoreStatistics}}
}
