% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Exponential.R
\name{Exponential}
\alias{Exponential}
\title{Exponential Distribution}
\description{
Mathematical and statistical functions for the exponential distribution parameterised
with rate or scale.
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{rate} \tab numeric \tab arrival rate. \cr
\code{scale} \tab numeric \tab scale parameter. \cr
\code{decorators} \tab Decorator \tab decorators to add functionality. See details. \cr
\code{...} \tab ANY \tab additional arguments for Distribution constructor. See details. \cr
}
}

\section{Constructor Details}{
 The exponential distribution can either be parameterised with a rate or
scale parameter. If neither are provided then rate parameterisation is used with rate = 1. If both are
provided then rate parameterisation is used with given rate. Scale is defined by
\deqn{scale = 1/rate}
The CoreStatistics and ExoticStatistics decorators can be added to the distribution for further
numeric functionality, but these are approximate calculations only. Additional arguments can be passed
to the Distribution constructor, including R62S3 to determine if S3 methods should be added for
the exponential distribution.
}

\section{Public Variables}{

 \tabular{lr}{
  \strong{Method} \tab \strong{Return} \cr
  \code{name} \tab "Exponential" \cr
  \code{short_name} \tab "Exp" \cr
  \code{traits} \tab List of exponential distribution traits. \cr
  \code{properties} \tab List of exponential distribution properties. \cr
  }
}

\section{Public Methods}{

 \tabular{lrr}{
  \strong{Method} \tab \strong{Return Type} \tab \strong{Details} \cr
  \code{pdf(x1, log = FALSE)} \tab character \tab Evaluates density at x1. \cr
  \code{cdf(x1, lower.tail = TRUE, log.p = FALSE)} \tab numeric \tab Evaluates distribution function at x1. \cr
  \code{quantile(p, lower.tail = TRUE, log.p = FALSE)} \tab numeric \tab Evalutes inverse distribution at p.  \cr
  \code{rand(n)} \tab numeric \tab Randomly generates n samples from the distribution.  \cr
  \code{expectation()} \tab numeric \tab Expectation.  \cr
  \code{var()} \tab numeric \tab Variance.  \cr
  \code{skewness()} \tab numeric \tab Skewness. \cr
  \code{kurtosis(excess = TRUE)} \tab numeric \tab Kurtosis. Kurtosis - 3 if excess = TRUE. \cr
  \code{entropy(base = 2)} \tab numeric \tab Entropy. Shannon if base = 2. \cr
  \code{mode()} \tab numeric \tab Mode. \cr
  \code{mgf(t)} \tab numeric \tab Evaluates moment generating function at t. \cr
  \code{cf(t)} \tab numeric \tab Evaluates characteristic function at t. \cr
  \code{survival(x1, log.p = FALSE)} \tab numeric \tab Evaluates survival function at x1. \cr
  \code{hazard(x1)} \tab numeric \tab Evaluates hazard function at x1. \cr
  \code{cumHazard(x1)} \tab numeric \tab Evaluates cumulative hazard function at x1. \cr
  }
}

\section{Public Methods Details}{

If \code{log.p} is TRUE then the natural logarithm of probabilities is returned. If \code{lower.tail}
is TRUE then distribution functions are evaluated at the lower tail of the distribution, otherwise
the upper tail (1 - p).
}

\seealso{
See \code{\link{Distribution}} for inherited methods and variables. See \code{\link{DistributionDecorator}}
for Decorator details as well as \code{\link{CoreStatistics}} and \code{\link{ExoticStatistics}}.
}
