% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_FunctionImputation.R
\name{FunctionImputation}
\alias{FunctionImputation}
\title{Numeric Pdf/Cdf/Quantile/Rand Functions}
\description{
Numeric statistical functions including density/mass function, distribution function,
inverse distribution and simulation.
}
\details{
Decorator objects add functionality to the given Distribution object
 by copying methods in the decorator environment to the chosen Distribution environment. Use the
 \code{\link{decorate}} function to decorate a Distribution. The functions are imputed based on
 which are already available in the distribution. For example if the pdf is known then this is used
 to generate the cdf, quantile and rand. But if the cdf and pdf are both known, then cdf is used to
 generate the quantile.

 All methods in this decorator use numerical approximations and therefore better results may be available
 from analytic computations.
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{dist} \tab distribution \tab Distribution to decorate. \cr
\code{R62S3} \tab logical \tab If TRUE (default), S3 methods are added for decorators in construction.
}
}

\section{Public Methods}{

\tabular{lll}{
\strong{Method} \tab \strong{Input -> Output} \tab \strong{Details} \cr
\code{pdf(x1)} \tab numeric x numeric -> numeric \tab Numeric pdf evaluated at x1. \cr
\code{cdf(x1)} \tab numeric x numeric -> numeric \tab Numeric cdf evaluated at x1. \cr
\code{quantile(p)} \tab numeric x numeric -> numeric \tab Numeric inverse cdf evaluated at x1. \cr
\code{rand(n)} \tab integer x numeric -> numeric \tab n simulations from distribution. \cr
}
}

\examples{
x = Distribution$new("Test", pdf = function(x) 1/(4-1), support = Interval$new(1,4),
type = Reals$new())
decorate(x, FunctionImputation)
plot(x$pdf(0:5))
plot(x$cdf(0:5))

x = Distribution$new("Test", pdf = function(x) 1/(4-1), decorators = ExoticStatistics)
x$cdf(1)
}
\seealso{
\code{\link{DistributionDecorator}}
}
