% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_helpers.R
\name{generalPNorm}
\alias{generalPNorm}
\title{Generalised P-Norm}
\usage{
generalPNorm(fun, p, lower, upper)
}
\arguments{
\item{fun}{function to calculate the p-norm of.}

\item{p}{the pth norm to calculate}

\item{lower}{lower bound for the integral}

\item{upper}{upper bound for the integral}
}
\description{
Calculate the p-norm of any function between given limits. Given by,
\deqn{(\int_S |f|^p d\mu)^1/p}
}
\examples{
generalPNorm(Exponential$new()$pdf,2,0,10)

}
