% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{genExp}
\alias{genExp}
\alias{genExp.Distribution}
\title{Generalised Expectation of a Distribution}
\usage{
genExp(trafo = NULL)
}
\arguments{
\item{trafo}{transformation for expectation calculation, see details.}
}
\description{
A generalised expectation function for distributions, for arithmetic mean and more complex
numeric calculations.
}
\details{
The expectation of a probability distribution can be numerically calculated in a variety
of different ways, some more efficient than others depending on what is available, this function first
checks which analytic methods are present before selecting a numeric strategy.

If trafo = NULL, then the arithmetic mean is calculated, i.e. the approximation to \eqn{E[X]}. Any
transformation must be given as a function, for example \code{trafo = function(x) x^2}
(which is the second moment).

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.
}
