% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDistribution_Geometric.R
\name{Geometric}
\alias{Geometric}
\title{Geometric Distribution}
\description{
Mathematical and statistical functions for the Geometric distribution parameterised
with prob or \eqn{qprob = 1 - prob}. The Geometric distribution for
the number of X trials before one success is defined by
the pmf,
\deqn{f(x) = (1 - p)^{k-1}p}
where x is the number of trials needed to get one success and \eqn{0 < p < 1} is the probability
of success.

Alternatively the Geometric distribution can also be defined by the pmf,
\deqn{f(x) = (1 - p)^{k}p}
where x is the number of failures before the first success and \eqn{0 < p \le 1} is the probability
of success.
}
\details{
The Geometric Distribution is parameterised by default with probability of success = 0.5
and such that \eqn{x} is the number of failures before the first success, as this is the implementation
in R stats. Once constructed, the type of Geometric distribution cannot be changed or updated as parameter,
only the probability of success/failure.
}
\section{Constructor}{
 Geometric$new(prob = 0.5, qprob = NULL, trials = FALSE, decorators = NULL, verbose = FALSE)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{prob} \tab numeric \tab probability of success. \cr
\code{qprob} \tab numeric \tab probability of failure. \cr
\code{trials} \tab logical \tab number of trials or failures, see details. \cr
\code{decorators} \tab Decorator \tab decorators to add functionality. \cr
\code{verbose} \tab logical \tab if TRUE parameterisation messages produced.
}
}

\section{Constructor Details}{
 The Geometric distribution is parameterised with probability of success,
prob = 0.5 by default. The logical parameter \code{trials} determines which Geometric distribution
is constructed and cannot be changed after construction. If \code{trials} is TRUE then
the Geometric distribution that calculates the probability of first success is after \eqn{x} trials.
Otherwise the Geometric distribution calculates probability of \eqn{y} failures before the first success.
Mathematically these are related by \eqn{Y = X - 1}.
}

\section{Public Variables}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Return} \cr
  \code{name} \tab Name of distribution. \cr
  \code{short_name} \tab Id of distribution. \cr
  \code{description} \tab Brief description of distribution. \cr
  \code{traits} \tab List: type, valueSupport, variateForm. \cr
  \code{package} \tab The package p/d/q/r are implemented in.
 }
}

\section{Public Methods}{

 \tabular{ll}{
  \strong{Accessor Methods} \tab \strong{Link} \cr
  \code{decorators()} \tab \code{\link{decorators}} \cr
  \code{valueSupport()} \tab \code{\link{valueSupport}} \cr
  \code{variateForm()} \tab \code{\link{variateForm}} \cr
  \code{type()} \tab \code{\link{type}} \cr
  \code{properties()} \tab \code{\link{properties}} \cr
  \code{support()} \tab \code{\link{support}} \cr
  \code{distrDomain()} \tab \code{\link{distrDomain}} \cr
  \code{symmetry()} \tab \code{\link{symmetry}} \cr
  \code{sup()}  \tab \code{\link{sup}} \cr
  \code{inf()} \tab \code{\link{inf}} \cr
  \code{dmax()}  \tab \code{\link{dmax}} \cr
  \code{dmin()} \tab \code{\link{dmin}} \cr
  \code{skewnessType()} \tab \code{\link{skewnessType}} \cr
  \code{kurtosisType()} \tab \code{\link{kurtosisType}} \cr

  \tab \cr \tab \cr \tab \cr

  \strong{Statistical Methods} \tab \strong{Link} \cr
  \code{pdf(x1, ..., log = FALSE, simplify = TRUE)} \tab \code{\link{pdf}} \cr
  \code{cdf(x1, ..., lower.tail = TRUE, log.p = FALSE, simplify = TRUE)} \tab \code{\link{cdf}}\cr
  \code{quantile(p, ..., lower.tail = TRUE, log.p = FALSE, simplify = TRUE)} \tab \code{\link{quantile.Distribution}} \cr
  \code{rand(n, simplify = TRUE)} \tab \code{\link{rand}} \cr
  \code{mean()} \tab \code{\link{mean.Distribution}} \cr
  \code{var()} \tab \code{\link{var}} \cr
  \code{prec()} \tab \code{\link{prec}} \cr
  \code{cor()} \tab \code{\link{cor}} \cr
  \code{skewness()} \tab \code{\link{skewness}} \cr
  \code{kurtosis(excess = TRUE)} \tab \code{\link{kurtosis}} \cr
  \code{entropy(base = 2)} \tab \code{\link{entropy}} \cr
  \code{mgf(t)} \tab \code{\link{mgf}} \cr
  \code{cf(t)} \tab \code{\link{cf}} \cr
  \code{pgf(z)} \tab \code{\link{pgf}} \cr
  \code{sd()} \tab \code{\link{sd}} \cr
  \code{median()} \tab \code{\link{median.Distribution}} \cr
  \code{iqr()} \tab \code{\link{iqr}} \cr

  \tab \cr \tab \cr \tab \cr

  \strong{Parameter Methods} \tab \strong{Link} \cr
  \code{parameters(id)} \tab \code{\link{parameters}} \cr
  \code{getParameterValue(id, error = "warn")}  \tab \code{\link{getParameterValue}} \cr
  \code{setParameterValue(lst, error = "warn")} \tab \code{\link{setParameterValue}} \cr

  \tab \cr \tab \cr \tab \cr

  \strong{Validation Methods} \tab \strong{Link} \cr
  \code{liesInSupport(x, all = TRUE, bound = FALSE)} \tab \code{\link{liesInSupport}} \cr
  \code{liesInType(x, all = TRUE)} \tab \code{\link{liesInType}} \cr
  \code{liesInDistrDomain(x, all = TRUE)} \tab \code{\link{liesInDistrDomain}} \cr

  \tab \cr \tab \cr \tab \cr

  \strong{Representation Methods} \tab \strong{Link} \cr
  \code{strprint()} \tab \code{\link{strprint}} \cr
  \code{print()} \tab \code{\link[base]{print}} \cr
  \code{summary(full = T)} \tab \code{\link{summary.Distribution}} \cr
  \code{plot()} \tab Coming Soon. \cr
  \code{qqplot()} \tab Coming Soon. \cr
  }
}

\examples{
# Different parameterisations
Geometric$new(prob = 0.2)
Geometric$new(qprob = 0.7)

# Different forms of the distribution
Geometric$new(trials = TRUE) # Number of trials before first success
Geometric$new(trials = FALSE) # Number of failures before first success

# Use description to see which form is used
Geometric$new(trials = TRUE)$description
Geometric$new(trials = FALSE)$description

x <- Geometric$new() # Default is prob = 0.5 and number of failures before first success

# Update parameters
x$setParameterValue(list(qprob = 0.2))  # When any parameter is updated, all others are too!
x$parameters()

# p/d/q/r
x$pdf(5)
x$cdf(5)
x$quantile(0.42)
x$rand(4)

# Statistics
x$mean()
x$var()

summary(x)

}
