% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{getParameterValue}
\alias{getParameterValue}
\title{Parameter Value Accessor}
\usage{
getParameterValue(object, id, error = "warn")
}
\arguments{
\item{object}{Distribution or ParameterSet.}

\item{id}{character, id of the parameter to return.}

\item{error}{character, value to pass to \code{stopwarn}.}
}
\value{
The current value of a given parameter as a numeric.
}
\description{
Returns the value of the given parameter.
}
