% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_HuberizedDistribution.R
\name{HuberizedDistribution}
\alias{HuberizedDistribution}
\title{Distribution Huberization Wrapper}
\description{
A wrapper for huberizing any probability distribution at given limits.
}
\details{
Huberizes a distribution at lower and upper limits, using the formula
\tabular{lll}{
 \tab \eqn{F(x)} \tab if \eqn{x \le lower} \cr
 f_H(x) = \tab \eqn{f(x)} \tab if \eqn{lower < x < upper} \cr
 \tab \eqn{1 - F(x)} \tab if \eqn{x \ge upper} \cr
}
where f_H is the pdf of the truncated distribution H = Huberize(X, lower, upper) and f_X/F_X is the
pdf/cdf of the original distribution.

If lower or upper are NULL they are taken to be \code{self$inf()} and \code{self$sup()} respectively.
}
\section{Constructor}{
 HuberizedDistribution$new(distribution, lower = NULL, upper = NULL)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{distribution} \tab distribution \tab Distribution to huberize. \cr
\code{lower} \tab numeric \tab Lower limit for huberization. \cr
\code{upper} \tab numeric \tab Upper limit for huberization.
}
}

\section{Public Methods}{

See \code{\link{Distribution}} and \code{\link{DistributionWrapper}}.
}

\examples{
hubBin <- HuberizedDistribution$new(Binomial$new(prob = 0.5, size = 10), lower = 2, upper = 4)
hubBin$getParameterValue("Binom_prob")
hubBin$getLowerLimit()
hubBin$pdf(2)
}
\seealso{
\code{\link{listWrappers}}.
}
