% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{inf}
\alias{inf}
\title{Infimum Accessor}
\usage{
inf(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
Infimum as a numeric.
}
\description{
Returns the distribution infimum as the infimum of the support.
}
\section{R6 Usage}{
 $inf
}

