% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\name{Integers}
\alias{Integers}
\title{Set of Integers}
\description{
The mathematical set of integers.
}
\details{
The set of Integers is defined as the set of numbers that can be written without a fractional
component, i.e.
\deqn{Integers = \{...,-3, -2, -1, 0, 1, 2, 3,...\}}{Integers = {...,-3, -2, -1, 0, 1, 2, 3,...}}
}
\section{Constructor}{
 Integers$new(dim = 1,...)
}

\section{Constructor Arguments}{

 \tabular{lll}{
   \strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
   \code{dim} \tab numeric \tab Dimension of the set. \cr
   \code{...} \tab ANY \tab Additional arguments.
}
}

\section{Constructor Details}{
 Generally the \code{...} argument should be ignored, its primary use-case
is for the child-classes.
}

\examples{
Integers$new()
Integers$new(dim = 2)

}
\seealso{
\code{\link{listSpecialSets}}
}
