% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_Interval.R
\name{Interval}
\alias{Interval}
\alias{numeric}
\title{Symbolic Interval Object}
\arguments{
\item{lower}{lower limit of the interval.}

\item{upper}{upper limit of the interval.}

\item{type}{endpoint type of the interval.}
}
\value{
\code{Interval$new} constructs an R6 object of class Interval

\code{numeric} gets numeric vector of the interval range.
}
\description{
An R6 set object for symbolic representation of mathematical intervals.
}
\section{Usage}{
 Interval$new(lower = -Inf, upper = Inf, type = "[]")

 $numeric()
}

\seealso{
The parent class \code{\link{SetInterval}} for a full list of inherited methods and variables.
}
