% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{iqr}
\alias{iqr}
\alias{iqr.Distribution}
\title{Interquartile Range}
\usage{
iqr()
}
\description{
Interquartile range of a distribution
}
\details{
The interquartile range of a distribution is defined by
\deqn{iqr_X = q(0.75) - q(0.25)}
where q is the quantile, or inverse distribution function.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.
}
