% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{iqr}
\alias{iqr}
\title{Distribution Interquartile Range}
\usage{
iqr(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
Interquartile range of distribution as a numeric.
}
\description{
Interquartile range of a distribution
}
