% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{kthmoment}
\alias{kthmoment}
\title{Kth Moment}
\arguments{
\item{object}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{k}{the kth moment to calculate}

\item{type}{one of 'central', 'standard' or 'zero', abbreviations allowed}
}
\description{
Kth standardised or central moment of a distribution
}
\details{
The kth central moment of a distribution is defined by
\deqn{CM(k)_X = E_X[(x - \mu)^k]}
the kth standardised moment of a distribution is defined by
\deqn{SM(k)_X = CM(k)/\sigma^k}
#' the kth zeroth moment of a distribution is defined by
\deqn{ZM(k)_X = E_X[(x)^k]}
where E_X is the expectation of distribution X, \eqn{\mu} is the mean of the distribution and \eqn{\sigma} is the
standard deviation of the distribution.

Abbreviations for the type are allowed but if an unfamiliar input is given then the central moment
is computed.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method.
}
