% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{kthmoment}
\alias{kthmoment}
\title{Kth Moment}
\usage{
kthmoment(object, k, type = c("central", "standard", "raw"), ...)
}
\arguments{
\item{object}{Distribution.}

\item{k}{the kth moment to calculate}

\item{type}{one of 'central', 'standard' or 'raw', abbreviations allowed}

\item{...}{Passed to \verb{$genExp}.}
}
\value{
If univariate, the given k-moment as a numeric, otherwise NULL.
}
\description{
Kth standardised or central moment of a distribution
}
