% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{kurtosis}
\alias{kurtosis}
\alias{kurtosis.Distribution}
\title{Kurtosis}
\usage{
kurtosis()
}
\description{
Kurtosis of a distribution
}
\details{
The kurtosis of a distribution is defined by the fourth standardised moment of the
distribution,
\deqn{k_X = E_X[(x - \mu)^4]/\sigma^4}
where E_X is the expectation of distribution X, \eqn{\mu} is the mean of the distribution and \eqn{\sigma} is the
standard deviation of the distribution.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.
}
