% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{kurtosis}
\alias{kurtosis}
\title{Distribution Kurtosis}
\usage{
kurtosis(object, excess = TRUE, ...)
}
\arguments{
\item{object}{Distribution.}

\item{excess}{logical, if TRUE (default) excess Kurtosis returned.}

\item{...}{Passed to \verb{$genExp}.}
}
\value{
Kurtosis as a numeric.
}
\description{
Kurtosis of a distribution
}
