% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{kurtosisType}
\alias{kurtosisType}
\title{Type of Kurtosis Accessor}
\usage{
kurtosisType(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
If the distribution kurtosis is present in properties, returns one of "platykurtic"/"mesokurtic"/"leptokurtic",
otherwise returns NULL.
}
\description{
Returns the type of kurtosis (in relation to Normal distribution)
}
\section{R6 Usage}{
 $kurtosisType()
}

\seealso{
\code{\link{kurtosis}}, \code{\link{properties}} and \code{\link{skewnessType}}
}
