% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{kurtosisType}
\alias{kurtosisType}
\title{Type of Kurtosis Accessor - Deprecated}
\usage{
kurtosisType(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
If the distribution kurtosis is present in properties, returns one of
"platykurtic"/"mesokurtic"/"leptokurtic", otherwise returns NULL.
}
\description{
Deprecated. Use $properties$kurtosis.
}
