% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDecorators.R
\name{listDecorators}
\alias{listDecorators}
\title{Lists Implemented Distribution Decorators}
\usage{
listDecorators(simplify = TRUE)
}
\arguments{
\item{simplify}{logical. If \code{TRUE} (default) returns results as characters,
otherwise as R6 classes.}
}
\value{
Either a list of characters (if \code{simplify} is \code{TRUE}) or a list of
\link{DistributionDecorator} classes.
}
\description{
Lists decorators that can decorate an R6 Distribution.
}
\examples{
listDecorators()
listDecorators(FALSE)
}
\seealso{
\code{\link{DistributionDecorator}}
}
