% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDistributions.R
\name{listDistributions}
\alias{listDistributions}
\title{Lists Implemented R6 Distributions}
\usage{
listDistributions(simplify = FALSE, filter = NULL)
}
\arguments{
\item{simplify}{logical.}

\item{filter}{list to filter distributions by.}
}
\description{
Lists R6 distributions in a data.table or a character vector, can be filtered by
traits and implemented package.
}
\examples{
listDistributions()
listDistributions(filter = list(VariateForm = "univariate"))
listDistributions(filter = list(ValueSupport = "discrete"))
}
