% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_helpers.R
\name{listDistributions}
\alias{listDistributions}
\title{Lists Implemented R6 Distributions}
\usage{
listDistributions(simplify = FALSE, traits = NULL, properties = NULL)
}
\arguments{
\item{simplify}{logical.}

\item{traits}{list of traits to filter distributions by.}

\item{properties}{list of traits to filter distributions by.}
}
\description{
Lists R6 distributions, either all in a data.frame or filtered by chosen
traits and/or properties.
}
\examples{
listDistributions()
listDistributions(traits = list(variateForm = "univariate"))
listDistributions(properties = list(symmetry = FALSE))
}
