% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listKernels.R
\name{listKernels}
\alias{listKernels}
\title{Lists Implemented R6 Distributions}
\usage{
listKernels(simplify = FALSE)
}
\arguments{
\item{simplify}{logical.}

\item{traits}{list of traits to filter distributions by.}

\item{view}{logical, if TRUE displays Distributions in Viewer. Ignored if \code{simplify} is FALSE.}
}
\description{
Lists R6 distributions, either all in a data.frame or filtered by chosen
traits and/or properties.
}
\examples{
listKernels()
listKernels(traits = list(VariateForm = "univariate"))
listKernels(traits = list(ValueSupport = "discrete"))
}
