% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listKernels.R
\name{listKernels}
\alias{listKernels}
\title{Lists Implemented R6 Kernels}
\usage{
listKernels(simplify = FALSE)
}
\arguments{
\item{simplify}{logical.}
}
\description{
Lists all implemented kernels in disrt6.
}
\examples{
listKernels()

}
