% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_helpers.R
\name{listSpecialSets}
\alias{listSpecialSets}
\title{Lists Implemented R6 Special Sets}
\usage{
listSpecialSets(simplify = FALSE, view = FALSE)
}
\arguments{
\item{simplify}{logical. If FALSE (default) returns data.frame of set name and symbol, otherwise character.}

\item{view}{logical, if TRUE displays Distributions in Viewer. Ignored if \code{simplify} is FALSE.}
}
\description{
Lists special sets that can be used in SetInterval.
}
\examples{
listSpecialSets()
listSpecialSets(TRUE)
}
