% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSpecialSets.R
\name{listSpecialSets}
\alias{listSpecialSets}
\title{Lists Implemented R6 Special Sets}
\usage{
listSpecialSets(simplify = FALSE)
}
\arguments{
\item{simplify}{logical. If FALSE (default) returns data.table of set name and symbol, otherwise set names as characters.}
}
\value{
Either a list of characters (if \code{simplify} is TRUE) or a data.table of \code{SpecialSet}s and their traits.
}
\description{
Lists special sets that can be used in SetInterval.
}
\examples{
listSpecialSets()
listSpecialSets(TRUE)
}
\seealso{
\code{\link{SpecialSet}}
}
