% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listWrappers.R
\name{listWrappers}
\alias{listWrappers}
\title{Lists Implemented Distribution Wrappers}
\usage{
listWrappers(simplify = TRUE)
}
\arguments{
\item{simplify}{logical. If TRUE (default) returns results as characters, otherwise as R6
classes.}
}
\value{
Either a list of characters (if \code{simplify} is \code{TRUE}) or a list of \code{Wrapper} classes.
}
\description{
Lists wrappers that can wrap an R6 Distribution.
}
\examples{
listWrappers()
listWrappers(TRUE)
}
\seealso{
\code{\link{DistributionWrapper}}
}
