% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDistribution_Lognormal.R
\name{Lognormal}
\alias{Lognormal}
\title{Lognormal Distribution}
\description{
Mathematical and statistical functions for the Lognormal distribution parameterised
with mean and standard deviation of the variable's natural logarithm.
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{meanlog} \tab numeric \tab mean of the distribution on the log scale. \cr
\code{varlog} \tab numeric \tab variance of the distribution on the log scale. \cr
\code{sdlog} \tab numeric \tab standard deviation of the distribution on the log scale. \cr
\code{precisionlog} \tab numeric \tab parameter. \cr
\code{decorators} \tab Decorator \tab decorators to add functionality. See details. \cr
\code{...} \tab ANY \tab additional arguments for Distribution constructor. See details. \cr
}
}

\section{Constructor Details}{
 The Normal distribution can either be parameterised with variance,
standard deviation or precision. If none are provided then var parameterisation is used with var = 1.
If multiple are provided then parameterisatin takes the hierarchy: var, sd, prec.
sd is defined by
\deqn{sd = var^2}
prec is defined by
\deqn{prec = var^-1}

The CoreStatistics and ExoticStatistics decorators can be added to the distribution for further
numeric functionality, but these are approximate calculations only. Additional arguments can be passed
to the Distribution constructor, including R62S3 to determine if S3 methods should be added for
the Lognormal distribution.
}

\section{Public Variables}{

 \tabular{lr}{
  \strong{Method} \tab \strong{Return} \cr
  \code{name} \tab "Lognormal" \cr
  \code{short_name} \tab "lnorm" \cr
  \code{traits} \tab List of Lognormal distribution traits. \cr
  \code{properties} \tab List of Lognormal distribution properties. \cr
  }
}

\section{Public Methods}{

 \tabular{lrr}{
  \strong{Method} \tab \strong{Return Type} \tab \strong{Details} \cr
  \code{pdf(x1, log = FALSE)} \tab character \tab Evaluates density at x1. \cr
  \code{cdf(x1, lower.tail = TRUE, log.p = FALSE)} \tab numeric \tab Evaluates distribution function at x1. \cr
  \code{quantile(p, lower.tail = TRUE, log.p = FALSE)} \tab numeric \tab Evalutes inverse distribution at p.  \cr
  \code{rand(n)} \tab numeric \tab Randomly generates n samples from the distribution.  \cr
  \code{normectation()} \tab numeric \tab normectation.  \cr
  \code{var()} \tab numeric \tab Variance.  \cr
  \code{skewness()} \tab numeric \tab Skewness. \cr
  \code{kurtosis(excess = TRUE)} \tab numeric \tab Kurtosis. Kurtosis - 3 if excess = TRUE. \cr
  \code{entropy(base = 2)} \tab numeric \tab Entropy. Shannon if base = 2. \cr
  \code{mode()} \tab numeric \tab Mode. \cr
  \code{mgf(t)} \tab numeric \tab Evaluates moment generating function at t. \cr
  \code{cf(t)} \tab numeric \tab Evaluates characteristic function at t. \cr
  \code{survival(x1, log.p = FALSE)} \tab numeric \tab Evaluates survival function at x1. \cr
  \code{hazard(x1)} \tab numeric \tab Evaluates hazard function at x1. \cr
  \code{cumHazard(x1)} \tab numeric \tab Evaluates cumulative hazard function at x1. \cr
  }
}

\section{Public Methods Details}{

If \code{log.p} is TRUE then the natural logarithm of probabilities is returned. If \code{lower.tail}
is TRUE then distribution functions are evaluated at the lower tail of the distribution, otherwise
the upper tail (1 - p).
}

\seealso{
See \code{\link{Distribution}} for inherited methods and variables. See \code{\link{DistributionDecorator}}
for Decorator details as well as \code{\link{CoreStatistics}} and \code{\link{ExoticStatistics}}.
}
