% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUniqueDistributions.R
\name{makeUniqueDistributions}
\alias{makeUniqueDistributions}
\title{De-Duplicate Distribution Names}
\usage{
makeUniqueDistributions(distlist)
}
\arguments{
\item{distlist}{list of Distributions.}
}
\value{
The list of inputted distributions except with the \code{short_name}s manipulated as
necessary to make them unique.
}
\description{
Helper function to lapply over the given distribution list, and make the
\code{short_name}s unique.
}
\details{
The \code{short_name}s are made unique by suffixing each with a consecutive
number so that the names are no longer duplicated.
}
\examples{
makeUniqueDistributions(list(Binomial$new(), Binomial$new()))
}
