% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{mean.Distribution}
\alias{mean.Distribution}
\title{Distribution Mean}
\usage{
\method{mean}{Distribution}(x, ...)
}
\arguments{
\item{x}{Distribution.}

\item{...}{Passed to \verb{$genExp}.}
}
\value{
Mean as a numeric.
}
\description{
Arithmetic mean for the probability distribution.
}
