% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{mean.Distribution}
\alias{mean.Distribution}
\title{Distribution Mean}
\usage{
\method{mean}{Distribution}(x, ...)
}
\arguments{
\item{x}{Distribution.}

\item{...}{Additional arguments.}
}
\description{
Arithmetic mean for the probability distribution.
}
\details{
The arithmetic mean of a (discrete) probability distribution X is the expectation
\deqn{E_X(X) = \sum p_X(x)*x}
with an integration analogue for continuous distributions.

If an analytic expression isn't available, returns error. To impute a numerical expression, use the
\code{\link{CoreStatistics}} decorator.
}
\section{R6 Usage}{
 $mean()
}

\seealso{
\code{\link{CoreStatistics}}, \code{\link{decorate}} and \code{\link{genExp}}.
}
