% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{mgf}
\alias{mgf}
\title{Moment Generating Function}
\arguments{
\item{object}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{t}{integer to evaluate moment generating function at.}
}
\description{
Moment generating function of a distribution
}
\details{
The moment generating function is defined by
\deqn{mgf_X(t) = E_X[exp(xt)]}
where X is the distribution and E_X is the expectation of the distribution X.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.

export
}
