% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{mgf}
\alias{mgf}
\title{Moment Generating Function}
\usage{
mgf(object, t)
}
\arguments{
\item{object}{Distribution.}

\item{t}{integer to evaluate moment generating function at.}
}
\description{
Moment generating function of a distribution
}
\details{
The moment generating function is defined by
\deqn{mgf_X(t) = E_X[exp(xt)]}
where X is the distribution and E_X is the expectation of the distribution X.

If an analytic expression isn't available, returns error. To impute a numerical expression, use the
\code{\link{CoreStatistics}} decorator.
}
\section{R6 Usage}{
 $mgf(t)
}

\seealso{
\code{\link{CoreStatistics}} and \code{\link{decorate}}
}
