% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{mgf}
\alias{mgf}
\title{Moment Generating Function}
\usage{
mgf(object, t, ...)
}
\arguments{
\item{object}{Distribution.}

\item{t}{integer to evaluate moment generating function at.}

\item{...}{Passed to \verb{$genExp}.}
}
\value{
Moment generating function evaluated at t as a numeric.
}
\description{
Moment generating function of a distribution
}
