% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_MixtureDistribution.R
\name{MixtureDistribution}
\alias{MixtureDistribution}
\title{Mixture Distribution Wrapper}
\description{
Wrapper used to construct a mixture of two or more distributions.
}
\details{
A mixture distribution is defined by

\deqn{F_P(x) = w_1 F_{X1}(x) * ... * w_n F_{XN}(x)}{F_P(x) = w_1 F_X1(x) * ... * w_n F_XN(x)} #nolint
where \eqn{F_P} is the cdf of the mixture distribution, \eqn{X1,...,XN} are
independent distributions, and \eqn{w1,...,wN} are weights for the mixture.
}
\examples{

## ------------------------------------------------
## Method `MixtureDistribution$new`
## ------------------------------------------------

MixtureDistribution$new(list(Binomial$new(prob = 0.5, size = 10), Binomial$new()),
  weights = c(0.2, 0.8)
)

## ------------------------------------------------
## Method `MixtureDistribution$pdf`
## ------------------------------------------------

m <- MixtureDistribution$new(list(Binomial$new(prob = 0.5, size = 10), Binomial$new()),
  weights = c(0.2, 0.8)
)
m$pdf(1:5)
m$pdf(1)
# also possible but unlikely to be used
m$pdf(1, 2)

## ------------------------------------------------
## Method `MixtureDistribution$rand`
## ------------------------------------------------

m <- MixtureDistribution$new(distribution = "Normal",
params = data.frame(mean = 1:2, sd = 1))
m$rand(5)
}
\seealso{
Other wrappers: 
\code{\link{Convolution}},
\code{\link{DistributionWrapper}},
\code{\link{HuberizedDistribution}},
\code{\link{ProductDistribution}},
\code{\link{TruncatedDistribution}},
\code{\link{VectorDistribution}}
}
\concept{wrappers}
\section{Super classes}{
\code{\link[distr6:Distribution]{distr6::Distribution}} -> \code{\link[distr6:DistributionWrapper]{distr6::DistributionWrapper}} -> \code{\link[distr6:VectorDistribution]{distr6::VectorDistribution}} -> \code{MixtureDistribution}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MixtureDistribution-new}{\code{MixtureDistribution$new()}}
\item \href{#method-MixtureDistribution-strprint}{\code{MixtureDistribution$strprint()}}
\item \href{#method-MixtureDistribution-pdf}{\code{MixtureDistribution$pdf()}}
\item \href{#method-MixtureDistribution-cdf}{\code{MixtureDistribution$cdf()}}
\item \href{#method-MixtureDistribution-quantile}{\code{MixtureDistribution$quantile()}}
\item \href{#method-MixtureDistribution-rand}{\code{MixtureDistribution$rand()}}
\item \href{#method-MixtureDistribution-clone}{\code{MixtureDistribution$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="confidence"><a href='../../distr6/html/Distribution.html#method-Distribution-confidence'><code>distr6::Distribution$confidence()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="correlation"><a href='../../distr6/html/Distribution.html#method-Distribution-correlation'><code>distr6::Distribution$correlation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="iqr"><a href='../../distr6/html/Distribution.html#method-Distribution-iqr'><code>distr6::Distribution$iqr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="liesInSupport"><a href='../../distr6/html/Distribution.html#method-Distribution-liesInSupport'><code>distr6::Distribution$liesInSupport()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="liesInType"><a href='../../distr6/html/Distribution.html#method-Distribution-liesInType'><code>distr6::Distribution$liesInType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="parameters"><a href='../../distr6/html/Distribution.html#method-Distribution-parameters'><code>distr6::Distribution$parameters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="prec"><a href='../../distr6/html/Distribution.html#method-Distribution-prec'><code>distr6::Distribution$prec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="print"><a href='../../distr6/html/Distribution.html#method-Distribution-print'><code>distr6::Distribution$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="setParameterValue"><a href='../../distr6/html/Distribution.html#method-Distribution-setParameterValue'><code>distr6::Distribution$setParameterValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="stdev"><a href='../../distr6/html/Distribution.html#method-Distribution-stdev'><code>distr6::Distribution$stdev()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="summary"><a href='../../distr6/html/Distribution.html#method-Distribution-summary'><code>distr6::Distribution$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="Distribution" data-id="workingSupport"><a href='../../distr6/html/Distribution.html#method-Distribution-workingSupport'><code>distr6::Distribution$workingSupport()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="cf"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-cf'><code>distr6::VectorDistribution$cf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="entropy"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-entropy'><code>distr6::VectorDistribution$entropy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="getParameterValue"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-getParameterValue'><code>distr6::VectorDistribution$getParameterValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="kurtosis"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-kurtosis'><code>distr6::VectorDistribution$kurtosis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="mean"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-mean'><code>distr6::VectorDistribution$mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="median"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-median'><code>distr6::VectorDistribution$median()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="mgf"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-mgf'><code>distr6::VectorDistribution$mgf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="mode"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-mode'><code>distr6::VectorDistribution$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="pgf"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-pgf'><code>distr6::VectorDistribution$pgf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="skewness"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-skewness'><code>distr6::VectorDistribution$skewness()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="variance"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-variance'><code>distr6::VectorDistribution$variance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="distr6" data-topic="VectorDistribution" data-id="wrappedModels"><a href='../../distr6/html/VectorDistribution.html#method-VectorDistribution-wrappedModels'><code>distr6::VectorDistribution$wrappedModels()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixtureDistribution-new"></a>}}
\if{latex}{\out{\hypertarget{method-MixtureDistribution-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixtureDistribution$new(
  distlist = NULL,
  weights = "uniform",
  distribution = NULL,
  params = NULL,
  shared_params = NULL,
  name = NULL,
  short_name = NULL,
  decorators = NULL,
  vecdist = NULL,
  ids = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distlist}}{\code{(list())}\cr
List of \link{Distribution}s.}

\item{\code{weights}}{\code{(character(1)|numeric())}\cr
Weights to use in the resulting mixture. If all distributions are weighted equally then
\code{"uniform"} provides a much faster implementation, otherwise a vector of length equal
to the number of wrapped distributions, this is automatically scaled internally.}

\item{\code{distribution}}{\code{(character(1))} \cr
Should be supplied with \code{params} and optionally \code{shared_params} as an alternative to \code{distlist}.
Much faster implementation when only one class of distribution is being wrapped. \code{distribution}
is the full name of one of the distributions in \code{\link[=listDistributions]{listDistributions()}}, or \code{"Distribution"} if
constructing custom distributions. See examples in \link{VectorDistribution}.}

\item{\code{params}}{\code{(list()|data.frame())} \cr
Parameters in the individual distributions for use with \code{distribution}. Can be supplied as a list,
where each element is the list of parameters to set in the distribution, or as an object
coercable to \code{data.frame}, where each column is a parameter and each row is a distribution.
See examples in \link{VectorDistribution}.}

\item{\code{shared_params}}{\code{(list())}\cr
If any parameters are shared when using the \code{distribution} constructor, this provides a much faster
implementation to list and query them together. See examples in \link{VectorDistribution}.}

\item{\code{name}}{\code{(character(1))}\cr
Optional name of wrapped distribution.}

\item{\code{short_name}}{\code{(character(1))}\cr
Optional short name/ID of wrapped distribution.}

\item{\code{decorators}}{\code{(character())} \cr
Decorators to add to the distribution during construction.}

\item{\code{vecdist}}{\link{VectorDistribution} \cr
Alternative constructor to directly create this object from an object inheriting from
\link{VectorDistribution}.}

\item{\code{ids}}{\code{(character())} \cr
Optional ids for wrapped distributions in vector, should be unique and of same length as
the number of distributions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{MixtureDistribution$new(list(Binomial$new(prob = 0.5, size = 10), Binomial$new()),
  weights = c(0.2, 0.8)
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixtureDistribution-strprint"></a>}}
\if{latex}{\out{\hypertarget{method-MixtureDistribution-strprint}{}}}
\subsection{Method \code{strprint()}}{
Printable string representation of the \code{MixtureDistribution}. Primarily used internally.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixtureDistribution$strprint(n = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{\code{(integer(1))}\cr
Number of distributions to include when printing.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixtureDistribution-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-MixtureDistribution-pdf}{}}}
\subsection{Method \code{pdf()}}{
Probability density function of the mixture distribution. Computed by
\deqn{f_M(x) = \sum_i (f_i)(x)*w_i}
where \eqn{w_i} is the vector of weights and \eqn{f_i} are the pdfs of the wrapped
distributions.

Note that as this class inherits from \link{VectorDistribution}, it is possible to evaluate
the distributions at different points, but that this is not the usual use-case for
mixture distributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixtureDistribution$pdf(..., log = FALSE, simplify = TRUE, data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{\code{(numeric())} \cr
Points to evaluate the function at Arguments do not need
to be named. The length of each argument corresponds to the number of points to evaluate,
the number of arguments corresponds to the number of variables in the distribution.
See examples.}

\item{\code{log}}{\code{(logical(1))} \cr
If \code{TRUE} returns the logarithm of the probabilities. Default is \code{FALSE}.}

\item{\code{simplify}}{\code{logical(1)} \cr
If \code{TRUE} (default) simplifies the return if possible to a \code{numeric}, otherwise returns a
\link[=data.table]{data.table::data.table}.}

\item{\code{data}}{\link{array} \cr
Alternative method to specify points to evaluate. If univariate then rows correspond with number
of points to evaluate and columns correspond with number of variables to evaluate. In the special
case of \link{VectorDistribution}s of multivariate distributions, then the third dimension corresponds
to the distribution in the vector to evaluate.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{m <- MixtureDistribution$new(list(Binomial$new(prob = 0.5, size = 10), Binomial$new()),
  weights = c(0.2, 0.8)
)
m$pdf(1:5)
m$pdf(1)
# also possible but unlikely to be used
m$pdf(1, 2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixtureDistribution-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-MixtureDistribution-cdf}{}}}
\subsection{Method \code{cdf()}}{
Cumulative distribution function of the mixture distribution. Computed by
\deqn{F_M(x) = \sum_i (F_i)(x)*w_i}
where \eqn{w_i} is the vector of weights and \eqn{F_i} are the cdfs of the wrapped
distributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixtureDistribution$cdf(
  ...,
  lower.tail = TRUE,
  log.p = FALSE,
  simplify = TRUE,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{\code{(numeric())} \cr
Points to evaluate the function at Arguments do not need
to be named. The length of each argument corresponds to the number of points to evaluate,
the number of arguments corresponds to the number of variables in the distribution.
See examples.
@examples
m <- MixtureDistribution$new(list(Binomial$new(prob = 0.5, size = 10), Binomial$new()),
weights = c(0.2, 0.8)
)
m$cdf(1:5)}

\item{\code{lower.tail}}{\code{(logical(1))} \cr
If \code{TRUE} (default), probabilities are \code{X <= x}, otherwise, \code{P(X > x)}.}

\item{\code{log.p}}{\code{(logical(1))} \cr
If \code{TRUE} returns the logarithm of the probabilities. Default is \code{FALSE}.}

\item{\code{simplify}}{\code{logical(1)} \cr
If \code{TRUE} (default) simplifies the return if possible to a \code{numeric}, otherwise returns a
\link[=data.table]{data.table::data.table}.}

\item{\code{data}}{\link{array} \cr
Alternative method to specify points to evaluate. If univariate then rows correspond with number
of points to evaluate and columns correspond with number of variables to evaluate. In the special
case of \link{VectorDistribution}s of multivariate distributions, then the third dimension corresponds
to the distribution in the vector to evaluate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixtureDistribution-quantile"></a>}}
\if{latex}{\out{\hypertarget{method-MixtureDistribution-quantile}{}}}
\subsection{Method \code{quantile()}}{
The quantile function is not implemented for mixture distributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixtureDistribution$quantile(
  ...,
  lower.tail = TRUE,
  log.p = FALSE,
  simplify = TRUE,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{\code{(numeric())} \cr
Points to evaluate the function at Arguments do not need
to be named. The length of each argument corresponds to the number of points to evaluate,
the number of arguments corresponds to the number of variables in the distribution.
See examples.}

\item{\code{lower.tail}}{\code{(logical(1))} \cr
If \code{TRUE} (default), probabilities are \code{X <= x}, otherwise, \code{P(X > x)}.}

\item{\code{log.p}}{\code{(logical(1))} \cr
If \code{TRUE} returns the logarithm of the probabilities. Default is \code{FALSE}.}

\item{\code{simplify}}{\code{logical(1)} \cr
If \code{TRUE} (default) simplifies the return if possible to a \code{numeric}, otherwise returns a
\link[=data.table]{data.table::data.table}.}

\item{\code{data}}{\link{array} \cr
Alternative method to specify points to evaluate. If univariate then rows correspond with number
of points to evaluate and columns correspond with number of variables to evaluate. In the special
case of \link{VectorDistribution}s of multivariate distributions, then the third dimension corresponds
to the distribution in the vector to evaluate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixtureDistribution-rand"></a>}}
\if{latex}{\out{\hypertarget{method-MixtureDistribution-rand}{}}}
\subsection{Method \code{rand()}}{
Simulation function for mixture distributions. Samples are drawn from a mixture by first
sampling Multinomial(probs = weights, size = n), then sampling each distribution according
to the samples from the Multinomial, and finally randomly permuting these draws.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixtureDistribution$rand(n, simplify = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{\code{(numeric(1))} \cr
Number of points to simulate from the distribution. If length greater than \eqn{1}, then
\code{n <- length(n)},}

\item{\code{simplify}}{\code{logical(1)} \cr
If \code{TRUE} (default) simplifies the return if possible to a \code{numeric}, otherwise returns a
\link[=data.table]{data.table::data.table}.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{m <- MixtureDistribution$new(distribution = "Normal",
params = data.frame(mean = 1:2, sd = 1))
m$rand(5)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixtureDistribution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MixtureDistribution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixtureDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
