% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{mode}
\alias{mode}
\title{Mode of a Distribution}
\usage{
mode(object, which = "all")
}
\arguments{
\item{object}{Distribution.}

\item{which}{which mode of the distribution should be returned, default is all.}
}
\description{
A numeric search for the mode(s) of a distribution.
}
\details{
If the distribution has multiple modes, all are returned by default. Otherwise the index
of the mode to return can be given or "all" if all should be returned.
}
\section{R6 Usage}{
 $mode(which = "all")
}

