% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{mode}
\alias{mode}
\alias{mode.Distribution}
\title{Mode of a Distribution}
\usage{
mode(which = 1)
}
\arguments{
\item{which}{which mode of the distribution should be returned, default is the first.}
}
\description{
A numeric search for the mode(s) of a distribution.
}
\details{
If the distribution has multiple modes, the first is returned by default, similarly if it has
one only. Otherwise the index of the mode to return can be given or "All" if all should be returned.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.
}
