% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Multinomial.R
\name{Multinomial}
\alias{Multinomial}
\title{Multinomial Distribution}
\description{
Mathematical and statistical functions for the Multinomial distribution parameterised
with mean and variance, sd or precision.
}
\section{Constructor}{
 Multinomial$new(size, probs, decorators = NULL)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{size} \tab integer \tab number of trials. See details. \cr
\code{probs} \tab numeric \tab vector of probabilities. See details. \cr
\code{decorators} \tab Decorator \tab decorators to add functionality. See details. \cr
}
}

\section{Constructor Details}{
 The Multinomial distribution is parameterised by size and prob.
Size, N, is given as a single integer greater than zero, such that if \eqn{x} is a vector of \eqn{K} parameters
passed to the pmf then it should be true that \eqn{\sum x_i = N}.
The length of the probability vector, \eqn{K}, tells the constructor how many arguments to expect
to be passed to the pmf function. The probability vector is automatically normalised with
\deqn{probs = probs/sum(probs)}.
}

\section{Statistical Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{mean()} \tab \code{\link{mean.Distribution}} \cr
  \code{var()} \tab \code{\link{var}} \cr
  \code{cov()} \tab \code{\link{cov}} \cr
  \code{cor()} \tab \code{\link{cor}} \cr
  \code{skewness()} \tab \code{\link{skewness}} \cr
  \code{kurtosis(excess = TRUE)} \tab \code{\link{kurtosis}} \cr
  \code{entropy(base = 2)} \tab \code{\link{entropy}} \cr
  \code{mgf(t)} \tab \code{\link{mgf}} \cr
  \code{cf(t)} \tab \code{\link{cf}} \cr
  \code{pgf(t)} \tab \code{\link{pgf}} \cr
  \code{sd()} \tab \code{\link{sd}} \cr
  }
}

\section{Public Variables}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Return} \cr
  \code{name} \tab Name of distribution. \cr
  \code{short_name} \tab Id of distribution. \cr
  \code{description} \tab Brief description of distribution. \cr
  \code{traits} \tab List: type, valueSupport, variateForm.
  }
}

\section{Accessor Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{decorators()} \tab \code{\link{decorators}} \cr
  \code{valueSupport()} \tab \code{\link{valueSupport}} \cr
  \code{variateForm()} \tab \code{\link{variateForm}} \cr
  \code{type()} \tab \code{\link{type}} \cr
  \code{properties()} \tab \code{\link{properties}} \cr
  \code{support()} \tab \code{\link{support}} \cr
  \code{distrDomain()} \tab \code{\link{distrDomain}} \cr
  \code{symmetry()} \tab \code{\link{symmetry}} \cr
  \code{sup()}  \tab \code{\link{sup}} \cr
  \code{inf()} \tab \code{\link{inf}} \cr
  \code{skewnessType()} \tab \code{\link{skewnessType}} \cr
  \code{kurtosisType()} \tab \code{\link{kurtosisType}} \cr
  }
}

\section{p/d/q/r Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{pdf(x1, ..., log = FALSE)} \tab \code{\link{pdf}} \cr
  \code{cdf(x1, ..., lower.tail = TRUE, log.p = FALSE)} \tab \code{\link{cdf}}\cr
  \code{quantile(p, ..., lower.tail = TRUE, log.p = FALSE)} \tab \code{\link{quantile.Distribution}} \cr
  \code{rand(n)} \tab \code{\link{rand}}
 }
}

\section{Parameter Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{parameters(id)} \tab \code{\link{parameters}} \cr
  \code{getParameterValue(id, error = "warn")}  \tab \code{\link{getParameterValue}} \cr
  \code{setParameterValue(lst, error = "warn")} \tab \code{\link{setParameterValue}} \cr
}
}

\section{Validation Methods}{

 \tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{liesInSupport(x, all = TRUE)} \tab \code{\link{liesInSupport}} \cr
  \code{liesInType(x, all = TRUE)} \tab \code{\link{liesInType}} \cr
  \code{liesInDistrDomain(x, all = TRUE)} \tab \code{\link{liesInDistrDomain}} \cr
}
}

\section{Representation Methods}{

\tabular{ll}{
  \strong{Method} \tab \strong{Link} \cr
  \code{strprint()} \tab \code{\link[RSmisc]{strprint}} \cr
  \code{print()} \tab \code{\link[base]{print}} \cr
  \code{summary(full = T)} \tab \code{\link{summary.Distribution}} \cr
  \code{plot()} \tab Coming Soon. \cr
  \code{qqplot()} \tab Coming Soon. \cr
}
}

