% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\name{Naturals}
\alias{Naturals}
\title{Set of Natural Numbers}
\arguments{
\item{dim}{numeric dimension of set.}
}
\value{
Returns an R6 object of class Naturals.
}
\description{
The mathematical set of natural numbers.
}
\section{Usage}{
 Naturals$new(dim = 1)
}

\seealso{
\code{\link{Naturals}}, \code{\link{PosNaturals}}, \code{\link{Integers}},
  \code{\link{PosIntegers}}, \code{\link{NegIntegers}}, \code{\link{Rationals}}, \code{\link{PosRationals}},
  \code{\link{NegRationals}}, \code{\link{Reals}}, \code{\link{PosReals}}, \code{\link{NegReals}},
  \code{\link{ExtendedReals}}, \code{\link{Complex}}.
}
