% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\name{Naturals}
\alias{Naturals}
\title{Set of Natural Numbers}
\description{
The mathematical set of natural numbers.
}
\section{Constructor Arguments}{

 \tabular{lll}{
   \strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
   \code{dim = 1} \tab numeric \tab Dimension of the set. \cr
}
}

\seealso{
\code{\link{Naturals}}, \code{\link{PosNaturals}}, \code{\link{Integers}},
  \code{\link{PosIntegers}}, \code{\link{NegIntegers}}, \code{\link{Rationals}}, \code{\link{PosRationals}},
  \code{\link{NegRationals}}, \code{\link{Reals}}, \code{\link{PosReals}}, \code{\link{NegReals}},
  \code{\link{ExtendedReals}}, \code{\link{Complex}}.
}
