% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\name{Naturals}
\alias{Naturals}
\title{Set of Natural Numbers}
\description{
The mathematical set of natural numbers.
}
\details{
The set of Naturals is defined as the counting numbers, i.e.
\deqn{Naturals = \{0, 1, 2,...\}}{Naturals = {0, 1, 2,...}}
}
\section{Constructor}{
 Naturals$new(dim = 1, lower = 0)
}

\section{Constructor Arguments}{

 \tabular{lll}{
   \strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
   \code{dim} \tab numeric \tab Dimension of the set. \cr
   \code{lower} \tab integer \tab Where to start the set. \cr
}
}

\section{Constructor Details}{
 Generally the \code{lower} argument should be ignored, its primary use-case
is for the \code{PosNaturals} child-class.
}

\examples{
Naturals$new()
Naturals$new(dim = 2)

}
\seealso{
\code{\link{listSpecialSets}}
}
