% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{operation}
\alias{operation}
\title{Symbolic Operations for SetInterval}
\usage{
operation(unicode, ..., lower = NULL, upper = NULL, type = NULL,
  dim = NULL)
}
\arguments{
\item{unicode}{unicode symbol for the operation.}

\item{...}{sets and/or intervals to combine via the operation.}

\item{lower}{lower bound of new SetInterval}

\item{upper}{upper bound of new SetInterval}

\item{type}{type of new SetInterval}

\item{dim}{dimension of new SetInterval}
}
\value{
An R6 object of class SetInterval.
}
\description{
Operations for SetInterval objects and subclasses, symbolic only.
}
\details{
Generally not recommended to use this function directly but instead
  via \code{\link{product}} or \code{\link{union}}.
}
\seealso{
\code{\link{SetInterval}}.
}
