% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{parameters}
\alias{parameters}
\title{Parameters Accessor}
\usage{
parameters(object, id = NULL)
}
\arguments{
\item{object}{Distribution or ParameterSet.}

\item{id}{character, see details.}
}
\value{
An R6 object of class ParameterSet or a data.table.
}
\description{
Returns some or all the parameters in a distribution.
}
