% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{pdfPNorm}
\alias{pdfPNorm}
\title{Probability Density Function P-Norm}
\usage{
pdfPNorm(object, p = 2, lower = NULL, upper = NULL)
}
\arguments{
\item{object}{Distribution.}

\item{p}{p-norm to calculate.}

\item{lower}{lower limit for integration, default is infimum.}

\item{upper}{upper limit for integration, default is supremum.}
}
\description{
The p-norm of the pdf evaluated between given limits or over the whole support.
}
\seealso{
\code{\link{ExoticStatistics}} and \code{\link{decorate}}
}
