% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{pgf}
\alias{pgf}
\alias{pgf.Distribution}
\title{Probability Generating Function}
\usage{
pgf(x, z)
}
\arguments{
\item{x}{decorated distribution, see details.}

\item{z}{integer to evaluate probability generating function at}
}
\description{
Probability generating function of a discrete distribution
}
\details{
The probability generating function is defined by
\deqn{cf_X(t) = E_X[exp(z^x)]}
where X is the distribution and E_X is the expectation of the distribution X.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method. CoreStatistics methods can only be used if the distribution has first been decorated
with \code{decorate(Distribution, CoreStatistics)}.
}
\seealso{
\code{\link{decorate}} for the decorator function and \code{\link{CoreStatistics}} and
\code{\link{ExoticStatistics}} for other available methods for decorating.
}
