% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\name{PosIntegers}
\alias{PosIntegers}
\title{Set of Positive Integers}
\description{
The mathematical set of positive integers.
}
\section{Constructor Arguments}{

 \tabular{lll}{
   \strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
   \code{dim = 1} \tab numeric \tab Dimension of the set. \cr
   \code{zero = FALSE} \tab logical \tab If TRUE, zero is included in the set. \cr
}
}

\seealso{
\code{\link{Naturals}}, \code{\link{PosNaturals}}, \code{\link{Integers}},
  \code{\link{PosIntegers}}, \code{\link{NegIntegers}}, \code{\link{Rationals}}, \code{\link{PosRationals}},
  \code{\link{NegRationals}}, \code{\link{Reals}}, \code{\link{PosReals}}, \code{\link{NegReals}},
  \code{\link{ExtendedReals}}, \code{\link{Complex}}.
}
