% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{prec}
\alias{prec}
\title{Precision of a Distribution}
\usage{
prec(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
Precision of a distribution assuming variance is provided.
}
\details{
The precision is analytically computed as the reciprocal of the variance.
If the variance is not found in the distribution (analytically or numerically), returns error.
}
\section{R6 Usage}{
 $prec()
}

\seealso{
\code{\link{var}}
}
