% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{product.SetInterval}
\alias{product.SetInterval}
\alias{*.SetInterval}
\title{Symbolic Cartesian product.SetInterval for SetInterval}
\usage{
product.SetInterval(...)

\method{*}{SetInterval}(x, y)
}
\arguments{
\item{...}{sets and/or intervals to take the cartesian product.SetInterval of.}

\item{x}{distribution}

\item{y}{distribution}
}
\value{
An R6 object of class SetInterval.
}
\description{
Makes a symbolic representation for the cartesian product.SetInterval of sets/intervals.
}
\details{
This does not calculate the cartesian product.SetInterval of the arguments but
  is just a symbolic representation using unicode.
}
\seealso{
\code{\link{union.SetInterval}} for the union.SetInterval of two or more intervals/sets.
}
