% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_ProductDistribution.R
\name{ProductDistribution}
\alias{ProductDistribution}
\title{Product Distribution}
\description{
A wrapper for creating the joint distribution of two indepdendent probability distributions.
}
\details{
Exploits the following relationships of independent distributions
\deqn{f_P(x1, x2) = f_X1(x1) * f_X2(x2)}
\deqn{F_P(x1, x2) = F_X1(x1) * F_X2(x2)}
where f_P/F_P is the pdf/cdf of the joint (product) distribution P and X1,X2 are independent distributions.

\code{ProductDistribution} inherits all methods from \code{\link{Distribution}} and \code{\link{DistributionWrapper}}.
}
\section{Constructor Arguments}{

\tabular{ll}{
\code{dist1} \tab Distribution. \cr
\code{dist2} \tab Distribution. \cr
}
}

\examples{
prodBin <- ProductDistribution$new(Binomial$new(prob = 0.5, size = 10),
                                   Binomial$new(prob = 0.5, size = 20))
prodBin$pdf(x1 = 2, x2 =3)
prodBin$cdf(x1 = 5, x2 = 10)
}
\seealso{
\code{\link{ArrayDistribution}} and \code{\link{Distribution}}. As well as \code{\link{DistributionWrapper}}
for wrapper details.
}
