% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{properties}
\alias{properties}
\title{Properties Accessor}
\usage{
properties(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
List of distribution properties.
}
\description{
Returns the properties of the distribution.
}
\section{R6 Usage}{
 $properties
}

