% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{rand}
\alias{rand}
\title{Random Simulation Function}
\usage{
rand(object, n, simplify = TRUE)
}
\arguments{
\item{object}{(\link{Distribution})}

\item{n}{\code{(numeric(1))} \cr
Number of points to simulate from the distribution. If length greater than \eqn{1}, then
\code{n <- length(n)},}

\item{simplify}{\code{logical(1)} \cr
If \code{TRUE} (default) simplifies the pdf if possible to a \code{numeric}, otherwise returns a
\link[=data.table]{data.table::data.table}.}
}
\value{
Simulations as either a numeric if \code{simplify} is TRUE
or as a \link[data.table:data.table]{data.table::data.table}.
}
\description{
See \link{Distribution}\verb{$rand}
}
