% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_Scale.R
\name{Scale}
\alias{Scale}
\title{Distribution Scale Wrapper}
\description{
Scales a distribution to a given mean and standard deviation. By default the distribution
is centered (\code{mean} = 0) with unit variance (\code{sd} = 1)
}
\details{
STILL IN DEVELOPMENT. Results are likely sub-optimal.
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{dist} \tab distribution \tab Distribution to scale. \cr
\code{mean} \tab numeric \tab Desired mean after distribution scale. \cr
\code{sd} \tab numeric \tab Desired standard deviation after distribution scale.\cr
\code{var}\tab numeric \tab Desired variance after distribution scale.\cr
}
}

\section{getParameterValue}{

\tabular{lll}{
\strong{Method} \tab \strong{Return Type} \tab \strong{Details} \cr
\code{getParameterValue("scaledmean")} \tab numeric \tab Return mean of scaled distribution. \cr
\code{getParameterValue("scaledsd")} \tab numeric \tab Return standard deviation of scaled distribution.\cr
\code{getParameterValue("scaledvar")} \tab numeric \tab Return variance of scaled distribution.\cr
}
}

\section{setParameterValue}{

\tabular{lll}{
\strong{Method} \tab \strong{Input Type} \tab \strong{Details} \cr
\code{setParameterValue(scaledmean=value)} \tab numeric \tab Set the mean of scale distribution to value. \cr
\code{setParameterValue(scaledsd=value))} \tab numeric \tab Set the standard deviation of scale distribution to a value. \cr
\code{setParameterValue(scaledvar=value))} \tab numeric \tab Set the variance of scale distribution to a value. \cr
}
}

\examples{
\dontrun{
Scale$new(Binomial$new())
}

}
\seealso{
See \code{\link{DistributionWrapper}} for inherited wrapper methods and see \code{\link{Distribution}}
for a full list of inherited distribution methods.
}
