% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_Scale.R
\docType{data}
\name{Scale}
\alias{Scale}
\title{Distribution Scale Wrapper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Scale
}
\description{
Scales a distribution to a given mean and standard deviation. By default the distribution
is centered (\code{mean} = 0) with unit variance (\code{sd} = 1)
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{distribution} \tab distribution \tab Distribution to scale. \cr
\code{mean} \tab numeric \tab Desired mean after distribution shift. \cr
\code{sd} \tab numeric \tab Desired standard deviation after distribution scale.
}
}

\section{Getters}{

\tabular{lll}{
\strong{Method} \tab \strong{Return Type} \tab \strong{Details} \cr
\code{getScaleMean()} \tab numeric \tab Return mean of scaled distribution. \cr
\code{getScaleSd()} \tab numeric \tab Return standard deviation of scaled distribution. \cr
}
}

\section{Setters}{

\tabular{lll}{
\strong{Method} \tab \strong{Input Type} \tab \strong{Details} \cr
\code{setScaleMean(mean)} \tab numeric \tab Set mean to scale distribution to. \cr
\code{setScaleSd(sd)} \tab numeric \tab Set standard deviation to scale distribution to. \cr
}
}

\examples{
\dontrun{
Scale$new(Binomial$new())
}

}
\seealso{
See \code{\link{DistributionWrapper}} for inherited wrapper methods and see \code{\link{Distribution}}
for a full list of inherited distribution methods.
}
\keyword{datasets}
