% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_SDistribution.R
\name{SDistribution}
\alias{SDistribution}
\title{Abstract Special Distribution Class}
\value{
Returns error. Abstract classes cannot be constructed directly.
}
\description{
Abstract class that cannot be constructed directly.
}
\section{Methods}{

\itemize{
\item \code{mode(which = "all")} - Mode of the distribution. \code{which} selects which
mode to return if multimodal
\item \code{mean(...)} \deqn{E_X(X) = \sum p_X(x)*x}
\item \code{variance(...)} \deqn{var_X = E[X^2] - E[X]^2}
\item \code{median()} Any real-valued m where
\deqn{P(X \le m) \ge 0.5 and P(X \ge m) \ge 0.5}
\item \code{skewness(...)}
\deqn{sk_X = E_X[\frac{x - \mu}{\sigma}^3]}{sk_X = E_X[((x - \mu)/\sigma)^3]}
\item \code{kurtosis(excess = TRUE, ...)} \cr
If \code{excess = FALSE}
\deqn{k_X = E_X[\frac{x - \mu}{\sigma}^4]}{k_X = E_X[((x - \mu)/\sigma)^4]}
If \code{excess = TRUE}
\deqn{k_X = E_X[\frac{x - \mu}{\sigma}^4] - 3}{k_X = E_X[((x - \mu)/\sigma)^4] - 3}
\item \code{entropy(base = 2, ...)} \deqn{- \sum (f_X)log_{base}(f_X)}
\item \code{mgf(t, ...)} \deqn{mgf_X(t) = E_X[exp(xt)]}
\item \code{cf(t, ...)} \deqn{cf_X(t) = E_X[exp(xti)]}
\item \code{pgf(z, ...)} \deqn{pgf_X(z) = E_X[exp(z^x)]}
}
}

\seealso{
\code{\link[=listDistributions]{listDistributions()}}
}
