% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{SetInterval}
\alias{SetInterval}
\alias{lower}
\alias{upper}
\alias{type}
\alias{dimension}
\alias{max}
\alias{min}
\alias{sup}
\alias{inf}
\alias{getSymbol}
\alias{print.SetInterval}
\title{R6 Generalised Class for Symbolic Sets and Intervals}
\arguments{
\item{symbol}{string representation of set/interval.}

\item{type}{string for interval type, one of (), (], [), []}

\item{lower}{numeric lower limit of set/interval}

\item{upper}{numeric upper limit of set/interval}

\item{dimension}{numeric dimension of set/interval}
}
\value{
\code{SetInterval$new} constructs an R6 object of class SetInterval.

\code{lower} gets lower limit of set/interval

\code{upper} gets upper limit of set/interval

\code{type} gets closed/open interval type

\code{dimension} gets dimension of set/interval

\code{max} gets maximum of set/interval

\code{min} gets minimum of set/interval

\code{sup} gets supremum of set/interval

\code{inf} gets infimum of set/interval

\code{getSymbol} gets symbolic representation of set/interval

\code{print} prints symbolic representation of set/interval
}
\description{
The parent class to sets and intervals in R6.
}
\details{
Whilst this is not an abstract class, direct construction is generally not advised.
  Construction should instead be called on 'Set' or 'Interval'
}
\section{Usage}{
 SetInterval$new(symbol, type, lower, upper, dimension)

 $lower()

 $upper()

 $type()

 $dimension()

 $max()

 $min()

 $sup()

 inf()

 $getSymbol()

 $print()
}

\seealso{
\code{\link{Set}} for R6 set objects and \code{\link{Interval}} for R6 interval objects.
}
