% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{SetInterval}
\alias{SetInterval}
\title{R6 Generalised Class for Symbolic Sets and Intervals}
\usage{
SetInterval$new(symbol, type, lower, upper, dimension)
}
\arguments{
\item{symbol}{string representation of set/interval.}

\item{type}{string for interval type, one of (), (], [), []}

\item{lower}{numeric lower limit of set/interval}

\item{upper}{numeric upper limit of set/interval}

\item{dimension}{numeric dimension of set/interval}
}
\value{
An R6 object of class SetInterval.
}
\description{
The parent class to sets and intervals in R6.
}
\details{
Whilst this is not an abstract class, direct construction is generally not advised.
  Construction should instead be called on 'sets' or 'interval'
}
\section{Methods}{


\code{$lower()} Get lower limit of set/interval

\code{$upper()} Get upper limit of set/interval

\code{$type()} Get closed/open interval type

\code{$dimension()} Get dimension of set/interval

\code{$max()} Get maximum of set/interval

\code{$min()} Get minimum of set/interval

\code{$sup()} Get supremum of set/interval

\code{$inf()} Get infimum of set/interval

\code{$getSymbol()} Get symbolic representation of set/interval

\code{$print()} Print symbolic representation of set/interval
}

\seealso{
\code{\link{set}} for R6 set objects and \code{\link{interval}} for R6 interval objects.
}
