% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{SetInterval}
\alias{SetInterval}
\title{R6 Generalised Class for Symbolic Sets and Intervals}
\description{
The parent class to sets and intervals in R6. Primarily used for inheriting methods.
}
\details{
Whilst this is not an abstract class, direct construction is generally not advised.
  Construction should instead be called on 'Set' or 'Interval'
}
\section{Constructor Arguments}{

 \tabular{lll}{
   \strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
   \code{symbol} \tab character \tab Representation of set/interval. \cr
   \code{type} \tab character \tab Interval type, one of (), (], [), []. \cr
   \code{lower} \tab numeric \tab Lower limit of set/interval. \cr
   \code{upper} \tab numeric \tab Upper limit of set/interval. \cr
   \code{class} \tab character \tab Atomic class, one of "numeric" or "integer". \cr
   \code{dimension} \tab integer \tab Dimension of set/interval.
}
}

\section{Accessor Methods}{

\tabular{lll}{
\strong{Method} \tab \strong{Return Type} \tab \strong{Details} \cr
\code{lower} \tab numeric \tab Lower limit of set/interval. \cr
\code{upper} \tab numeric \tab Upper limit of set/interval. \cr
\code{type} \tab  character \tab Closed/open interval type. \cr
\code{class} \tab  character \tab Class of set/interval. \cr
\code{dimension} \tab integer \tab Dimension of set/interval. \cr
\code{max} \tab  numeric \tab Maximum of set/interval. \cr
\code{min} \tab  numeric \tab Minimum of set/interval. \cr
\code{sup} \tab  numeric \tab Supremum of set/interval. \cr
\code{inf} \tab  numeric \tab Infimum of set/interval. \cr
\code{getSymbol} \tab  character \tab Symbolic representation of set/interval.
}
}

\section{Representation Methods}{

\tabular{ll}{
  \strong{Method} \tab \strong{Details} \cr
  \code{print} \tab  Prints symbolic representation of set/interval \cr
}
}

\seealso{
\code{\link{Set}} for R6 Set objects and \code{\link{Interval}} for R6 Interval objects.
}
