% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{setParameterValue}
\alias{setParameterValue}
\title{Parameter Value Setter}
\usage{
setParameterValue(object, lst, error = "warn")
}
\arguments{
\item{object}{Distribution or ParameterSet.}

\item{lst}{list, see details.}

\item{error}{character, value to pass to \code{stopwarn}.}
}
\description{
Returns the value of the given parameter.
}
\details{
A list is supplied to the function, the list names are parameter IDs and the list values are
the respective values to set the parameters.

\code{stopwarn} either breaks the code with an error if "error" is given or returns \code{NULL}
with warning otherwise.
}
\section{R6 Usage}{
 $setParameterValue(lst, error = "warn")
}

\seealso{
\code{\link{parameters}} and \code{\link{setParameterValue}}
}
