% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSymbol.R
\name{setSymbol}
\alias{setSymbol}
\title{Unicode Symbol of Special Sets}
\usage{
setSymbol(set)
}
\arguments{
\item{set}{special set}
}
\description{
Gets the unicode symbol for standard mathematical special sets.
}
\details{
Special set can be supplied as a character string or class, case-insensitive.
  See \code{\link{listSpecialSets}} for full list of currently supported sets.
}
\examples{
# Supplied as class
setSymbol(empty)

# Supplied as string
setSymbol("empty")

# Case-insensitive
setSymbol(EmPtY)


}
\seealso{
\code{\link{SpecialSet}}, \code{\link{listSpecialSets}}
}
