% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\description{
Skewness of a distribution
}
\details{
The skewness of a distribution is defined by the third standardised moment of the
distribution,
\deqn{sk_X = E_X[(x - \mu)^3]/\sigma^3}
where E_X is the expectation of distribution X, \eqn{\mu} is the mean of the distribution and \eqn{\sigma} is the
standard deviation of the distribution.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method.
}
