% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{skewness}
\alias{skewness}
\title{Distribution Skewness}
\usage{
skewness(object, ...)
}
\arguments{
\item{object}{Distribution.}

\item{...}{Passed to \verb{$genExp}.}
}
\value{
Skewness as a numeric.
}
\description{
Skewness of a distribution
}
