% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{skewnessType}
\alias{skewnessType}
\title{Type of Skewness Accessor - Deprecated}
\usage{
skewnessType(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
If the distribution skewness is present in properties, returns one of
"negative skew", "no skew", "positive skew", otherwise returns NULL.
}
\description{
Deprecated. Use $properties$skewness.
}
