% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\docType{data}
\name{SpecialSet}
\alias{SpecialSet}
\alias{naturals}
\alias{posNaturals}
\alias{integers}
\alias{posIntegers}
\alias{negIntegers}
\alias{rationals}
\alias{posRationals}
\alias{negRationals}
\alias{reals}
\alias{posReals}
\alias{negReals}
\alias{extendedReals}
\alias{complex}
\title{Special Mathematical Sets}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpecialSet$new(dim = 1, lower = -Inf, upper = Inf, type = "()", ...)

naturals$new(dim = 1)

posNaturals$new(dim = 1)

integers$new(dim = 1, lower = -Inf, upper = Inf, type = "()")

posIntegers$new(dim = 1)

negIntegers$new(dim = 1)

rationals$new(dim = 1, lower = -Inf, upper = Inf, type = "()")

posRationals$new(dim = 1)

negRationals$new(dim = 1)

reals$new(dim = 1, lower = -Inf, upper = Inf, type=  "()")

posReals$new(dim = 1, zero = FALSE)

negReals$new(dim = 1, zero = FALSE)

extendedReals$new(dim = 1)

complex$new(dim = 1, lower = -Inf, upper = Inf, type = "()")
}
\arguments{
\item{dim}{numeric dimension of set.}

\item{lower}{lower limit of the set.}

\item{upper}{upper limit of the set.}

\item{type}{interval type, one of "()","(]","[)","[]".}

\item{zero}{logical; if FALSE (default), zero is not included in the set.}
}
\value{
An R6 object of class SpecialSet.
}
\description{
R6 representations of the 'special' mathematical sets.
}
\details{
All arguments passed to the constructor are elements in the set. See
  \code{\link{SetInterval}} for inherited methods and variables.
}
\section{Methods}{


\code{$lower()} Get lower limit of set/interval

\code{$upper()} Get upper limit of set/interval

\code{$type()} Get closed/open interval type

\code{$dimension()} Get dimension of set/interval

\code{$max()} Get maximum of set/interval

\code{$min()} Get minimum of set/interval

\code{$sup()} Get supremum of set/interval

\code{$inf()} Get infimum of set/interval

\code{$getSymbol()} Get symbolic representation of set/interval

\code{$print()} Print symbolic representation of set/interval
}

\seealso{
The parent classes \code{\link{set}} and \code{\link{SetInterval}}.
}
\keyword{datasets}
