% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\docType{data}
\name{SpecialSet}
\alias{SpecialSet}
\alias{Naturals}
\alias{PosNaturals}
\alias{Integers}
\alias{PosIntegers}
\alias{NegIntegers}
\alias{Rationals}
\alias{PosRationals}
\alias{NegRationals}
\alias{Reals}
\alias{PosReals}
\alias{NegReals}
\alias{ExtendedReals}
\alias{Complex}
\title{Special Mathematical Sets}
\format{}
\usage{
$new(dim = 1, lower = -Inf, upper = Inf, type = "()", ...)

Naturals$new(dim = 1)

PosNaturals$new(dim = 1)

Integers$new(dim = 1, lower = -Inf, upper = Inf, type = "()")

PosIntegers$new(dim = 1)

NegIntegers$new(dim = 1)

Rationals$new(dim = 1, lower = -Inf, upper = Inf, type = "()")

PosRationals$new(dim = 1)

NegRationals$new(dim = 1)

Reals$new(dim = 1, lower = -Inf, upper = Inf, type=  "()")

PosReals$new(dim = 1, zero = FALSE)

NegReals$new(dim = 1, zero = FALSE)

ExtendedReals$new(dim = 1)

Complex$new(dim = 1, lower = -Inf, upper = Inf, type = "()")
}
\arguments{
\item{dim}{numeric dimension of set.}

\item{lower}{lower limit of the set.}

\item{upper}{upper limit of the set.}

\item{type}{interval type, one of "()","(]","[)","[]".}

\item{zero}{logical; if FALSE (default), zero is not included in the set.}
}
\value{
Returns an R6 object of the given class.
}
\description{
R6 representations of the 'special' mathematical sets.
}
\details{
Whilst these could all inherit from the class Interval. By convention we instead use Set.
}
\seealso{
The parent class \code{\link{SetInterval}} for a full list of inherited methods and variables.
}
\keyword{datasets}
