% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_SpecialSet.R
\name{SpecialSet}
\alias{SpecialSet}
\title{Special Mathematical Sets}
\description{
Abstract class for the representation of the 'special' mathematical sets.
}
\details{
Special sets refer to the most commonly used (and important) sets in mathematics. Including
the sets of natural numbers, integers and reals.

This is an abstract class that cannot be constructed, instead construct one of the implemented
SpecialSet child-classes. For a full list of these see \code{\link{listSpecialSets}}.
}
\seealso{
\code{\link{listSpecialSets}}
}
