% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Kernel.R
\name{squared2Norm}
\alias{squared2Norm}
\title{Squared Probability Density Function 2-Norm}
\usage{
squared2Norm(object)
}
\arguments{
\item{object}{Distribution.}
}
\description{
The squared 2-norm of the Kernel pdf evaluated over the whole support.
}
\details{
The squared 2-norm of the pdf is defined by
\deqn{\int (f_X(u))^2 du}
where X is the Kernel and \eqn{f_X} is its pdf.
}
\section{R6 Usage}{
 $squared2Norm()
}

