% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_helpers.R
\name{strprint}
\alias{strprint}
\alias{strprint.R6}
\alias{strprint.list}
\title{String Representation of Print}
\usage{
strprint(x,...)

\method{strprint}{R6}(x, ...)

\method{strprint}{list}(x, ...)
}
\arguments{
\item{x}{R6 object}

\item{...}{Additional arguments}
}
\description{
Gets the result of a print method as a character string without printing. Useful for
methods such as \code{cat}, \code{summary} and \code{print}. Additionally can be used to easily
parse R6 objects into data-frames.
}
\examples{
strprint(Binomial$new())

}
