% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDistribution_Arrdist.R
\name{[.Arrdist}
\alias{[.Arrdist}
\title{Extract one or more Distributions from an Array distribution}
\usage{
\method{[}{Arrdist}(ad, i = NULL, j = NULL)
}
\arguments{
\item{ad}{\link{Arrdist} from which to extract Distributions.}

\item{i}{indices specifying distributions (first dimension) to extract, all returned if NULL.}

\item{j}{indices specifying curves (third dimension) to extract, all returned if NULL.}
}
\value{
If \code{length(i) == 1} and \code{length(j) == 1} then returns a \link{WeightedDiscrete} otherwise if
\code{j} is \code{NULL} returns an \link{Arrdist}. If \code{length(i)} is greater than 1 or \code{NULL} returns a
\link{Matdist} if \code{length(j) == 1}.
}
\description{
Extract a \link{WeightedDiscrete} or \link{Matdist} or \link{Arrdist} from a \link{Arrdist}.
}
\examples{
pdf <- runif(400)
arr <- array(pdf, c(20, 10, 2), list(NULL, sort(sample(1:20, 10)), NULL))
arr <- aperm(apply(arr, c(1, 3), function(x) x / sum(x)), c(2, 1, 3))
darr <- as.Distribution(arr, fun = "pdf")
# WeightDisc
darr[1, 1]
# Matdist
darr[1:2, 1]
# Arrdist
darr[1:3, 1:2]
darr[1, 1:2]
}
