% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{support}
\alias{support}
\title{Support Accessor - Deprecated}
\usage{
support(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
An R6 object of class \link[set6:Set]{set6::Set}.
}
\description{
Deprecated. Use $properties$support
}
\details{
The support of a probability distribution is defined as the interval where the pmf/pdf
is greater than zero,
\deqn{Supp(X) = \{x \ \in R: \ f_X(x) \ > \ 0\}}{Supp(X) = {x \epsilon R: f_X(x) > 0}}
where \eqn{f_X} is the pmf if distribution \eqn{X} is discrete, otherwise the pdf.
}
\section{R6 Usage}{
 $support
}

