% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_ExoticStatistics.R
\name{survivalPNorm}
\alias{survivalPNorm}
\title{Survival Function P-Norm}
\arguments{
\item{p}{p-norm to calculate.}

\item{lower}{lower limit for integration, default is infimum.}

\item{upper}{upper limit for integration, default is supremum.}
}
\description{
The p-norm of the survival function evaluated between given limits or over the whole support.
}
\details{
The p-norm of the survival function is defined by
\deqn{\int_a^b |S_X|^p d\mu)^1/p}
where X is the distribution, \eqn{S_X} is the survival function and a,b are the limits of integration.

Returns NULL if distribution is not continuous.

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method.
}
