% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{symmetry}
\alias{symmetry}
\title{Symmetry Accessor - Deprecated}
\usage{
symmetry(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
One of "symmetric" or "asymmetric".
}
\description{
Deprecated. Use $properties$symmetry.
}
