% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testContinuous}
\alias{testContinuous}
\alias{checkContinuous}
\alias{assertContinuous}
\title{assert/check/test/Continuous}
\usage{
testContinuous(
  object,
  errormsg = paste(object$short_name, "is not continuous")
)

checkContinuous(
  object,
  errormsg = paste(object$short_name, "is not continuous")
)

assertContinuous(
  object,
  errormsg = paste(object$short_name, "is not continuous")
)
}
\arguments{
\item{object}{Distribution}

\item{errormsg}{custom error message to return if assert/check fails}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if Distribution is continuous.
}
\examples{
testContinuous(Binomial$new()) # FALSE
}
