% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testDiscrete}
\alias{testDiscrete}
\alias{checkDiscrete}
\alias{assertDiscrete}
\alias{isDiscrete}
\title{assert/check/test/Discrete}
\usage{
testDiscrete(object)

checkDiscrete(object)

assertDiscrete(object)

isDiscrete(object)
}
\arguments{
\item{object}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is discrete.
}
\examples{
testDiscrete(Binomial$new()) # FALSE

}
