% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testDistribution}
\alias{testDistribution}
\alias{checkDistribution}
\alias{assertDistribution}
\title{assert/check/test/Distribution}
\usage{
testDistribution(
  object,
  errormsg = paste(object, "is not an R6 Distribution object")
)

checkDistribution(
  object,
  errormsg = paste(object, "is not an R6 Distribution object")
)

assertDistribution(
  object,
  errormsg = paste(object, "is not an R6 Distribution object")
)
}
\arguments{
\item{object}{object to test}

\item{errormsg}{custom error message to return if assert/check fails}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is a \link{Distribution}.
}
\examples{
testDistribution(5) # FALSE
testDistribution(Binomial$new()) # TRUE
}
