% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testMatrixvariate}
\alias{testMatrixvariate}
\alias{checkMatrixvariate}
\alias{assertMatrixvariate}
\title{assert/check/test/Matrixvariate}
\usage{
testMatrixvariate(object)

checkMatrixvariate(object)

assertMatrixvariate(object)
}
\arguments{
\item{object}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if Distribution is matrixvariate.
}
\examples{
testMatrixvariate(Binomial$new()) # FALSE

}
