% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testMatrixvariate}
\alias{testMatrixvariate}
\alias{checkMatrixvariate}
\alias{assertMatrixvariate}
\title{assert/check/test/Matrixvariate}
\usage{
testMatrixvariate(
  object,
  errormsg = paste(object$short_name, "is not matrixvariate")
)

checkMatrixvariate(
  object,
  errormsg = paste(object$short_name, "is not matrixvariate")
)

assertMatrixvariate(
  object,
  errormsg = paste(object$short_name, "is not matrixvariate")
)
}
\arguments{
\item{object}{Distribution}

\item{errormsg}{custom error message to return if assert/check fails}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if Distribution is matrixvariate.
}
\examples{
testMatrixvariate(Binomial$new()) # FALSE
}
