% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testMesokurtic}
\alias{testMesokurtic}
\alias{checkMesokurtic}
\alias{assertMesokurtic}
\alias{isMesokurtic}
\title{assert/check/test/Mesokurtic}
\usage{
assertMesokurtic(x)
checkMesokurtic(x)
testMesokurtic(x)
isMesokurtic(x)

checkMesokurtic(x)

assertMesokurtic(x)

isMesokurtic(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is mesokurtic.
}
\examples{
testMesokurtic(Binomial$new())

}
