% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testMesokurtic}
\alias{testMesokurtic}
\alias{checkMesokurtic}
\alias{assertMesokurtic}
\title{assert/check/test/Mesokurtic}
\usage{
testMesokurtic(
  object,
  errormsg = paste(object$short_name, "is not mesokurtic")
)

checkMesokurtic(
  object,
  errormsg = paste(object$short_name, "is not mesokurtic")
)

assertMesokurtic(
  object,
  errormsg = paste(object$short_name, "is not mesokurtic")
)
}
\arguments{
\item{object}{Distribution}

\item{errormsg}{custom error message to return if assert/check fails}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if Distribution is mesokurtic.
}
\examples{
testMesokurtic(Binomial$new())
}
