% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testMixture}
\alias{testMixture}
\alias{checkMixture}
\alias{assertMixture}
\title{assert/check/test/Mixture}
\usage{
testMixture(object, errormsg = paste(object$short_name, "is not mixture"))

checkMixture(object, errormsg = paste(object$short_name, "is not mixture"))

assertMixture(object, errormsg = paste(object$short_name, "is not mixture"))
}
\arguments{
\item{object}{Distribution}

\item{errormsg}{custom error message to return if assert/check fails}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if Distribution is mixture.
}
\examples{
testMixture(Binomial$new()) # FALSE
}
