% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testPlatykurtic}
\alias{testPlatykurtic}
\alias{checkPlatykurtic}
\alias{assertPlatykurtic}
\title{assert/check/test/Platykurtic}
\usage{
testPlatykurtic(
  object,
  errormsg = paste(object$short_name, "is not platykurtic")
)

checkPlatykurtic(
  object,
  errormsg = paste(object$short_name, "is not platykurtic")
)

assertPlatykurtic(
  object,
  errormsg = paste(object$short_name, "is not platykurtic")
)
}
\arguments{
\item{object}{Distribution}

\item{errormsg}{custom error message to return if assert/check fails}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if Distribution is platykurtic.
}
\examples{
testPlatykurtic(Binomial$new())
}
