% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testPositiveSkew}
\alias{testPositiveSkew}
\alias{checkPositiveSkew}
\alias{assertPositiveSkew}
\alias{isPositiveSkew}
\title{assert/check/test/PositiveSkew}
\usage{
assertPositiveSkew(x)
checkPositiveSkew(x)
testPositiveSkew(x)
isPositiveSkew(x)

checkPositiveSkew(x)

assertPositiveSkew(x)

isPositiveSkew(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}/\code{is}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test}/\code{is} return \code{FALSE}.
}
\description{
Validation checks to test if Distribution is positive skew.
}
\examples{
testPositiveSkew(Binomial$new())

}
