% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testPositiveSkew}
\alias{testPositiveSkew}
\alias{checkPositiveSkew}
\alias{assertPositiveSkew}
\title{assert/check/test/PositiveSkew}
\usage{
testPositiveSkew(
  object,
  errormsg = paste(object$short_name, "is not positive skew")
)

checkPositiveSkew(
  object,
  errormsg = paste(object$short_name, "is not positive skew")
)

assertPositiveSkew(
  object,
  errormsg = paste(object$short_name, "is not positive skew")
)
}
\arguments{
\item{object}{Distribution}

\item{errormsg}{custom error message to return if assert/check fails}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if Distribution is positive skew.
}
\examples{
testPositiveSkew(Binomial$new())
}
