% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testSymmetric}
\alias{testSymmetric}
\alias{checkSymmetric}
\alias{assertSymmetric}
\title{assert/check/test/Symmetric}
\usage{
testSymmetric(object, errormsg = paste(object$short_name, "is not symmetric"))

checkSymmetric(object, errormsg = paste(object$short_name, "is not symmetric"))

assertSymmetric(
  object,
  errormsg = paste(object$short_name, "is not symmetric")
)
}
\arguments{
\item{object}{Distribution}

\item{errormsg}{custom error message to return if assert/check fails}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if Distribution is symmetric.
}
\examples{
testSymmetric(Binomial$new()) # FALSE
}
