% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{type}
\alias{type}
\title{Type Accessor - Deprecated}
\usage{
type(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
An R6 object of class \link[set6:Set]{set6::Set}.
}
\description{
Deprecated. Use $traits$type
}
\section{R6 Usage}{
 $type
}

