% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_operations.R
\name{union.SetInterval}
\alias{union.SetInterval}
\alias{+.SetInterval}
\title{Symbolic Unions for SetInterval}
\usage{
union.SetInterval(..., dim = 1)

\method{+}{SetInterval}(x, y)
}
\arguments{
\item{x}{SetInterval}

\item{y}{SetInterval}

\item{...}{SetIntervals to take the union of.}

\item{dim}{dimension of new SetInterval.}
}
\description{
Makes a symbolic representation for the union of sets/intervals.
}
\details{
This does not calculate the union of the arguments but
  is just a symbolic representation using unicode.
}
\examples{
PosNaturals$new() * Reals$new()
product.SetInterval(PosNaturals$new(), Reals$new())

}
\seealso{
\code{\link{product.SetInterval}}, \code{\link{complement.SetInterval}},
\code{\link{power.SetInterval}}
}
