% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{valueSupport}
\alias{valueSupport}
\title{Value Support Accessor}
\usage{
valueSupport(object)
}
\arguments{
\item{object}{distribution.}
}
\description{
Returns the valueSupport of the distribution, one of discrete/continuous/mixture.
}
\section{R6 Usage}{
 $valueSupport()
}

