% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{valueSupport}
\alias{valueSupport}
\title{Value Support Accessor - Deprecated}
\usage{
valueSupport(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
One of "discrete"/"continuous"/"mixture".
}
\description{
Deprecated. Use $traits$valueSupport
}
