% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{var.Distribution}
\alias{var.Distribution}
\title{Numeric Variance of a Distribution}
\arguments{
\item{object}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A numeric variance calculation for distributions.
}
\details{
The variance of a probability distribution can be numerically calculated via the generalised
expectation function and the formula
\deqn{var_X = E[X^2] - E[X]^2}

Documentation is for the S3 method, the first parameter can be omitted if calling as
an R6 method.
}
