% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionDecorator_CoreStatistics.R
\name{variance}
\alias{variance}
\title{Distribution Variance}
\usage{
variance(object, ...)
}
\arguments{
\item{object}{Distribution.}

\item{...}{Passed to \verb{$genExp}.}
}
\value{
Variance as a numeric.
}
\description{
The variance or covariance of a distribution, either calculated analytically if
or estimated numerically.
}
