% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{variateForm}
\alias{variateForm}
\title{Variate Form Accessor}
\usage{
variateForm(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
One of "univariate"/"multivariate"/"matrixvariate".
}
\description{
Returns the variateForm of the distribution.
}
\section{R6 Usage}{
 $variateForm()
}

