% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{variateForm}
\alias{variateForm}
\title{Variate Form Accessor - Deprecated}
\usage{
variateForm(object)
}
\arguments{
\item{object}{Distribution.}
}
\value{
One of "univariate"/"multivariate"/"matrixvariate".
}
\description{
Deprecated. Use $traits$variateForm
}
