% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{wrappedModels}
\alias{wrappedModels}
\title{Gets Internally Wrapped Models}
\usage{
wrappedModels(object, model = NULL)
}
\arguments{
\item{object}{Distribution.}

\item{model}{character, see details.}
}
\value{
If \code{model} is NULL then returns list of models that are wrapped by the wrapper.
Otherwise returns model given in \code{model}.
}
\description{
Returns either a list of all the wrapped models or the models named by parameters.
}
